Correspondence Analysis Results

An S4 class to store the results of a simple correspondence analysis.

Slots

data

A numeric matrix

dimension

An integer giving the dimension of the solution.

row_names

A character vector specifying the row names.

row_coordinates

A numeric matrix giving the row standard coordinates.

row_distances

A numeric vector giving the row chi-square distances to centroid.

row_inertias

A numeric vector giving the row inertias.

row_masses

A numeric vector giving the row masses.

row_supplement

A logical vector specifying if a row is a supplementary point.

column_names

A character vector specifying the column names.

column_coordinates

A numeric matrix giving the column standard coordinates.

column_distances

A numeric vector giving the column chi-square distances to centroid.

column_inertias

A numeric vector giving the column inertias.

column_masses

A numeric vector giving the column masses.

column_supplement

A logical vector specifying if a column is a supplementary point.

singular_values

A numeric vector giving the singular values.

Subset

In the code snippets below, x is a CA object.

x[[i]]

Extracts informations from a slot selected by subscript i. i is a length-one character vector. Returns the corresponding slot values.

See also

Author

N. Frerebeau

Examples

## Create count data matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)
B <- matrix(data = sample(12:20, 42, TRUE), nrow = 7, ncol = 6)

## Compute correspondence analysis
X <- ca(A, sup_rows = 8:10, sup_columns = 7:10)

## Get row coordinates
get_coordinates(X, margin = 1, standard = FALSE)

#>            CA1         CA2          CA3          CA4         CA5  .sup
#> 1  -0.12666650 -0.07214315 -0.278176418 -0.043944892  0.02702429 FALSE
#> 2   0.42822559 -0.30767284  0.002813518  0.045837642  0.02194862 FALSE
#> 3  -0.01008883  0.47382411 -0.117848155 -0.007964038 -0.04796407 FALSE
#> 4  -0.30866834 -0.26536066  0.107452438 -0.043081177 -0.05754344 FALSE
#> 5   0.19771575  0.12418978  0.015553309  0.158720623 -0.04287627 FALSE
#> 6   0.20723850  0.17840794  0.166308837 -0.156854971  0.02330331 FALSE
#> 7  -0.40601769  0.12154788  0.161700910  0.125104372  0.08185251 FALSE
#> 8  -0.19745347 -0.51219154 -0.118417216  0.119744868  0.22432756  TRUE
#> 9   0.20416072  0.41229302 -0.501631433  0.360773939  0.21422002  TRUE
#> 10  0.02218649 -0.32815808 -0.119033603  0.268197617 -0.12047693  TRUE

## Get column coordinates
get_coordinates(X, margin = 2, standard = FALSE)

#>            CA1         CA2          CA3         CA4          CA5  .sup
#> 1   0.08497257  0.26634343  0.028356927 -0.03184594  0.075142569 FALSE
#> 2   0.21013486  0.20030503 -0.250239068  0.12081842 -0.040192701 FALSE
#> 3  -0.02695405  0.13380491  0.117592195 -0.10622146 -0.056089107 FALSE
#> 4   0.04111899 -0.40504544 -0.232074657 -0.13882958  0.016933470 FALSE
#> 5  -0.59264289 -0.06548465  0.001780906  0.08253799  0.006350240 FALSE
#> 6   0.27703815 -0.28799489  0.192634046  0.09818637  0.002950031 FALSE
#> 7   0.14342540  0.51733650  0.293785984 -0.05674685 -0.381646168  TRUE
#> 8   0.15758598  0.24996552 -0.259694669 -0.25977386 -0.063981499  TRUE
#> 9   0.10012968  0.49844012  0.210706441  0.13933424  0.224586858  TRUE
#> 10 -0.01658832  0.32755206 -0.209232929  0.09905221 -0.247159401  TRUE

## Get row distances to centroid
get_distances(X, margin = 1)

#>         1         2         3         4         5         6         7         8 
#> 0.3182650 0.5297457 0.4907782 0.4270898 0.2859845 0.3571867 0.4776216 0.6164514 
#>         9        10 
#> 0.7995932 0.4569394 

## Get row inertias
get_inertia(X, margin = 1)

#>          1          2          3          4          5          6          7 
#> 0.01896346 0.04100537 0.02419631 0.03165030 0.01083026 0.01980730 0.02395806 
#>          8          9         10 
#>         NA         NA         NA 

## Get row contributions
get_contributions(X, margin = 1)

#>           CA1       CA2          CA3         CA4       CA5
#> 1  3.97857818  1.676687 58.043633869  3.69055467  6.688581
#> 2 35.49079389 23.801603  0.004634251  3.13390259  3.443546
#> 3  0.01354331 38.809219  5.589827094  0.06503997 11.305664
#> 4 21.89718409 21.024931  8.026870507  3.28737833 28.107141
#> 5  6.85647710  3.514371  0.128343453 34.05302455 11.908948
#> 6  8.83161810  8.503255 17.204406080 38.99118813  4.124349
#> 7 22.93180533  2.669934 11.002284747 16.77891176 34.421771

## Get eigenvalues
get_eigenvalues(X)

#>     eigenvalues percentage cumulative
#> CA1 0.075497990  44.303457   44.30346
#> CA2 0.058113626  34.102028   78.40549
#> CA3 0.024958919  14.646303   93.05179
#> CA4 0.009796364   5.748667   98.80045
#> CA5 0.002044158   1.199545  100.00000