Multivariate Data Analysis Results

An S4 class to store the results of a multivariate data analysis.

Slots

names

A character vector specifying the row names.

principal

A numeric matrix giving the principal coordinates.

standard

A numeric matrix giving the standard coordinates.

contributions

A numeric matrix giving the contributions to the definition of the dimensions.

cosine

A numeric matrix giving the \(cos^2\) values.

distances

A numeric vector giving the distances to centroid.

weights

A numeric vector giving the masses/weights.

supplement

A logical vector specifying the supplementary points.

order

An integer vector giving the original indices of the data (computation moves all supplementary points at the end of the results).

groups

A character vector specifying the class for each observation.

See also

Author

N. Frerebeau