PCA Results

An S4 class to store the results of a principal components analysis.

Slots

center

A numeric vector giving the column mean of the initial dataset (active individuals only).

scale

A numeric vector giving the column standard deviations of the initial dataset (active individuals only).

See also

Author

N. Frerebeau

Examples

## Load data
data("iris")

## Compute principal components analysis
X <- pca(iris, scale = TRUE, sup_row = sample(150, 10), sup_col = 2)
#> 1 qualitative variable was removed: Species.

## Get row coordinates
head(get_coordinates(X, margin = 1))
#>          F1         F2  .sup
#> 1 -1.991606  0.2856546 FALSE
#> 2 -2.128178  0.0864059 FALSE
#> 3 -2.298294 -0.0979697 FALSE
#> 4 -2.299490 -0.2273402 FALSE
#> 5 -2.059892  0.1860302 FALSE
#> 6 -1.533919  0.4046200 FALSE

## Get column coordinates
head(get_coordinates(X, margin = 2))
#>                      F1         F2  .sup
#> Sepal.Length  0.9288160  0.3695705 FALSE
#> Petal.Length  0.9839026 -0.1189751 FALSE
#> Petal.Width   0.9659451 -0.2341780 FALSE
#> Sepal.Width  -0.3615820  0.4537703  TRUE

## Get correlations between variables and dimensions
head(get_correlations(X))
#>                      F1         F2  .sup
#> Sepal.Length  0.9288160  0.3695705 FALSE
#> Petal.Length  0.9839026 -0.1189751 FALSE
#> Petal.Width   0.9659451 -0.2341780 FALSE
#> Sepal.Width  -0.3615820  0.4537703  TRUE

## Get eigenvalues
get_eigenvalues(X)
#>    eigenvalues  variance cumulative
#> F1   2.7638135 93.076802    93.0768
#> F2   0.2055768  6.923198   100.0000