Correspondence Analysis

Computes a simple correspondence analysis based on the singular value decomposition.

ca(object, ...)

# S4 method for data.frame
ca(object, rank = NULL, sup_row = NULL, sup_col = NULL)

# S4 method for matrix
ca(object, rank = NULL, sup_row = NULL, sup_col = NULL)

Arguments

object
A \(m \times p\) numeric matrix or a data.frame.

...
Currently not used.

rank
An integer value specifying the maximal number of components to be kept in the results. If NULL (the default), \(min(m, p) - 1\) components will be returned.

sup_row
A numeric or logical vector specifying the indices of the supplementary rows.

sup_col
A numeric or logical vector specifying the indices of the supplementary columns.

Value

A CA object.

See also

Author

N. Frerebeau

Examples

## Load data
library(codex)
data("compiegne")

## The chi square of independence between the two variables
stats::chisq.test(compiegne)

#> 
#> 	Pearson's Chi-squared test
#> 
#> data:  compiegne
#> X-squared = 67342, df = 60, p-value < 2.2e-16
#> 

## Compute correspondence analysis
X <- ca(compiegne, sup_col = 7:10)

## Get row coordinates
get_coordinates(X, margin = 1)

#>          CA1        CA2          CA3           CA4  .sup
#> 5 -0.6173651  0.2304056 -0.171327462 -0.0256156412 FALSE
#> 4 -0.2445848 -0.0902560  0.227869066  0.0009719754 FALSE
#> 3  0.1446483 -0.4759092 -0.099562140  0.0515217147 FALSE
#> 2  0.3880319  0.1299915 -0.012410526 -0.1763300900 FALSE
#> 1  0.3426913  0.4717792  0.002915709  0.1476656291 FALSE

## Get column coordinates
get_coordinates(X, margin = 2)

#>           CA1          CA2          CA3          CA4  .sup
#> A -0.26427888 -0.002106395 -0.077200851  0.007388896 FALSE
#> B  0.46579472 -0.265657215 -0.154574599  0.082193083 FALSE
#> C -0.18480877  0.261017642  0.083303541  0.003280442 FALSE
#> D  0.70606307  0.478395251 -0.005216167 -0.183986480 FALSE
#> E  0.40464014 -0.164513952  0.178014560 -0.196317101 FALSE
#> F  0.73946815  0.415115174 -0.016745192  0.159141489 FALSE
#> K -0.09746209 -0.435491389  0.243454796  0.049741374 FALSE
#> L  0.31510500 -0.633198145 -0.184727450 -0.040600432 FALSE
#> M  0.64089862  0.336922595  0.145420743 -0.135857618 FALSE
#> N -0.80499701  0.394831734 -0.376091538 -0.261346933 FALSE
#> O  0.23619808  0.865964911  0.031396693  0.420363758 FALSE
#> P  0.24789451 -0.245701716  0.420019296 -0.424330190 FALSE
#> G  0.71818947  0.404225421 -0.002320782 -0.101761055  TRUE
#> H  0.86789801  0.917187746 -0.003955767  0.487586192  TRUE
#> I  1.01229735  0.704811250 -0.069378353 -0.581275667  TRUE
#> J  0.90561140  0.703631246 -0.068820407  0.104582825  TRUE

## Get row distances to centroid
get_distances(X, margin = 1)

#>         5         4         3         2         1 
#> 0.4642357 0.1198931 0.2599798 0.1987129 0.3618266 

## Get row inertias
get_inertia(X, margin = 1)

#>          5          4          3          2          1 
#> 0.07771673 0.02891810 0.06388755 0.03658339 0.05845345 

## Get row contributions
get_contributions(X, margin = 1)

#>         CA1       CA2          CA3          CA4
#> 5 49.055540  8.417617 24.688593097  1.097458438
#> 4 11.093341  1.861036 62.923451874  0.002276605
#> 3  3.953042 52.717159 12.238590957  6.517166729
#> 2 21.311824  2.946559  0.142463831 57.188980025
#> 1 14.586254 34.057629  0.006900241 35.194118203

## Get eigenvalues
get_eigenvalues(X)

#>     eigenvalues  variance cumulative
#> CA1  0.13006850 48.979094   48.97909
#> CA2  0.10557785 39.756801   88.73590
#> CA3  0.01990369  7.495010   96.23091
#> CA4  0.01000918  3.769094  100.00000