Correspondence Analysis

Computes a simple correspondence analysis based on the singular value decomposition.

ca(object, ...)

# S4 method for matrix
ca(object, n = NULL, sup_rows = NULL, sup_columns = NULL)

# S4 method for CA
get_coordinates(x, margin = 1, standard = FALSE, sup = TRUE, sup_name = ".sup")

# S4 method for CA
get_inertia(x, margin = 1)

# S4 method for CA
get_distances(x, margin = 1)

# S4 method for CA
get_contributions(x, margin = 1)

# S4 method for CA
get_eigenvalues(x)

Arguments

object
A matrix or a data.frame.

...
Currently not used.

n
An integer value specifying the number of dimensions to be kept in the results (default to 5).

sup_rows
A numeric or logical vector specifying the indices of the supplementary rows.

sup_columns
A numeric or logical vector specifying the indices of the supplementary columns.

x
A CA object.

margin
A length-one numeric vector giving the subscript which the data will be returned: 1 indicates rows (the default), 2 indicates columns.

standard
A logical scalar: should standard coordinates be returned instead of principal coordinates (the default)?

sup
A logical scalar: should supplementary points be returned?

sup_name
A character string specifying the name of the column to create for supplementary points attribution (see below).

Value

ca returns a CA object.

get_* returns a data.frame.

Details

get_coordinates returns a data.frame of coordinates. If sup is TRUE, an extra column (named after sup_name) is added specifying whether an observation is a supplementary point or not.

get_eigenvalues returns a data.frame with the following columns: eigenvalues, percentage (percentage of variance) and cumulative (cumulative percentage of variance).

See also

Author

N. Frerebeau

Examples

## Create count data matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)
B <- matrix(data = sample(12:20, 42, TRUE), nrow = 7, ncol = 6)

## Compute correspondence analysis
X <- ca(A, sup_rows = 8:10, sup_columns = 7:10)

## Get row coordinates
get_coordinates(X, margin = 1, standard = FALSE)

#>            CA1          CA2         CA3          CA4          CA5  .sup
#> 1  -0.31660277  0.266513464  0.06201539 -0.058995592 -0.033987087 FALSE
#> 2   0.42459662 -0.297838583  0.04823677  0.104647464 -0.009595252 FALSE
#> 3  -0.25868011  0.219153199 -0.06142161  0.103162958 -0.057292890 FALSE
#> 4   0.33544164 -0.199983418 -0.16245303 -0.109838753 -0.046101577 FALSE
#> 5   0.50613774  0.303173306  0.13017351 -0.023333145  0.042521221 FALSE
#> 6  -0.52561448 -0.374094083  0.20584807 -0.036178736  0.013475053 FALSE
#> 7  -0.27039410 -0.003696783 -0.18937626  0.015013851  0.080133111 FALSE
#> 8  -0.31622187  0.242944872  0.32398317 -0.117182564  0.114428420  TRUE
#> 9  -0.33173967 -0.337753976  0.09396313  0.194509297  0.484361396  TRUE
#> 10 -0.08744743  0.341807902  0.26696296 -0.006800214 -0.143754280  TRUE

## Get column coordinates
get_coordinates(X, margin = 2, standard = FALSE)

#>             CA1         CA2          CA3          CA4          CA5  .sup
#> 1   0.146431009  0.05060717 -0.161029453  0.042133230 -0.059057217 FALSE
#> 2   0.170568501 -0.17322856  0.252168672  0.001790289 -0.043834960 FALSE
#> 3   0.026575389  0.58846794  0.083845162  0.017418107  0.033193118 FALSE
#> 4  -0.616486754 -0.02540980 -0.035899304 -0.120086734 -0.008416061 FALSE
#> 5   0.573507590 -0.16479439 -0.067260937 -0.063669097  0.058358131 FALSE
#> 6  -0.420791477 -0.23282828 -0.008731804  0.121378211  0.053624235 FALSE
#> 7  -0.003663127 -0.29707813 -0.168023785  0.023385171  0.048544579  TRUE
#> 8  -0.036490977 -0.60479743  0.031324554  0.002893499  0.182533993  TRUE
#> 9  -0.283030738 -0.18334229  0.056990265  0.176641395 -0.134693097  TRUE
#> 10 -0.230222949 -0.02729560  0.003361082 -0.074091603 -0.009042995  TRUE

## Get row distances to centroid
get_distances(X, margin = 1)

#>         1         2         3         4         5         6         7         8 
#> 0.4239673 0.5313756 0.3641994 0.4394262 0.6061247 0.6782926 0.3400540 0.5392673 
#>         9        10 
#> 0.7109134 0.4652531 

## Get row inertias
get_inertia(X, margin = 1)

#>          1          2          3          4          5          6          7 
#> 0.02588375 0.04065985 0.01856977 0.02703335 0.05731239 0.05520970 0.01803933 
#>          8          9         10 
#>         NA         NA         NA 

## Get row contributions
get_contributions(X, margin = 1)

#>         CA1          CA2       CA3       CA4        CA5
#> 1  9.620384 15.390189736  3.004174  9.056033  7.4182912
#> 2 17.302787 19.220628236  1.817534 28.494187  0.5912740
#> 3  6.243881 10.117351534  2.865063 26.922289 20.4947675
#> 4 10.499352  8.424791889 20.042289 30.519372 13.2700667
#> 5 26.635604 21.574944649 14.339484  1.534644 12.5790998
#> 6 22.096139 25.268886094 27.582810  2.838079  0.9717525
#> 7  7.601854  0.003207862 30.348647  0.635397 44.6747483

## Get eigenvalues
get_eigenvalues(X)

#>     eigenvalues percentage cumulative
#> CA1 0.150037390 61.8180301   61.81803
#> CA2 0.066459463 27.3824616   89.20049
#> CA3 0.018434711  7.5954235   96.79592
#> CA4 0.005534312  2.2802333   99.07615
#> CA5 0.002242263  0.9238515  100.00000