dimensio: Multivariate Data Analysis

Simple Principal Components Analysis (PCA) and Correspondence Analysis (CA) based on the Singular Value Decomposition (SVD). This package provides S4 classes and methods to compute, extract, summarize and visualize results of multivariate data analysis. It also includes methods for partial bootstrap validation described in Greenacre (1984) <isbn: 978-0-12-299050-2> and Lebart et al. (2006) <isbn: 978-2-10-049616-7>.

Details

Package:	dimensio

Type:	Package

Version:	0.2.1

License:	GPL-3

Zenodo:	doi: 10.5281/zenodo.4478530

See also

Author

Full list of authors and contributors (alphabetic order)

Jean-Baptiste Fourvel	CNRS, France

Nicolas Frerebeau	Université Bordeaux Montaigne, France

Brice Lebrun	Université Bordeaux Montaigne, France

Package maintainer

Nicolas Frerebeau
nicolas.frerebeau@u-bordeaux-montaigne.fr

IRAMAT-CRP2A (UMR 5060)
Maison de l'Archéologie
Université Bordeaux Montaigne
F-33607 Pessac cedex
France