Predict New Coordinates

Operators acting on objects to extract or replace parts.

# S4 method for CA
predict(object, data, margin = 1)

Arguments

object
A CA object..

data
An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin
A length-one numeric vector giving the subscript which the data will be returned: 1 indicates rows (the default), 2 indicates columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create count data matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)
B <- matrix(data = sample(12:20, 42, TRUE), nrow = 7, ncol = 6)

## Compute correspondence analysis
X <- ca(A, sup_rows = 8:10, sup_columns = 7:10)

## Predict new row coordinates
predict(X, B, margin = 1)

#>          CA1          CA2         CA3          CA4         CA5
#> 1 0.12084567 -0.012387980 -0.01259173  0.030339222 -0.02368704
#> 2 0.10157572 -0.049825142 -0.06976088 -0.002084658  0.06596631
#> 3 0.12472176  0.042139194  0.08540809  0.061140599  0.05232613
#> 4 0.05038707  0.007811258  0.06819297  0.011614377  0.01458099
#> 5 0.26260592 -0.111680290 -0.07738512  0.040082685  0.06105519
#> 6 0.31138805  0.017279733 -0.02073207  0.017064266  0.15564815
#> 7 0.24608316 -0.089490106  0.02051403  0.077141808  0.19517947

## Predict new column coordinates
predict(X, B, margin = 2)

#>            CA1         CA2         CA3         CA4         CA5
#> 1  0.008723511 -0.07188704 -0.03395671 -0.02168364 -0.07941615
#> 2 -0.110175690 -0.01576383 -0.23594613 -0.04912538  0.11236624
#> 3 -0.035205261 -0.11085686 -0.16776901 -0.05241581  0.15649626
#> 4 -0.012625740 -0.02023128 -0.14111440 -0.06187186  0.14598378
#> 5 -0.112705538 -0.04727639 -0.14340637 -0.03920645  0.14758876
#> 6 -0.024340539 -0.05563047 -0.21120203 -0.10265776  0.07172390