Non-Overlapping Text Labels

Non-Overlapping Text Labels

Usage

viz_labels(
  x,
  y,
  z,
  labels,
  ...,
  margin = 1,
  axes = c(1, 2),
  top = 10,
  box = FALSE,
  segment = FALSE,
  cex = graphics::par("cex"),
  col = graphics::par("fg"),
  bg = graphics::par("bg"),
  font = graphics::par("font")
)

Arguments

x, y

A numeric vector giving the x and y coordinates of a set of points. If y is missing, x must be a CA, MCA or PCA object.

z

A numeric vector giving an extra variable for subsetting.

labels

A character vector specifying the text to be written.

margin

A length-one numeric vector giving the subscript which the data will be returned: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

axes

A length-two numeric vector giving the dimensions to be plotted.

top

An integer specifying the number of labels to draw. Only the labels of the top \(n\) observations along z will be drawn. If NULL, all labels are drawn.

box

A logical scalar: should a box be drawn underneath labels?

segment

A logical scalar: should segments be drawn?