Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object
A CA or PCA object.

newdata
An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin
A length-one numeric vector giving the subscript which the data will be returned: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)

#>            CA1          CA2         CA3         CA4         CA5
#> 1  -0.26310025 -0.319993962  0.13962702  0.06999116  0.14113471
#> 2  -0.06821901 -0.150823784 -0.15673721 -0.26870632 -0.29972501
#> 3   0.19378526 -0.124887078 -0.51497068 -0.49649503  0.68824356
#> 4  -0.14814982 -0.325207918 -0.11182528  0.12284549 -0.16118129
#> 5  -0.20430209 -0.229795708  0.03266171  0.37746067 -0.04723501
#> 6  -0.14514126 -0.161380172 -0.30376738  0.12724360 -0.20080141
#> 7   0.19065221 -0.020002582 -0.07844325  0.22824642 -0.48405752
#> 8  -0.10834874 -0.017735509 -0.13232860 -0.29954598  0.21218811
#> 9   0.11457036 -0.047713307 -0.30532216 -0.41504297  0.09339018
#> 10 -0.19878519 -0.003225746  0.12555267  0.23816223 -0.13783604
#> 11  0.18307318 -0.309379534  0.08142047 -0.31325968 -0.09470154
#> 12 -0.11846867 -0.092618567  0.06512530  0.36634198 -0.37948591
#> 13 -0.26576809 -0.310175263 -0.21815995 -0.12495089 -0.12256998
#> 14  0.33611401  0.004625258  0.29357350  0.26631157  0.26761369
#> 15  0.03132380 -0.435744023  0.26922840 -0.01445938 -0.25167750
#> 16  0.17038171 -0.348175347 -0.41153532 -0.17535648  0.19682126
#> 17  0.16336733  0.317299104 -0.48420857  0.45880400  0.11149019
#> 18  0.18141750 -0.060733176 -0.21254195 -0.35653235  0.23035301
#> 19 -0.13391785 -0.115098185 -0.37223759 -0.37871763  0.27627281
#> 20  0.14553621  0.117931763  0.12277983  0.25976805 -0.17476193

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)

#>             CA1          CA2          CA3         CA4         CA5
#> 1   0.152504594 -0.229352303  0.180468031  0.20877614 -0.45724437
#> 2  -0.163954296  0.126801098  0.224986726  0.34159781 -0.08459503
#> 3  -0.123278363  0.176317824 -0.136462621  0.14171388 -0.06439708
#> 4  -0.138082055  0.166997321  0.036861437  0.16397205 -0.16824257
#> 5  -0.014542605  0.180900451  0.037690539 -0.04496971  0.03821410
#> 6  -0.066955759 -0.247722708  0.138591402  0.34669893 -0.08951825
#> 7   0.079390554  0.348719309  0.006440379 -0.24717338 -0.20718286
#> 8   0.312218605  0.176472102  0.318345557  0.34513378  0.18084205
#> 9   0.105284832  0.173587553 -0.036948854  0.10377186  0.05027168
#> 10 -0.239512798  0.287700526  0.063952507 -0.22224979 -0.02335059
#> 11 -0.306800767  0.169060711  0.188407406 -0.09361800  0.13915461
#> 12  0.524241600 -0.224136843 -0.243741677  0.48540372 -0.07856375
#> 13  0.002623805  0.821326732  0.259699972 -0.15692498 -0.00547570
#> 14  0.224584069 -0.120857376 -0.171035132  0.32985462 -0.08157549
#> 15  0.061811802  0.293084885 -0.024059126  0.21994626  0.24678748
#> 16  0.069508054 -0.007768587 -0.031221425 -0.05919848  0.19128374
#> 17 -0.154787306 -0.015512396 -0.312220306 -0.08813826 -0.14910468
#> 18 -0.233371408  0.470404835 -0.147178868 -0.02633813  0.02287626
#> 19  0.235274907  0.237388001 -0.322919766 -0.03848287 -0.06424274
#> 20  0.303989426  0.303692312  0.017141352 -0.06450347 -0.07332883