Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object
A CA or PCA object.

newdata
An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin
A length-one numeric vector giving the subscript which the data will be returned: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)

#>             CA1           CA2         CA3          CA4         CA5
#> 1   0.177369696  0.2317509755 -0.29854996  0.272754223  0.52419728
#> 2   0.482031505 -0.1710497430  0.05442428 -0.138288319 -0.22052740
#> 3   0.004853266 -0.3456297224  0.28009163  0.024492826  0.35776587
#> 4   0.462416913  0.2707611000 -0.01878772  0.019291678  0.03312502
#> 5  -0.076044799  0.1459814723 -0.01088903  0.079832545  0.17313257
#> 6  -0.032090528  0.2443513823 -0.25518035 -0.021166793 -0.21811681
#> 7   0.231139884  0.0617827365  0.22157966 -0.192649900 -0.12123044
#> 8  -0.410854508 -0.1218313422 -0.18286493  0.220072724 -0.18220300
#> 9   0.358581015  0.3138344545 -0.34410850  0.136108644  0.13262760
#> 10  0.207711720 -0.1217398497 -0.47878097 -0.009821116  0.13841145
#> 11 -0.121864407  0.0077348099 -0.23099801  0.427127629 -0.06163083
#> 12 -0.089629355 -0.0008245752 -0.25591095 -0.073124709 -0.04145854
#> 13  0.195084471 -0.2453241526 -0.10986510  0.236489420  0.32494545
#> 14 -0.216085091  0.2158019613 -0.11224067 -0.186337923 -0.12220742
#> 15  0.121983112  0.2204089582  0.08526321  0.102158847  0.17669301
#> 16  0.115525571  0.1551355358 -0.21596446 -0.053987245 -0.35677308
#> 17  0.060927624 -0.1065372039 -0.08839891  0.041358743 -0.22844249
#> 18  0.047710473 -0.1055854274 -0.33851819  0.166668364  0.11484142
#> 19 -0.041845409 -0.1918726066 -0.07682455  0.262960842 -0.13494241
#> 20 -0.569866067 -0.3909571496  0.11423997  0.109581789 -0.08067347

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)

#>            CA1         CA2         CA3          CA4          CA5
#> 1  -0.06539726  0.07055871 -0.21954027  0.001903712  0.069607011
#> 2  -0.18017105 -0.18084490 -0.26296402  0.195682438  0.302919196
#> 3  -0.36198017  0.17520724  0.23537315 -0.243213682  0.150971984
#> 4  -0.37088482 -0.09094025  0.28091945 -0.225931187 -0.514913935
#> 5  -0.27158892  0.06561160  0.10144165  0.102998798  0.209954942
#> 6  -0.06953374  0.22325905  0.25599489 -0.023485348 -0.001839841
#> 7   0.10832448 -0.22298626  0.04868477 -0.337774327 -0.280707878
#> 8   0.05588293  0.41208216 -0.01903509 -0.214370384 -0.199582246
#> 9   0.12143591  0.02106864 -0.01122085 -0.029596416 -0.266808813
#> 10 -0.08261756 -0.04316200  0.14722578  0.190514404 -0.109946854
#> 11 -0.43047715  0.23217777  0.18246820 -0.093383749  0.051514522
#> 12 -0.15391352 -0.07542618  0.33241867 -0.141426790 -0.005176298
#> 13  0.07373832  0.08014375  0.19747163 -0.120606794 -0.150260125
#> 14 -0.28193232 -0.24303176 -0.05879125 -0.016010831 -0.116469567
#> 15 -0.22557989  0.12418612 -0.05662343  0.328917803 -0.567986445
#> 16 -0.28214367  0.12087086  0.21391232 -0.024804355 -0.011357656
#> 17 -0.25670820 -0.13456031  0.26570712  0.158881799  0.345802215
#> 18 -0.40123057  0.02623313 -0.20517888  0.294901305 -0.064389215
#> 19 -0.41766253 -0.03389670  0.09724893 -0.108651030  0.050623724
#> 20  0.17946714 -0.11578870  0.01400709 -0.118973438 -0.016126217