Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>             F1           F2           F3          F4           F5
#> 1  -0.07390590  0.001783069 -0.119730519 -0.06358456 -0.188785073
#> 2   0.25082659 -0.169936180  0.173456767 -0.04278524 -0.120049825
#> 3   0.36267460 -0.092211852 -0.409132423  0.17062180  0.126287593
#> 4   0.06871337  0.039814824  0.003108246  0.36253838  0.205172775
#> 5  -0.19637132 -0.007118503  0.016729930  0.30745268  0.159327549
#> 6  -0.08802478  0.107587351 -0.481890285  0.12555711  0.007320991
#> 7   0.36747057  0.133608972 -0.459806678  0.09955478 -0.032212198
#> 8   0.21836777 -0.229476663  0.359598774  0.14510220  0.196473517
#> 9   0.20675067 -0.003159585  0.169484372  0.07759449 -0.032647062
#> 10 -0.05362780 -0.425701327 -0.185684830  0.21491180 -0.364134708
#> 11  0.27260734  0.165628820 -0.049253094 -0.41366866 -0.193433342
#> 12  0.25396987 -0.188293692  0.145620325  0.16256375 -0.431697741
#> 13  0.09019668  0.011515776  0.020791781 -0.03820306  0.098152807
#> 14  0.42064974  0.030340070  0.240436718 -0.08128810 -0.060445588
#> 15 -0.09051300  0.371772672  0.142596155  0.23322989 -0.015776662
#> 16  0.15050200  0.017387566 -0.163599935 -0.07661879 -0.106057629
#> 17 -0.29848018  0.034450409 -0.144009152  0.59294614 -0.216770045
#> 18  0.24173781  0.082033204  0.010414521 -0.15639493 -0.062674313
#> 19 -0.02400424  0.140679083  0.194173362 -0.05605293  0.082186271
#> 20 -0.18374779 -0.104180180  0.045260290  0.53182466 -0.109135424

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>              F1          F2          F3          F4           F5
#> 1   0.065749888 -0.05175383  0.13904799  0.18705741  0.037670616
#> 2   0.181903508 -0.36322521 -0.05354137 -0.01318748 -0.008776219
#> 3   0.300072158 -0.18190788  0.03022343  0.11824669  0.064895793
#> 4  -0.088435185 -0.02471004  0.17653879  0.03575744  0.023165204
#> 5   0.260625837 -0.18295715  0.03860776  0.05272710 -0.494506019
#> 6  -0.435148094 -0.09014585  0.14087441 -0.29801127 -0.070284717
#> 7   0.177163895 -0.06893103 -0.25494136 -0.10573437 -0.097841957
#> 8   0.440461854 -0.10992405  0.12736746  0.03502158 -0.075580495
#> 9  -0.095010491 -0.13479960  0.16659132 -0.05156633 -0.287931756
#> 10  0.763105057 -0.31556117  0.19348540 -0.03012250 -0.105843905
#> 11  0.586713293 -0.04380290 -0.14410248  0.19339010 -0.004132880
#> 12  0.120214502 -0.14259940  0.40506239  0.16788201 -0.261657348
#> 13  0.110638583  0.08568584  0.14810596  0.44248565 -0.436829846
#> 14 -0.099911950  0.21284652 -0.05958097  0.05294197  0.054889262
#> 15  0.174576904 -0.25959941  0.08512918  0.04160900  0.302005061
#> 16  0.207928542  0.31852880  0.20706541  0.09268719  0.202567331
#> 17  0.406193869 -0.06682681  0.08434230 -0.08361823  0.100931804
#> 18 -0.007934822 -0.17022202  0.28119546  0.32992460  0.151496121
#> 19  0.305801982 -0.24869672 -0.00843572  0.02186660  0.082916836
#> 20  0.114579759 -0.06044203  0.06188108  0.13276333 -0.127539954