Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>              F1           F2           F3           F4           F5
#> 1   0.215660690 -0.350223068 -0.466105323  0.163102338 -0.246084142
#> 2  -0.258285936  0.136638174  0.044383310  0.210810672 -0.124087028
#> 3   0.373100582  0.106696327  0.108605417  0.168231480  0.140714943
#> 4  -0.166017115  0.058978028 -0.370743047  0.474481755 -0.260805225
#> 5   0.114816285 -0.001159879  0.021752721  0.428396332  0.147212396
#> 6  -0.249290192  0.144783383  0.028086540  0.796509220  0.064215777
#> 7  -0.202363317 -0.055785073 -0.005319663  0.300301578  0.080669571
#> 8   0.160247130 -0.198306250 -0.061072099 -0.029570991  0.122804593
#> 9  -0.319612431 -0.038939169  0.208763412  0.001830194 -0.141335621
#> 10 -0.331691522 -0.027654534  0.154119054  0.092055812 -0.114595177
#> 11 -0.009013655  0.078993777 -0.084575409  0.309464418  0.200641621
#> 12 -0.373664965  0.095918201 -0.049286872  0.196964536  0.178678590
#> 13  0.199402666 -0.280944916 -0.281702916  0.406235271 -0.106024353
#> 14  0.199648748  0.028295488  0.165838295  0.308174438 -0.094020490
#> 15  0.210288884  0.133368830  0.257892751 -0.371185972 -0.008091406
#> 16 -0.136918792 -0.034337224 -0.044824802  0.271405280 -0.270429817
#> 17  0.373117588  0.052463670  0.023045630  0.153534308  0.066694477
#> 18 -0.329649792  0.041276719  0.097688552  0.406185437 -0.215366342
#> 19 -0.188970124  0.097632671  0.043043259  0.339172458 -0.260104316
#> 20  0.017988803  0.013610514 -0.104105342  0.517833161 -0.015605911

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>             F1           F2          F3           F4           F5
#> 1  -0.20204377 -0.113607680  0.03519338 -0.226709633  0.089292108
#> 2   0.24744652 -0.243949654  0.01553072  0.295289655 -0.431700809
#> 3  -0.03757401 -0.075240482 -0.03820217 -0.107735049 -0.139219966
#> 4  -0.16291173  0.017619591 -0.03840924  0.023552029  0.031898066
#> 5   0.11890901  0.053492550  0.35914607 -0.134613399 -0.422837839
#> 6   0.28902624 -0.566685032 -0.12016387  0.008986502  0.006043441
#> 7  -0.01936094 -0.049053352 -0.18707160  0.489122377  0.002524304
#> 8  -0.37004077  0.022372655  0.14883796  0.080068576 -0.463209730
#> 9  -0.10423118 -0.826602532  0.06704328 -0.022495933 -0.276169829
#> 10  0.03860882 -0.004832407 -0.07557478  0.029133807  0.079897797
#> 11 -0.05342905  0.027578010 -0.19054895  0.101776400 -0.335808349
#> 12  0.17151954 -0.011541525  0.04106311 -0.050000032 -0.052910010
#> 13 -0.16144984 -0.062237415  0.10553881 -0.300835713  0.053037464
#> 14 -0.14271233 -0.250811804  0.07418655  0.255073913 -0.092038813
#> 15 -0.23681727  0.003772744  0.09204046  0.001488046 -0.288416478
#> 16  0.10225432  0.043291819 -0.28890115 -0.325827137 -0.025096923
#> 17 -0.38524305 -0.345647550  0.18182821 -0.021824052 -0.137313461
#> 18  0.27701572 -0.054659822  0.10904292 -0.165212260 -0.108060912
#> 19 -0.26285879  0.281879743  0.12809215 -0.124056083 -0.311186957
#> 20  0.09254685 -0.257265931  0.07752544 -0.250223255 -0.508216052