Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>             F1           F2          F3          F4            F5
#> 1  -0.19332158  0.078744812 -0.04513442  0.28255231  0.3591813788
#> 2   0.55643753 -0.027455223 -0.05153851 -0.06940140 -0.1971343954
#> 3   0.53371927  0.023186694  0.42914174  0.07904470  0.3487221433
#> 4  -0.04539172  0.189708929  0.18386943 -0.04348177  0.2371105809
#> 5  -0.15642612  0.207942578  0.58504453  0.24777013 -0.2103878020
#> 6   0.15022808  0.091624098  0.26988624 -0.35677323  0.1742157746
#> 7   0.15347933 -0.040583274  0.06177837  0.08099771 -0.0006415857
#> 8  -0.24696495 -0.555622332  0.44207513 -0.19834536 -0.1579242894
#> 9   0.32927691  0.111694553 -0.10513151  0.75128409 -0.1248229945
#> 10 -0.49487945 -0.106583687  0.19119203 -0.33246164  0.2494478615
#> 11 -0.20389927 -0.204837583 -0.10604456  0.35770575  0.0735887043
#> 12  0.19088722  0.113171753  0.14439236 -0.03849687  0.3085049172
#> 13 -0.29916469  0.096687850  0.48150125 -0.04838797  0.0624809565
#> 14 -0.36391156 -0.400769517  0.46400637  0.19761882 -0.2320220884
#> 15 -0.14046151  0.448340216  0.04556680  0.07742675  0.2002893194
#> 16  0.46858947  0.002927467  0.41053336  0.38473219  0.5550382996
#> 17  0.10634393 -0.171046847  0.19847033 -0.01100720  0.2952179793
#> 18 -0.37649653  0.157921729  0.13232284  0.10243492 -0.3779553982
#> 19  0.30134051  0.211267149  0.67310855 -0.18213902  0.2875895824
#> 20  0.16628087 -0.059435795  0.15177826 -0.13975424  0.1551599678

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>             F1          F2          F3           F4           F5
#> 1   0.01633306  0.47503939 -0.15262639 -0.408900497 -0.017594601
#> 2  -0.35259384 -0.04939816  0.01016489  0.247140409  0.047718841
#> 3   0.21488582  0.09265837 -0.07417993  0.459459958 -0.424353880
#> 4   0.05052564 -0.13068505  0.11019691  0.244128718 -0.161908174
#> 5  -0.21290097 -0.19162090 -0.24349574 -0.651958772  0.005308587
#> 6   0.29359273  0.04863160  0.16464161 -0.015408805 -0.200742938
#> 7   0.26986987 -0.23371496 -0.08521534  0.030353031 -0.294892672
#> 8  -0.29235626  0.31954304 -0.09105535  0.036808492  0.097766006
#> 9  -0.34415666 -0.16348513  0.48385265  0.027524992 -0.394939934
#> 10 -0.28114767  0.34079511  0.38221346  0.155774912 -0.147446970
#> 11 -0.36257949  0.01370318  0.29261236  0.205728938 -0.375127339
#> 12 -0.10823744  0.30971593  0.44914059  0.077225738  0.318337455
#> 13  0.01521608 -0.08700485  0.29629606 -0.130666480 -0.134467530
#> 14 -0.09663652 -0.12166064  0.36672837  0.119065980  0.226723189
#> 15  0.24971803 -0.19685180  0.46533552  0.001703263 -0.008018625
#> 16  0.24338807  0.21290363 -0.06048821 -0.357487028 -0.258664262
#> 17  0.19584333 -0.04475438  0.35065789  0.295834428 -0.232956727
#> 18 -0.18767761 -0.19519557 -0.34973317 -0.075766302 -0.289291322
#> 19  0.07547395 -0.21184215 -0.01899937  0.209331320 -0.180751648
#> 20 -0.30153635  0.27606389 -0.08552942 -0.066095227  0.038138853