Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>             F1           F2          F3          F4           F5
#> 1   0.32722425  0.428294578  0.17045064 -0.21799032 -0.009735764
#> 2   0.25688691  0.382946804 -0.16304094 -0.07016705  0.094690373
#> 3   0.40181755  0.320363194  0.07048214  0.05769096  0.114149827
#> 4   0.01582835  0.324327523  0.31207259 -0.06563055  0.077239305
#> 5  -0.06312776 -0.263784376  0.16819459 -0.22562014 -0.117166384
#> 6  -0.13592447  0.037318626 -0.00968123  0.73812634  0.004837371
#> 7   0.50404474 -0.053658312 -0.16009274 -0.12970761  0.050644824
#> 8   0.14605290 -0.092254565  0.10658356  0.25070903  0.345173715
#> 9  -0.08147699 -0.083752570  0.02396736 -0.32542100  0.250292970
#> 10 -0.03922270  0.583295061  0.17504859  0.37374292  0.105299659
#> 11  0.13805514 -0.352241353  0.33039760  0.25999430 -0.123709842
#> 12  0.16594147  0.326335649  0.14859164 -0.32631766 -0.171359069
#> 13  0.12817120  0.363948926  0.55945495 -0.14632884 -0.318546994
#> 14  0.25133003  0.294343881  0.66954824  0.25941504  0.075886040
#> 15  0.19278901  0.075372867 -0.20082471 -0.02001985  0.058805695
#> 16  0.22297898  0.004313104  0.17384577 -0.08491327  0.411268305
#> 17  0.27708881  0.187393225 -0.23477643 -0.08098601 -0.173986834
#> 18  0.17773876  0.268019568  0.31031030  0.01248916  0.050796697
#> 19 -0.08542335  0.313396373  0.52206067  0.15956323  0.015482961
#> 20 -0.05923748  0.136382001  0.03504872  0.28079944 -0.221435135

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>             F1           F2          F3          F4           F5
#> 1   0.03797545  0.198251319 -0.03804260 -0.38206675 -0.097568483
#> 2  -0.28621092 -0.008385716  0.35506473  0.05578357  0.313871706
#> 3  -0.46102072 -0.003163349 -0.36753715  0.08303730  0.124518180
#> 4   0.01480667  0.008217229 -0.11017914  0.16591300  0.132669163
#> 5  -0.34461996 -0.040814037  0.07117658  0.27801082 -0.060483718
#> 6  -0.04340056 -0.241487650  0.21264736  0.09299901  0.139858040
#> 7  -0.18458733 -0.264352045  0.23382836  0.06625091  0.202818226
#> 8  -0.34240982  0.161491802 -0.35029302  0.08231787  0.225339324
#> 9  -0.22317409 -0.037003336  0.03102932  0.02123709 -0.113831092
#> 10 -0.38330230  0.097788034 -0.25651745 -0.21249314  0.382665653
#> 11 -0.41291506 -0.544720987  0.33779772 -0.08796489 -0.090457618
#> 12 -0.34566105 -0.126250778  0.10542954 -0.21002848  0.438652501
#> 13  0.05091648 -0.028500949 -0.44912050 -0.08132052  0.173103478
#> 14 -0.24683727  0.119104521  0.02034949 -0.27321919  0.380905214
#> 15 -0.19644912  0.292960565 -0.40168666  0.17383936  0.195620794
#> 16 -0.09104942 -0.046371172 -0.06972212 -0.21414780  0.370883885
#> 17 -0.47435027 -0.002383899 -0.41095002 -0.36743491  0.396997756
#> 18 -0.45216230 -0.159601619 -0.15124013 -0.31367855  0.394420485
#> 19 -0.23497531 -0.369753677  0.10712528  0.04118157 -0.001373646
#> 20 -0.22787496 -0.306614978  0.14579855  0.18105769 -0.018172075