Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>             F1          F2          F3          F4           F5
#> 1  -0.01412355  0.19398888 0.384577712 -0.47162917 -0.023410095
#> 2  -0.32951150  0.01693295 0.648320508  0.18979003  0.183318006
#> 3  -0.29107999 -0.16824603 0.422802777  0.01390620  0.148171473
#> 4  -0.45131002 -0.17522950 0.502028474  0.01268521 -0.205671185
#> 5  -0.23712257 -0.20365918 0.273761657  0.02643894 -0.003785588
#> 6  -0.18065349 -0.18031434 0.102057596  0.16223378 -0.089034159
#> 7  -0.17021695  0.09501005 0.246035651 -0.02756320  0.395522745
#> 8  -0.09656041 -0.02543845 0.117961364 -0.35410584  0.306013186
#> 9   0.07612015  0.11255102 0.503419990 -0.12240493 -0.046675016
#> 10  0.34059777  0.13656059 0.283418659  0.35063816  0.250579076
#> 11  0.36109858  0.29616525 0.249116040 -0.08543542  0.154015997
#> 12  0.12214283  0.23150107 0.001940084  0.09323026  0.569936509
#> 13 -0.30181743 -0.03607143 0.362565094  0.11716861  0.496080782
#> 14 -0.44330943 -0.29680641 0.201061977  0.06824659  0.141234258
#> 15  0.15090604  0.24917438 0.324650592 -0.15950619 -0.034181858
#> 16  0.38807238 -0.14321304 0.080484913  0.33542627 -0.303697661
#> 17  0.18407172  0.45963055 0.233733916  0.05381526  0.245625363
#> 18 -0.07856236 -0.06876986 0.067507440  0.08970055  0.081080351
#> 19 -0.24363676 -0.06568177 0.347838239  0.02904751 -0.115018861
#> 20 -0.12275356  0.28010925 0.319690319  0.03087345 -0.053725015

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>             F1          F2            F3            F4          F5
#> 1   0.07126518 -0.21678526 -0.1988745448 -0.2375053066  0.07103763
#> 2  -0.15584427  0.36961867 -0.0479482874  0.1771109469 -0.03870818
#> 3  -0.12047575 -0.07707445  0.3232725073 -0.2396431634  0.25724480
#> 4  -0.19800695  0.16729525  0.0753069755  0.0005654106 -0.20721061
#> 5  -0.20483193 -0.22159738  0.0109103770  0.2013427774  0.16266965
#> 6   0.42657564 -0.15140652  0.0222278123  0.1138840733 -0.07168602
#> 7  -0.27333713 -0.02176597 -0.1009461950  0.0887269088  0.06313982
#> 8   0.29133329  0.24996925 -0.3143516389 -0.1284547662 -0.14237407
#> 9   0.04746971  0.27013672 -0.1691469359  0.0820329116  0.17875572
#> 10 -0.15962354  0.27498559 -0.1440035471 -0.0675904433  0.31590774
#> 11  0.11460954  0.36049843  0.0270726140  0.1613310322 -0.32254410
#> 12  0.01951468 -0.17195264  0.1338098190 -0.1613341868 -0.04005375
#> 13  0.02997585  0.18540750  0.3752844529 -0.0119729491  0.10603522
#> 14  0.02766649  0.01352276  0.3174715621  0.1873131907  0.17649414
#> 15 -0.36454638 -0.15646161 -0.1475626830 -0.0108882001 -0.16109504
#> 16  0.18493007  0.23767749  0.1117795749  0.0076556712  0.04881633
#> 17  0.07770409  0.09877000 -0.0040805376 -0.0833059684  0.24941914
#> 18  0.12549617 -0.11775763  0.0549153655 -0.3752530703 -0.10638497
#> 19  0.08360939  0.18417814  0.0239380690  0.3178282280  0.21395293
#> 20  0.57579796  0.06503376  0.0006027458 -0.1576530879 -0.18386156