Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>             F1           F2           F3          F4          F5
#> 1  -0.10190590  0.107356323  0.541362120  0.11982138  0.11187941
#> 2  -0.06635992 -0.501677972  0.399882300 -0.15882161 -0.02296358
#> 3  -0.18945973  0.119891437  0.086436142 -0.09072282 -0.20032856
#> 4  -0.40753612  0.474324071  0.015530528  0.07031980  0.24436021
#> 5   0.29225991 -0.091493161  0.276443194  0.27951734 -0.20575019
#> 6  -0.18272980  0.033483932  0.097932270 -0.30537663  0.23272704
#> 7   0.37411690 -0.013158092  0.014140613  0.12286812  0.37742602
#> 8   0.32197563  0.094408910  0.551699206  0.10936644 -0.16155921
#> 9  -0.08494349  0.085478999 -0.283702941  0.24254607  0.21686476
#> 10 -0.01500920  0.220215322 -0.206592802 -0.01230219  0.31523960
#> 11  0.44305041  0.137422486  0.009497657 -0.18742181  0.02847299
#> 12  0.31804195  0.162076632 -0.022332378  0.10799850 -0.47578118
#> 13 -0.25021791  0.297110389  0.488684949  0.01389344  0.12540705
#> 14  0.16231992  0.013618899  0.304232355  0.03129726 -0.48011992
#> 15  0.40363387  0.187975262 -0.062780832 -0.47310141  0.16225221
#> 16 -0.33410892 -0.251412763  0.344007428 -0.11336153  0.14570628
#> 17  0.13834780  0.056581204 -0.030068379  0.15310332 -0.07457581
#> 18 -0.01647955  0.112896147 -0.189502632 -0.23776634  0.29919322
#> 19 -0.29017616  0.249747074 -0.062168370 -0.21240769 -0.08291059
#> 20  0.10843178 -0.001565905 -0.132881058  0.33712654  0.20427254

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>             F1          F2          F3           F4           F5
#> 1   0.11730232 -0.26769686 -0.04149253  0.247060848  0.104442284
#> 2   0.14868293 -0.08970197 -0.10668585  0.163941890  0.131645731
#> 3  -0.08904023 -0.16557918 -0.43419284  0.029420539  0.061899631
#> 4  -0.11184590 -0.14137372  0.26399932 -0.087661098 -0.072771939
#> 5   0.01624972  0.27697826 -0.18951996 -0.138221781  0.088545608
#> 6  -0.36420992  0.36124527 -0.06988578  0.326203999  0.239833728
#> 7  -0.18702177 -0.05116674  0.06236834 -0.043727103 -0.009602507
#> 8   0.11373235 -0.02172611 -0.45577558 -0.098139149 -0.004339168
#> 9  -0.21628160  0.12624988 -0.22164701 -0.141751307 -0.050344995
#> 10 -0.66948046  0.17117691 -0.48310885 -0.207889434 -0.061275609
#> 11 -0.55480892 -0.07092769 -0.01430952 -0.198988223  0.046328595
#> 12 -0.11232406  0.32259843 -0.30360105  0.283119302 -0.031342778
#> 13 -0.55778336  0.14602009 -0.10683243  0.195325247  0.007484956
#> 14 -0.30931994 -0.26072146 -0.21654181 -0.041720581 -0.159881293
#> 15  0.03845761  0.01909596 -0.37538271  0.131556653  0.221640118
#> 16 -0.05948684 -0.09477190 -0.14266967  0.144273700  0.524913621
#> 17 -0.06499138 -0.08805616  0.16112104  0.109780250  0.038625227
#> 18 -0.08616252 -0.08612273  0.23198853 -0.075312488  0.055363463
#> 19  0.10712468  0.05454489 -0.13234842 -0.365700589  0.027985120
#> 20  0.08937116  0.16777130  0.02001826  0.002968204  0.304351425