Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>             F1          F2          F3          F4           F5
#> 1   0.15167876 0.146955743 -0.05467790 -0.00245418 -0.054665695
#> 2   0.33690629 0.049946172  0.36453283  0.14080356  0.069808554
#> 3   0.49583306 0.008238112 -0.07125904  0.19851884  0.301717150
#> 4   0.06572409 0.148195412 -0.28389526 -0.16534742 -0.178250326
#> 5   0.22595622 0.388116187 -0.16517724  0.26035135 -0.001457947
#> 6  -0.01192670 0.267556261 -0.27890737  0.05526825 -0.186593429
#> 7  -0.10054308 0.200403620 -0.28161227  0.03257705  0.322858511
#> 8   0.25125871 0.060915229  0.09683419  0.02972263 -0.241168788
#> 9   0.38491619 0.530319520 -0.13058934  0.11342803  0.095649911
#> 10  0.08603218 0.304892010  0.44215720  0.03124757  0.057345327
#> 11  0.09806467 0.261328372 -0.20685066 -0.15422475 -0.043226384
#> 12 -0.01261305 0.276991975 -0.04964994  0.06800143  0.439380294
#> 13  0.26889815 0.104871193  0.45301370  0.13845677  0.320942848
#> 14  0.42862991 0.377645914  0.06766282 -0.11149100  0.429453866
#> 15  0.34414873 0.302721092  0.01833833  0.23950188  0.353640678
#> 16  0.59142997 0.667125463  0.35616328 -0.28740068  0.766168113
#> 17  0.06514117 0.071958864 -0.61136869 -0.18126170 -0.152549030
#> 18  0.40549067 0.352275877 -0.37072341 -0.04845683  0.140600819
#> 19 -0.13839278 0.043640873  0.01012154  0.10441796 -0.018179121
#> 20  0.48850963 0.186280506  0.06128242  0.12502317  0.143895819

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>             F1          F2          F3          F4          F5
#> 1  -0.06933221  0.18071281  0.26693356  0.41424565 -0.15729910
#> 2   0.04592059  0.11872756 -0.13405114 -0.34968176  0.06804315
#> 3  -0.13377681  0.00332031 -0.20835646  0.23732764  0.37460076
#> 4  -0.12704145 -0.14499127  0.09110455  0.06126536  0.61963297
#> 5  -0.05369722  0.08712387  0.20709210 -0.33142224  0.13639354
#> 6   0.13865610 -0.02697667 -0.61281545  0.16798035  0.27595777
#> 7  -0.49268515  0.23297190  0.16151748  0.12737367  0.14236644
#> 8  -0.12147651 -0.32362416 -0.24342813  0.59718977 -0.06240690
#> 9  -0.18889554 -0.60926836  0.24695885 -0.19160041  0.40201183
#> 10 -0.03842646  0.21472603 -0.02498343 -0.11038843  0.43481785
#> 11  0.17833602  0.36630316 -0.29705014 -0.05699892  0.11965235
#> 12  0.05492272  0.04200899  0.07529725  0.07899572  0.50380464
#> 13  0.25904667 -0.09417388  0.34899054 -0.06008784  0.09974337
#> 14  0.50262205 -0.02790541  0.09520137 -0.16176658  0.22793810
#> 15  0.25511190  0.28019300 -0.06408519  0.30755649  0.15174402
#> 16  0.13596359 -0.04020258  0.16956790 -0.13192501 -0.06409698
#> 17  0.28851653  0.09504129  0.07777962 -0.03506183  0.12432709
#> 18  0.25564379 -0.05164026  0.11627002  0.14954373  0.28633405
#> 19 -0.05841302 -0.23254208 -0.18519809 -0.19691685  0.25848982
#> 20 -0.62636146  0.23809838 -0.44001570  0.05143253  0.26637959