Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>             F1          F2          F3          F4          F5
#> 1  -0.04439788  0.13308750  0.02500310  0.23452873 -0.08721799
#> 2   0.36804589 -0.47804184  0.18156889  0.08398878  0.16814562
#> 3   0.36332679 -0.07704566 -0.08193928  0.06220585 -0.16769071
#> 4  -0.27295642  0.15787718  0.06778428 -0.01396348  0.24121509
#> 5  -0.09320885 -0.34329379 -0.19173697  0.06632394 -0.08160131
#> 6   0.12125285 -0.16673951 -0.20711541  0.22643784  0.07341751
#> 7  -0.02866735  0.17617811  0.31268820 -0.05802691 -0.02748847
#> 8   0.08659241 -0.32896472 -0.08660391  0.15157719 -0.31236478
#> 9  -0.20056857 -0.11676306  0.02118591  0.21005246 -0.14482675
#> 10  0.19849988 -0.06228495 -0.20728291 -0.03201876  0.17402461
#> 11 -0.05129413 -0.31159171  0.12889258  0.38637201 -0.42047993
#> 12  0.07696675 -0.36741555 -0.10548775  0.36429635 -0.04897699
#> 13 -0.21386740  0.14114655 -0.20517080  0.31635928  0.05452140
#> 14 -0.06867775  0.34851340  0.03423553  0.53418828 -0.17208765
#> 15  0.15151623  0.37903245  0.24867293 -0.13899959 -0.06134274
#> 16 -0.04675585 -0.27351371 -0.06609880  0.20017106 -0.20936709
#> 17  0.08704187  0.18406516  0.42631304 -0.02296057 -0.07778875
#> 18 -0.20980633 -0.03510166  0.42574554 -0.47698047  0.39597150
#> 19  0.32467736 -0.03736385  0.06618164 -0.29586787  0.24005672
#> 20  0.06068226  0.06251798 -0.04548030  0.09518243  0.36316755

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>             F1          F2           F3          F4          F5
#> 1  -0.30700551  0.05234177  0.347391970 -0.03556555 -0.01584941
#> 2  -0.20209456  0.18463039  0.280443831 -0.31576286 -0.14764263
#> 3  -0.17821881 -0.23146777  0.379262549  0.05065359 -0.08838915
#> 4  -0.02520912 -0.15693275 -0.117951532 -0.01759229 -0.28587512
#> 5   0.08168139  0.02491684 -0.009941468 -0.55674771  0.42216994
#> 6   0.03322227  0.23185798 -0.017157059 -0.02671105  0.11778936
#> 7  -0.14076079 -0.07525360  0.161575716 -0.27775989  0.26863324
#> 8  -0.15863630 -0.25769640  0.035377042  0.22808716 -0.06499253
#> 9  -0.20149669 -0.24098415  0.108621655 -0.15938167  0.18532027
#> 10  0.29382338 -0.19407221 -0.107006487 -0.06029567 -0.08045726
#> 11 -0.34536680 -0.18664434  0.045674511 -0.05669650 -0.16050746
#> 12 -0.18583683  0.23403176 -0.073212675 -0.20326768  0.19004468
#> 13 -0.22055162 -0.54989103 -0.140407124  0.13859595  0.04933687
#> 14  0.22769560 -0.01059047 -0.010075993 -0.11796310  0.14261302
#> 15 -0.22602364 -0.39336239 -0.183135862 -0.10585873  0.14217488
#> 16 -0.01443735 -0.18539513  0.017899350  0.12182089  0.18532306
#> 17  0.03312324 -0.39150297  0.188506240 -0.16833854  0.01234801
#> 18 -0.12514879  0.00898898 -0.190789057  0.15972691  0.04829861
#> 19 -0.31132336 -0.14085893 -0.077830063 -0.32782849  0.07073946
#> 20 -0.23774569 -0.18704668  0.465982892  0.02769014  0.07677367