Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>             F1          F2          F3           F4          F5
#> 1   0.06519054  0.01283390  0.31752074 -0.359215216 -0.22263126
#> 2   0.11370656  0.32106437  0.18368293  0.193181396 -0.08948827
#> 3  -0.13857797  0.10071381  0.07583201  0.024118161 -0.08180841
#> 4  -0.12502179  0.27572170  0.22210560  0.484686355 -0.08632334
#> 5   0.14796113  0.49598001  0.09275346 -0.185453759 -0.29166499
#> 6  -0.23214261 -0.01557094  0.13397271  0.030014983  0.11443669
#> 7   0.20708260  0.12084189 -0.15918613 -0.004353981  0.30905584
#> 8  -0.58002726  0.70360616  0.64245900 -0.238553495 -0.08868132
#> 9   0.68328892  0.04992519  0.36687714  0.357632526  0.18250726
#> 10 -0.43861958 -0.21768160 -0.11584844  0.167943595  0.16894546
#> 11  0.37839279  0.08463843  0.36366609  0.092162651  0.01781536
#> 12  0.19198407  0.12927124  0.14776282  0.693855853 -0.07332414
#> 13 -0.01852185  0.04209046  0.19606767  0.402075610  0.04085662
#> 14  0.17760508 -0.04817510 -0.22886446  0.387284795 -0.01942941
#> 15 -0.23852351  0.45004344  0.05492475  0.184320079  0.22800066
#> 16  0.01496026  0.05734501  0.02819244 -0.395741752  0.05265132
#> 17  0.11402309  0.07566262  0.14546042  0.438327644 -0.06755431
#> 18 -0.02748196 -0.01533842 -0.01166328  0.292492241 -0.47529884
#> 19  0.09068743  0.45019616 -0.01966247 -0.005160655 -0.10623617
#> 20  0.03769847  0.19085902 -0.27732124 -0.043116790  0.10539526

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>             F1           F2            F3          F4           F5
#> 1  -0.35142191  0.079323943 -0.3440333726 -0.87362879  0.233046285
#> 2  -0.07414428  0.261906743 -0.0002002315  0.10056470  0.014478527
#> 3  -0.09318177  0.114179573  0.0711196611  0.18176453 -0.494519019
#> 4   0.19230813  0.006435064 -0.1637073871  0.28939305 -0.042419663
#> 5  -0.27386613  0.101692755 -0.3411422196 -0.16512116  0.026740878
#> 6  -0.13464517 -0.386693432 -0.2180851498 -0.11213828 -0.166781449
#> 7  -0.40306294 -0.170210012 -0.0480230543 -0.04109337  0.114704501
#> 8   0.12946459 -0.235107514 -0.0970208524 -0.23023283  0.093583838
#> 9  -0.11741431 -0.066645185  0.2386203759 -0.48461837  0.007473921
#> 10 -0.02733633  0.028100282 -0.4863625693 -0.02596233 -0.237527991
#> 11  0.10205125 -0.165380263 -0.0725900498 -0.04290805  0.025533111
#> 12  0.10471425 -0.365057399  0.2548818740  0.20247349  0.131079749
#> 13  0.16978548 -0.204271917 -0.1589038540 -0.04179088 -0.025077914
#> 14  0.03096879  0.218605465 -0.0913749657  0.07841007  0.027241844
#> 15  0.11379847 -0.321201059 -0.0327606038 -0.12127935  0.045472843
#> 16  0.14401386 -0.361716546  0.2797309682 -0.29916307  0.011824005
#> 17 -0.03463632 -0.342814045 -0.1196412579 -0.17940519  0.001028117
#> 18  0.04480570  0.051643005 -0.2725607679  0.28729581 -0.114366069
#> 19  0.01462790 -0.022508021 -0.1685802495  0.02338384 -0.103087005
#> 20 -0.14893900 -0.417824324 -0.1969779997 -0.18937098  0.176903755