Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>              F1          F2          F3          F4          F5
#> 1   0.207520397 -0.18849116  0.14803448  0.12534658  0.10097431
#> 2   0.632324378 -0.02639126 -0.02403821  0.09666094 -0.03212619
#> 3  -0.197064734  0.01846526  0.02313404  0.22693514  0.14975557
#> 4   0.419529387 -0.07689392 -0.11489744  0.06974550  0.17172917
#> 5   0.206696512  0.34606496 -0.12845444  0.25431016 -0.08204402
#> 6   0.751982297  0.25919132 -0.16201649  0.22020403  0.20252732
#> 7   0.326600601  0.02148166 -0.31783180  0.52083498 -0.02795974
#> 8   0.187368767 -0.32415193 -0.54734100  0.25379464  0.28265699
#> 9   0.003958498  0.01264781 -0.29872316  0.32615616  0.24042801
#> 10  0.390578830  0.15875173 -0.36212174 -0.08343555  0.08475384
#> 11  0.118437528  0.06606329 -0.14269209  0.21760795  0.27314295
#> 12  0.611882419  0.25499797  0.09784346  0.15428780  0.25295030
#> 13 -0.421196164  0.13750672  0.41279547  0.28555035 -0.16631486
#> 14 -0.326248837 -0.38963101  0.05338453 -0.25691259  0.13749813
#> 15  0.202323854  0.23686504 -0.03784891 -0.46159670  0.04736240
#> 16 -0.047182885 -0.12495416  0.31602376  0.05334442  0.41686917
#> 17  0.146275745  0.12156831  0.44124982 -0.20011640  0.26275300
#> 18 -0.165805382 -0.11809098 -0.09214549  0.10761104 -0.26502571
#> 19 -0.166609975 -0.03884447  0.22380161 -0.27084880  0.35360255
#> 20 -0.321759840 -0.21377163  0.38460044  0.26193022 -0.13546568

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>             F1           F2          F3          F4          F5
#> 1  -0.30533747  0.007516554 -0.17212289  0.20697418 -0.04716989
#> 2  -0.43068461 -0.197625961 -0.11356001  0.10759739  0.04799492
#> 3  -0.00404756  0.609618129 -0.09986503  0.24307590  0.17834863
#> 4  -0.09359359  0.003452653  0.02002360  0.05456105  0.15225330
#> 5  -0.03597369 -0.048895765 -0.02561918 -0.05842995 -0.10137187
#> 6  -0.03659141 -0.488767602 -0.44340313 -0.03664669 -0.21064939
#> 7  -0.24065997  0.053974351  0.12266806  0.01999111 -0.13480492
#> 8   0.03416054 -0.076528107  0.06567956 -0.37106705 -0.25049552
#> 9  -0.04072037 -0.056334010 -0.22823074 -0.13041728 -0.14280201
#> 10 -0.31138396 -0.126594025  0.06347401  0.05731313 -0.26405777
#> 11 -0.05794810 -0.252622227  0.04872657 -0.19203284 -0.22388055
#> 12 -0.09071818 -0.053451755  0.23982417  0.23440797 -0.20115591
#> 13  0.22975431  0.043878515 -0.04666735  0.19011761 -0.01709057
#> 14  0.12247789 -0.408048453  0.16641973  0.05952963 -0.11763962
#> 15  0.01479176  0.339556409 -0.02727324 -0.04679590  0.13959945
#> 16 -0.05818244 -0.030281413 -0.27552719  0.16280246  0.10119930
#> 17  0.13693846 -0.176749998  0.35221749  0.04992092  0.14469637
#> 18  0.33085395 -0.451119809 -0.12199290  0.02621556 -0.36233767
#> 19  0.20461172 -0.164293341 -0.21319277  0.27293383  0.21323534
#> 20  0.16667032 -0.058429429 -0.06218116 -0.09165612 -0.32134835