Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>              F1          F2          F3           F4           F5
#> 1   0.134260391  0.06885217 -0.06854247  0.096841591 -0.223896515
#> 2   0.222598914  0.19310658 -0.13738820 -0.175244877 -0.188034106
#> 3   0.163755226 -0.24229339  0.51569342 -0.205070174  0.075485246
#> 4  -0.120892754 -0.03970504 -0.18014032 -0.028472808  0.090897566
#> 5   0.308781819  0.19003997  0.65327479  0.399717320  0.084274134
#> 6  -0.101393053  0.27718418  0.05951069  0.095026476  0.209792491
#> 7  -0.044615081  0.13098063  0.23773602  0.069332173  0.103860622
#> 8  -0.221756159 -0.17902082  0.32405184 -0.371553357  0.045898922
#> 9   0.052233859 -0.47629167 -0.10370833  0.232982649  0.175210687
#> 10  0.402437036  0.59262562  0.24358439  0.086039042  0.236000826
#> 11  0.199580023 -0.09834864  0.23804798 -0.003168326  0.186169820
#> 12  0.032438392  0.08162842  0.20489444 -0.151020751  0.268801555
#> 13  0.148706187  0.20142386  0.01220857  0.070245133 -0.139524285
#> 14 -0.082237537 -0.13749425 -0.03091044  0.430692935 -0.058585058
#> 15  0.007598442  0.08864502  0.18499455 -0.109254623 -0.005862424
#> 16  0.106226407  0.23283681  0.14073934 -0.086751210 -0.301339639
#> 17 -0.034575264 -0.14790270  0.14280269  0.178003913 -0.178364525
#> 18 -0.374102640  0.06534926  0.14272259 -0.286866576  0.004912132
#> 19  0.012594822  0.25536415  0.29178259 -0.342022926 -0.380645806
#> 20  0.065785213 -0.08259675  0.05380984 -0.133789060  0.208448175

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>              F1            F2          F3           F4          F5
#> 1  -0.361472222  0.2068814587 -0.11421460 -0.403044725 -0.08453992
#> 2  -0.032038477  0.4241532803  0.08783119 -0.439902312 -0.01978871
#> 3  -0.033957172 -0.2512350627  0.19505755 -0.161582850 -0.46680915
#> 4  -0.491726662 -0.0057893445  0.17963801 -0.182543242 -0.27826814
#> 5  -0.062162810 -0.1545371077  0.24186289 -0.253955208 -0.06554111
#> 6   0.198904265 -0.1501406976 -0.00590225 -0.414620291  0.10885179
#> 7   0.006986162  0.2364390835 -0.04093711 -0.163883056 -0.15395147
#> 8   0.176958437 -0.1748770562 -0.09014312  0.072358736  0.12650776
#> 9   0.208927917  0.1534083622  0.40336594 -0.225324377  0.18902356
#> 10 -0.142511027  0.1808072251 -0.02009609 -0.105864715  0.45576258
#> 11 -0.221313230  0.2116910890 -0.07780448 -0.252934007  0.04017298
#> 12  0.226627529  0.0827832168  0.03108498  0.118843231 -0.07374375
#> 13 -0.255158495  0.5027522166 -0.32636906 -0.049192181  0.05959173
#> 14 -0.116907730 -0.1039383175  0.36365575 -0.558959727  0.01321892
#> 15  0.180727507  0.1299626895 -0.15584827 -0.160305575  0.29169284
#> 16  0.013927807 -0.0626827968  0.03429629  0.372580386  0.33536240
#> 17 -0.246030811  0.1123823078  0.09730094  0.095980538 -0.10019111
#> 18  0.037173339 -0.0001905529  0.06830282  0.006313207 -0.27300168
#> 19 -0.153616780  0.0848594923 -0.27380138 -0.544605558  0.15293636
#> 20 -0.068121795  0.1686618433  0.23268617 -0.032182368  0.45916448