Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>              F1           F2          F3          F4          F5
#> 1   0.524923409 -0.027107911  0.16983488  0.31161164 -0.22548306
#> 2  -0.200434807 -0.100016119  0.16986266  0.01729010  0.31044787
#> 3  -0.247116959 -0.364609483  0.09864937  0.33519283  0.44132608
#> 4   0.393935369 -0.150332924 -0.05549504 -0.09021334  0.09235926
#> 5   0.382344611 -0.034609082  0.41739337  0.47202904 -0.21929311
#> 6   0.155304755 -0.148914648  0.03357467  0.41310280  0.24417759
#> 7   0.269173921  0.115515345  0.06019720  0.31525114  0.24502464
#> 8   0.300935585  0.014645677  0.04381968  0.04321915  0.02848476
#> 9  -0.035010115 -0.300189666 -0.21317864 -0.03247516 -0.23962688
#> 10  0.688378047  0.085769281 -0.35767457  0.28171880  0.39724691
#> 11  0.322771821 -0.002339031 -0.02581740  0.06696978 -0.07618337
#> 12  0.098030020  0.371629459  0.05830453  0.20195601  0.10294147
#> 13  0.239822956  0.155372425 -0.03312348 -0.05359883  0.39846339
#> 14  0.324903230 -0.307427501  0.34007076  0.09546729 -0.10432930
#> 15  0.132108601  0.221167356  0.28707719  0.36134942 -0.10618194
#> 16 -0.048857365 -0.002357368  0.13873302  0.12665486  0.45455313
#> 17  0.495609781 -0.403784794  0.11379924 -0.20047033  0.24076505
#> 18  0.062333484  0.214878929  0.07069458  0.31502628 -0.18424076
#> 19  0.007729284 -0.091440633 -0.25919253 -0.05014716 -0.27966823
#> 20  0.499320834  0.322623541 -0.07011995  0.07267119  0.01188836

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>             F1          F2           F3           F4           F5
#> 1  -0.01266197  0.03150405 -0.123451833  0.033307678 -0.041373907
#> 2   0.20447602 -0.18582331  0.015977173 -0.107864528  0.087242213
#> 3  -0.08993810  0.42309419  0.144547663 -0.348293117  0.009891004
#> 4  -0.17610256  0.08486327 -0.111810866  0.001197477  0.061491045
#> 5   0.13640590  0.52880369  0.284055991 -0.140312305  0.076669271
#> 6   0.13711222  0.55044751  0.262034552 -0.013072197  0.418858760
#> 7   0.37506462  0.47674384  0.094707230 -0.138123527 -0.215959087
#> 8  -0.26952631 -0.15035836 -0.127570674 -0.047677117  0.089019051
#> 9   0.22738855  0.01532234 -0.135171389 -0.038680273  0.106003025
#> 10 -0.08153631  0.21740591  0.352902506 -0.001410280  0.095821767
#> 11 -0.01666711  0.36540127 -0.045378418 -0.214019068 -0.008892573
#> 12 -0.02281386 -0.17282494  0.073065580 -0.095682702  0.107828867
#> 13 -0.12272222  0.27502448 -0.249597502  0.395001689 -0.220721068
#> 14 -0.37298711  0.23835455 -0.113990738 -0.214755987  0.246992336
#> 15  0.33257853  0.54440690 -0.000283363 -0.185868577  0.057373505
#> 16  0.26729691  0.14830289  0.081469044  0.536685958  0.004406877
#> 17 -0.17011060  0.31423378 -0.153069534 -0.063381019  0.212665740
#> 18  0.24310281 -0.02841014  0.242980494 -0.342410578  0.379727645
#> 19  0.07657461  0.19561742  0.029085888 -0.120994431  0.218222584
#> 20  0.44409204  0.05096094  0.001022103  0.137334969  0.301879734