Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>              F1           F2          F3          F4          F5
#> 1  -0.011905008  0.638730018 -0.03818602  0.03699500  0.10232320
#> 2   0.122534060  0.146710185 -0.06529248  0.01981144  0.07686027
#> 3  -0.156271352  0.216238511  0.19763203  0.19988944  0.05395006
#> 4  -0.251226424 -0.304706185 -0.09621000  0.48525615  0.17740635
#> 5  -0.198412261  0.150349921  0.33753708  0.20084606 -0.13300980
#> 6  -0.055995222  0.592214077 -0.03976553  0.10640967 -0.11399361
#> 7   0.183205616  0.447754866 -0.11968635  0.33018690 -0.09462141
#> 8  -0.122302529  0.020404343 -0.09580810  0.19340482 -0.20872266
#> 9   0.318760754  0.313732975 -0.15224512 -0.00241254  0.13695185
#> 10 -0.009004554 -0.132512479 -0.04886994  0.37396639  0.05100148
#> 11  0.146387990  0.304980507 -0.05084292  0.20218222  0.24323617
#> 12  0.214434332  0.247855614  0.26287741  0.17733243 -0.28375632
#> 13 -0.024101008  0.181724508  0.54202200 -0.16409320  0.10292351
#> 14  0.360345108  0.338382746 -0.08706363  0.06105293  0.15530907
#> 15 -0.088246646 -0.007061624 -0.02690994  0.32076420 -0.10297466
#> 16 -0.321304848  0.021618177  0.24008137  0.06456246  0.15128119
#> 17  0.112641011  0.057175334  0.20496223  0.25692991 -0.21239074
#> 18  0.115195662  0.012422809 -0.07106907  0.27365106  0.18609247
#> 19 -0.183959598  0.118502548  0.06071294  0.05902176 -0.14706455
#> 20 -0.179956849  0.012393753  0.11399909  0.04987242  0.19848045

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>             F1          F2            F3          F4          F5
#> 1   0.13328391 -0.10513002 -6.415945e-02 -0.13760708 -0.25420159
#> 2  -0.10242948  0.25253756 -4.510737e-01  0.01005832  0.01150702
#> 3  -0.33300678  0.17710358 -1.427794e-01  0.02743626  0.03323725
#> 4   0.09450878  0.06490217 -7.341829e-02  0.14980980  0.01639445
#> 5   0.42530726 -0.12927631  8.077346e-02 -0.54615117  0.12893419
#> 6   0.30788798  0.14876689 -3.102261e-01  0.55129838 -0.10060815
#> 7  -0.06689436  0.27053087  2.079210e-01 -0.47579032  0.14523019
#> 8   0.18247942  0.12476875  1.326034e-01 -0.20524680  0.14143520
#> 9   0.05659236 -0.28078398  1.520398e-01  0.03205190  0.20260199
#> 10 -0.15271503  0.02393932  2.160812e-01 -0.34700659  0.18575445
#> 11 -0.20010136 -0.19008710 -3.266724e-01 -0.02054999  0.03872146
#> 12  0.11327797 -0.05209112 -4.579629e-01  0.40277380 -0.06716310
#> 13  0.17474067  0.28071172  1.199475e-01  0.02770731 -0.04303240
#> 14  0.08761577  0.13846577 -1.006868e-01  0.25928562  0.21245226
#> 15  0.13263146  0.01906988  3.518918e-01 -0.06687293 -0.23948657
#> 16 -0.01880629  0.09408789 -9.911822e-05 -0.14093957 -0.18696444
#> 17 -0.03290009 -0.03607357  1.399217e-01 -0.35595541 -0.16858432
#> 18 -0.34192528 -0.14729979 -2.366607e-01  0.44446503 -0.13593443
#> 19 -0.06444525 -0.17855641  2.193952e-01 -0.24967270 -0.49901592
#> 20  0.20139827 -0.16637795 -2.436061e-01 -0.37882115 -0.10004479