Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>             F1          F2           F3            F4          F5
#> 1   0.50018827 -0.10352851 -0.052960363  0.2678875466 -0.30999273
#> 2   0.07759842 -0.50435317  0.337039818 -0.1410853476  0.10970555
#> 3  -0.31357430  0.18550372  0.115342731 -0.1298352060  0.55797724
#> 4  -0.03893154  0.08464636 -0.277596455  0.0277596738  0.36356707
#> 5   0.33723363 -0.31930172  0.148714689  0.2512066092 -0.28721230
#> 6  -0.03730205  0.01976595  0.052805559 -0.0165765163  0.26047404
#> 7  -0.17774108  0.17469923  0.434808717 -0.1039238020 -0.03228823
#> 8  -0.15195047  0.11933751 -0.097528161  0.1035951513 -0.06807832
#> 9   0.26586303 -0.35164753  0.164891084 -0.0037114492  0.66552552
#> 10  0.49871386 -0.06345417 -0.609777211  0.2240864342  0.44516652
#> 11 -0.22926658 -0.24927054 -0.008968972 -0.2108748808  0.26336465
#> 12 -0.19942186 -0.44407480 -0.058647426 -0.1566952753  0.16846823
#> 13  0.07066358  0.11281143  0.056649474  0.1057289425  0.15930280
#> 14 -0.20332605  0.16709406 -0.190544638  0.2515609831  0.03460544
#> 15  0.06319195  0.30172460  0.141097776  0.3783738928 -0.17280190
#> 16  0.04738950  0.16766768 -0.131326975  0.2440441380  0.09377255
#> 17 -0.14218122 -0.17772036 -0.228868449 -0.0005192386  0.28434879
#> 18 -0.26276820 -0.36360805  0.022359210  0.2065395554  0.17211031
#> 19  0.24548196  0.17022793 -0.036307657  0.0458933139 -0.17874016
#> 20  0.42603733 -0.03675648 -0.225341376  0.2161740333  0.23194800

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>             F1           F2            F3          F4          F5
#> 1   0.15273681  0.130530842 -0.1848430841 -0.45969677 -0.06220156
#> 2  -0.26121783 -0.333578575 -0.0856601142 -0.22644917 -0.33799473
#> 3   0.37526152  0.012614542  0.3489299466 -0.27727979  0.16446699
#> 4  -0.06867854  0.008525850 -0.1201356646  0.00587404 -0.30520389
#> 5   0.07934201 -0.177192668 -0.2930017783 -0.14340149 -0.29310406
#> 6   0.11135880  0.027649225 -0.1278349559  0.03255659  0.24205320
#> 7  -0.44684881 -0.071729080 -0.0561844329  0.04192664 -0.01520807
#> 8   0.09260378  0.006194008  0.1483801104  0.43317816 -0.05827229
#> 9  -0.06947042 -0.102746435 -0.1879903109 -0.03319347  0.11604051
#> 10  0.12615099  0.053947820  0.0610091680  0.48885744  0.23645515
#> 11 -0.18765521  0.010901150  0.0961549592 -0.41898242  0.11801860
#> 12  0.42979381 -0.065658642 -0.2856729571 -0.16211659  0.13147592
#> 13 -0.57867563 -0.137358886  0.0116281942  0.18651872  0.07137314
#> 14  0.20893684 -0.087578171 -0.2229719655  0.05389358  0.01167541
#> 15 -0.05405675 -0.190617938 -0.1944297463  0.12515288 -0.06932595
#> 16  0.23897711 -0.207741402 -0.3456644107  0.10652819 -0.02644717
#> 17 -0.26720541 -0.024478879 -0.1968094201  0.19475009 -0.01997724
#> 18 -0.34541554 -0.226186222  0.0824374109 -0.03580846 -0.10477417
#> 19  0.05020170 -0.008694760 -0.1534906424 -0.09739291 -0.06150772
#> 20  0.28993285  0.085701980  0.0003503337 -0.04894639 -0.17375851