Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>              F1           F2           F3          F4          F5
#> 1  -0.416677195 -0.030240251 -0.141553319  0.02817815 -0.09107499
#> 2  -0.248859863 -0.103212055  0.357142311  0.10497739 -0.13952758
#> 3   0.142598593  0.216206319 -0.169433369 -0.19731026 -0.19508891
#> 4   0.060127494  0.379271254 -0.027703429  0.35985833  0.16177316
#> 5   0.006639203  0.447611302  0.132500360  0.33416356 -0.29135349
#> 6  -0.168452543  0.265870024  0.641687019  0.54435735  0.05774761
#> 7   0.062836575  0.241589993 -0.076579549  0.52879486  0.26214133
#> 8  -0.117535202 -0.074489650 -0.504690269  0.12672614 -0.08936817
#> 9   0.068349173  0.094713205  0.337686282  0.04968668  0.33137459
#> 10  0.243675673 -0.006866845 -0.100211335  0.12497987  0.43959153
#> 11 -0.476292044  0.335244597 -0.004386604  0.02616215  0.02095051
#> 12 -0.186152778  0.186882390  0.643127385  0.46779581 -0.19501209
#> 13 -0.062330110  0.370152925  0.088270984  0.35840937 -0.13032935
#> 14  0.045692969 -0.256612337  0.028792834  0.32847335  0.34155153
#> 15 -0.255184066  0.284711483  0.135696937  0.40976073  0.37268774
#> 16  0.185218581 -0.172046504  0.011871839  0.50574728 -0.19984727
#> 17 -0.060614335  0.283037890 -0.023992262  0.48726780 -0.04912065
#> 18 -0.075286299  0.013936770  0.401740628  0.40924295  0.01856638
#> 19 -0.012274456  0.359604144  0.217144439  0.14419483 -0.42167095
#> 20 -0.216816757  0.374959081 -0.199871069  0.15573247  0.10362962

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>             F1           F2           F3           F4          F5
#> 1   0.09559969  0.008164781  0.107149744 -0.487973779 -0.09439769
#> 2  -0.06830293  0.153763377 -0.090567571 -0.050217279 -0.40524171
#> 3  -0.20089825  0.190719948 -0.146198677  0.018024507  0.12356608
#> 4  -0.45236202 -0.096991369  0.002903921 -0.034931365 -0.08657877
#> 5  -0.02975857  0.247586631  0.443177299  0.175941691  0.19757294
#> 6  -0.15495080 -0.062712947 -0.092232512  0.291319146  0.12640350
#> 7  -0.18882254 -0.083338831 -0.014681645 -0.177057008  0.02387525
#> 8   0.15837775  0.274540597 -0.086027950 -0.295247330  0.16902090
#> 9  -0.13585313 -0.072370738  0.657122726  0.052909262  0.36770281
#> 10 -0.11652841  0.531827115  0.303046789 -0.101292780 -0.08010810
#> 11 -0.14389870  0.000557521  0.177789662 -0.262916053  0.01356256
#> 12 -0.07447345 -0.118436724  0.267991575  0.740893377  0.21039968
#> 13  0.27128674  0.148054052 -0.050943608  0.033400890 -0.01862293
#> 14 -0.09532512 -0.158695636 -0.077788696  0.062884485  0.03136730
#> 15 -0.44879709  0.485549761 -0.219358703  0.003850283 -0.65237978
#> 16 -0.35629055 -0.301728888  0.496414632  0.435876150  0.08865358
#> 17 -0.34462344  0.017495514  0.025476714 -0.030166008  0.13443682
#> 18 -0.09766463  0.121425071 -0.260082024  0.158246121 -0.38488208
#> 19 -0.32587774  0.207696584  0.346452967 -0.029950836 -0.12180073
#> 20 -0.10651452 -0.057354571  0.292970865 -0.141194628 -0.34681372