Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>             F1          F2          F3          F4          F5
#> 1  -0.44396581 -0.50960821  0.44521328 -0.09770053  0.19779361
#> 2   0.19836692  0.28394335 -0.08031012 -0.51922149  0.28705168
#> 3  -0.14828872 -0.07140818  0.68064737  0.26048664  0.08963952
#> 4   0.04632205  0.40020759  0.15806088  0.03129221 -0.12639647
#> 5   0.33755045  0.17157959 -0.01355668 -0.20094255  0.15179069
#> 6   0.21722785  0.04239695  0.24485768 -0.06068072  0.31685301
#> 7   0.45136318  0.26013909  0.24718420  0.33535267  0.34573930
#> 8  -0.05059758 -0.18276555  0.14466080  0.16930475  0.17045871
#> 9   0.32840651  0.21439863  0.65551072 -0.07000846 -0.19508160
#> 10  0.15328381 -0.06551832 -0.29128753 -0.30113902 -0.06946447
#> 11 -0.02630774  0.18617672  0.28477312  0.21727088  0.06039769
#> 12  0.33902976  0.23489949 -0.06362375  0.01758543 -0.29224505
#> 13  0.27388911  0.11625301  0.50640086  0.06492469  0.24160926
#> 14  0.15696908 -0.11715021  0.38554698 -0.01077783  0.08172497
#> 15  0.21872640 -0.08520612  0.08375413  0.49815924 -0.22868599
#> 16 -0.16272667  0.12336792  0.42885241  0.06363075 -0.07825940
#> 17  0.26248106  0.33507917  0.20327736  0.32884948  0.37402828
#> 18  0.24615051  0.40284117  0.49777530 -0.06869328  0.26444720
#> 19  0.42466799  0.07774676  0.42707271  0.26049874  0.07735685
#> 20 -0.16884380  0.22022691  0.36517303  0.15716462 -0.03671480

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>               F1          F2          F3          F4           F5
#> 1  -0.1018982365  0.29882807 -0.27140920 -0.18039316  0.052531985
#> 2   0.2940803070  0.34894911 -0.19760172 -0.31854592  0.147804092
#> 3  -0.0866274641 -0.04704703  0.36507461  0.16909461  0.288678190
#> 4   0.2264308353 -0.45054827 -0.14049575 -0.34590057  0.389833718
#> 5   0.0151806240  0.39716068 -0.12193430  0.31923682  0.279931589
#> 6  -0.2636794882  0.13111318 -0.59403188 -0.17977473  0.295699739
#> 7   0.2120152064 -0.20202350  0.39028813 -0.28120128  0.470056421
#> 8  -0.2960293331  0.27368247  0.20664072  0.26069896 -0.212168230
#> 9  -0.1205503901  0.24424172  0.02458879 -0.31233567  0.368516988
#> 10 -0.2653849039 -0.02381985 -0.32195073  0.17872148 -0.008980461
#> 11 -0.0721349372 -0.08015807 -0.22080949  0.18964535 -0.013628829
#> 12 -0.1374204174  0.04208102 -0.23254432 -0.22256464  0.142539719
#> 13 -0.1385523930 -0.04779562 -0.43994198 -0.05507586  0.303967597
#> 14  0.2118293055  0.35526431  0.03897595 -0.24048200 -0.069765068
#> 15  0.1622483149  0.04367921  0.26549873  0.58227678 -0.112120438
#> 16 -0.2645058157 -0.13938609 -0.60611914 -0.14292183  0.317423370
#> 17 -0.0003190798 -0.49058572 -0.30606841 -0.10380016  0.390737894
#> 18 -0.1525116212  0.38692138 -0.10264325  0.11090332 -0.123887174
#> 19  0.3838415162 -0.08328281 -0.17426651  0.10062214  0.434387259
#> 20  0.3058252336 -0.23074533 -0.13948352  0.01718978  0.260282581