Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>              F1          F2          F3           F4           F5
#> 1   0.082560175  0.13247936 -0.19662295 -0.087232304  0.550897647
#> 2   0.250479310  0.17901283 -0.03206165  0.197423639 -0.001232598
#> 3   0.022158948  0.17621692  0.23593380  0.130317133 -0.370820737
#> 4  -0.007310331  0.22342003  0.27398193  0.370690400 -0.245899598
#> 5  -0.149575615  0.02477130 -0.44296864  0.003779628  0.208539664
#> 6  -0.182070840  0.47294059  0.10945549 -0.001251017  0.453027935
#> 7   0.342529740  0.09103138  0.01844433  0.127460035  0.288760017
#> 8   0.433533960 -0.15779726  0.18422587 -0.394415230  0.251845505
#> 9  -0.254965816  0.83954502 -0.37165047 -0.096564459 -0.193473409
#> 10  0.371600386  0.11353031 -0.08229424  0.311370540 -0.020223008
#> 11 -0.085822608  0.27669910 -0.21456490 -0.181224751 -0.009366760
#> 12  0.159613160  0.43282019 -0.08549381  0.375447735 -0.073494419
#> 13 -0.072802758 -0.01463266  0.07001770  0.024665391  0.128252382
#> 14 -0.157633255  0.04388985 -0.05409940 -0.223665589 -0.347952571
#> 15  0.270396834  0.25352427 -0.10203846 -0.223603667 -0.008193776
#> 16  0.091981913  0.05681676  0.08057109  0.046762831  0.227074243
#> 17  0.002980450  0.18678109  0.15479147 -0.425166416  0.268360175
#> 18  0.130921851  0.28295161  0.18368059  0.296619046 -0.375054197
#> 19 -0.141887815  0.13031674 -0.05597187 -0.292706651  0.158050430
#> 20 -0.094399940  0.04251802  0.10700917 -0.401858618  0.020166036

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>             F1           F2           F3            F4           F5
#> 1  -0.30691723  0.112936929 -0.169320655 -0.0675419243  0.217617848
#> 2  -0.17931262 -0.120689626 -0.086631110 -0.0829820367  0.432408653
#> 3  -0.20735835 -0.076200391  0.128760105 -0.1104123230  0.217964298
#> 4  -0.11571181  0.004218342  0.086189983  0.0617914858  0.264945178
#> 5  -0.03644221  0.059506309  0.148148719  0.0079420268  0.073901544
#> 6   0.05568491 -0.070782484 -0.242608899 -0.1292176870  0.101714646
#> 7  -0.17150704 -0.140650583  0.452267769 -0.0504947720 -0.105710840
#> 8   0.05769262 -0.201773859  0.005260265 -0.0004696693  0.330383723
#> 9  -0.09150798  0.036942876 -0.324098121  0.0157541279  0.070661327
#> 10 -0.29526425  0.289980399 -0.254825487 -0.1500962007  0.006030233
#> 11 -0.22000898  0.083547858  0.186529025 -0.1158718421 -0.082956179
#> 12 -0.19260162 -0.166946560 -0.166322781 -0.0975290327 -0.056797761
#> 13 -0.09331239 -0.319847547  0.098726043 -0.2307244460  0.341748652
#> 14 -0.48734316 -0.260018563 -0.102756408  0.2046792109  0.026327118
#> 15  0.01336539 -0.196476245 -0.033055719  0.3877943953 -0.237781261
#> 16 -0.15868733 -0.065492102 -0.060828889  0.1746818706 -0.003577880
#> 17 -0.15959653 -0.219217665 -0.084206225 -0.4380295298  0.052085702
#> 18 -0.34289713  0.037258188 -0.123152668 -0.0859688641  0.042699819
#> 19 -0.69373045  0.494238172 -0.158866492 -0.5451223819 -0.024759528
#> 20 -0.24599708  0.250476073 -0.669872692 -0.1379304269 -0.149278801