Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>             F1           F2           F3           F4          F5
#> 1   0.31973467  0.234913503 -0.012461435  0.129009278  0.24010394
#> 2   0.11464890 -0.157884736 -0.358767459  0.117143967  0.17529624
#> 3  -0.12986800  0.216325455 -0.041472027 -0.304795054  0.21642393
#> 4  -0.32601610  0.307540314 -0.044920355  0.221089152 -0.28908057
#> 5  -0.11170241  0.313991929 -0.117537225  0.055186457  0.21163274
#> 6  -0.32338391  0.002366019  0.360245297 -0.354837864  0.04011300
#> 7   0.15889468 -0.253474880 -0.280120283  0.004122666  0.04251915
#> 8   0.08547589 -0.144506042 -0.045497101 -0.042567818 -0.32977589
#> 9  -0.01974396  0.107761639  0.076003641  0.296158639 -0.23594628
#> 10  0.16821266  0.553721213  0.423777143  0.175946406 -0.04148388
#> 11 -0.14532573  0.030783048 -0.136760475  0.022735827 -0.36759976
#> 12  0.45619495  0.228330747  0.002548617  0.429636225  0.17810286
#> 13 -0.04990927  0.418100804 -0.076514576  0.593287092  0.39368838
#> 14  0.10117443  0.362494670 -0.262066206  0.259533289  0.06307516
#> 15 -0.06621368 -0.187904028  0.086468785  0.121593467  0.20016703
#> 16 -0.01639962  0.286317544  0.167712163  0.064715820 -0.05858775
#> 17  0.27415479  0.224645205  0.362152853  0.104980147  0.25934071
#> 18 -0.09154809  0.221519891  0.055415989  0.334980947 -0.29122520
#> 19 -0.07800345 -0.290474952  0.268860564 -0.094794858  0.40539341
#> 20  0.20501299  0.248553877  0.181425460  0.042569078  0.14752163

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>             F1          F2          F3          F4          F5
#> 1  -0.15310617 -0.01501286  0.29626244  0.11299698 -0.15399328
#> 2  -0.05691570 -0.10230891 -0.28835461  0.09132194 -0.25583635
#> 3  -0.06582397 -0.06354893 -0.01026532  0.03772817 -0.16625205
#> 4   0.18083863 -0.49999812 -0.16500282  0.13043363 -0.47812561
#> 5  -0.05027523 -0.19388204 -0.06683621 -0.14758622  0.08453954
#> 6   0.06107452 -0.14741605 -0.09199445 -0.05169064 -0.02894394
#> 7  -0.40838829 -0.37661082 -0.38546239 -0.07714767 -0.03170103
#> 8   0.01062406 -0.09230667 -0.40574644  0.01338968 -0.03074054
#> 9  -0.27951559 -0.03096983  0.12730970  0.36118320 -0.29322337
#> 10  0.02099335 -0.03713966 -0.32491828 -0.26144311  0.29649363
#> 11 -0.31126318 -0.15184900  0.12293830  0.38395449  0.24004073
#> 12  0.03885801 -0.16685254  0.06419892  0.32022186 -0.28067564
#> 13  0.15670321 -0.32849014 -0.25594018 -0.03352516  0.19492157
#> 14  0.42053270 -0.18116540  0.33469840 -0.17731351  0.07145132
#> 15  0.48195450  0.14419863 -0.47028501  0.11680286  0.09820248
#> 16 -0.22775442 -0.35468415 -0.43210014  0.03733068 -0.24338284
#> 17 -0.10878863 -0.01550320  0.11551541 -0.33767759  0.19951084
#> 18  0.20951175 -0.13083337 -0.27632598  0.08613722  0.09004203
#> 19 -0.05858961 -0.37335271 -0.27973261  0.02217102  0.13907816
#> 20 -0.11873442 -0.22770148 -0.32270876  0.21972746 -0.29186996