Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>             F1          F2          F3           F4           F5
#> 1  -0.29196524  0.07748438 -0.45558750 -0.033239587 -0.373057966
#> 2  -0.13104341  0.07612678  0.52247075  0.524610278  0.172767004
#> 3  -0.21913413 -0.23776753  0.15350796 -0.209570077  0.018702858
#> 4  -0.17559038  0.01116228  0.37342713 -0.361708582  0.217557104
#> 5   0.09072907  0.18510957  0.30655678 -0.159143681  0.250697893
#> 6   0.15833446 -0.01471752  0.66380370 -0.186255809 -0.053870108
#> 7   0.52975842 -0.16385584 -0.01986406  0.110877737  0.321678043
#> 8   0.18524024 -0.10531416  0.17174857  0.071461291 -0.121621937
#> 9   0.31465292 -0.22626411 -0.02601511  0.448383054 -0.118450203
#> 10  0.38942590  0.26651462  0.70552384 -0.107926940 -0.262059431
#> 11 -0.09892202 -0.18442072  0.23664242  0.229587782  0.038595629
#> 12  0.19928313 -0.16406573  0.18272863  0.095333058 -0.237503844
#> 13 -0.05153133 -0.01083177 -0.11480542  0.027170613  0.280536496
#> 14  0.77779430  0.17058235 -0.04558974 -0.036715510  0.006149563
#> 15  0.26125234 -0.14567107  0.01398588 -0.195718634  0.119561822
#> 16  0.14278867 -0.20711753 -0.11876505  0.097381621  0.200725516
#> 17  0.23667235  0.22598453  0.11901861 -0.016192666 -0.342515431
#> 18  0.05864264  0.31618074  0.24319890  0.007498009  0.196946598
#> 19 -0.17072972  0.37279904  0.41625501  0.347417897 -0.124199708
#> 20  0.53180755 -0.14198408  0.55349618 -0.013510315 -0.506171556

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>              F1           F2          F3          F4           F5
#> 1  -0.261041610  0.108918735 -0.21139768  0.11296193 -0.154261081
#> 2  -0.347674324  0.173515718  0.11970708 -0.24644363  0.009992543
#> 3  -0.252482964  0.386466347 -0.23313618 -0.03912742 -0.169040152
#> 4  -0.288684478  0.164490539  0.11882568  0.13810813  0.227339233
#> 5  -0.149058856 -0.008528007 -0.42081866  0.19905288 -0.120927747
#> 6  -0.092405019  0.213716800  0.17862622  0.10095678  0.358508924
#> 7  -0.246231454  0.175627485  0.01437501 -0.17716898 -0.231275967
#> 8   0.048901849 -0.246125167  0.28016037  0.13700335 -0.012625808
#> 9   0.420870404  0.024942794 -0.16552759 -0.32644241  0.132374379
#> 10 -0.020357577  0.222839136 -0.14474945  0.26854657  0.210798667
#> 11 -0.054813994  0.092330608 -0.14708503  0.22155438  0.120888632
#> 12  0.254216139  0.271685382 -0.70383744  0.03626192  0.079187901
#> 13 -0.002381215  0.097731435  0.02125065 -0.15687176  0.313928726
#> 14 -0.490921698  0.052276585  0.04971103  0.41383873  0.215603297
#> 15 -0.152603361 -0.092637995 -0.32253466  0.58523657 -0.018624776
#> 16 -0.044115017 -0.048007875 -0.06450985 -0.01744344 -0.046242662
#> 17 -0.329372442 -0.052017998  0.06840053  0.15888547  0.114888876
#> 18 -0.157554956 -0.182521727 -0.73972703  0.20258952 -0.150384369
#> 19  0.097451624  0.084702044 -0.06476218  0.29125957 -0.163104894
#> 20 -0.307671070  0.092133682 -0.37383612  0.30082660 -0.109274860