Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>              F1          F2           F3           F4          F5
#> 1  -0.222740191  0.20211407 -0.016734490 -0.006462172  0.45472180
#> 2   0.580283207  0.06706329 -0.024064267  0.316955204  0.56717797
#> 3   0.159548608  0.32586313  0.402446266  0.251623045  0.20110251
#> 4  -0.408122843  0.20119057  0.035786203  0.243861386 -0.04021654
#> 5  -0.234365246  0.03690557  0.174632181  0.467400496 -0.05821230
#> 6  -0.094661658 -0.16839359 -0.042191451  0.052327536 -0.18517784
#> 7  -0.119082734  0.11714122 -0.005008864  0.413141507 -0.15372993
#> 8  -0.145413718  0.12913926  0.255837956  0.308178697 -0.34348053
#> 9  -0.212217628  0.11044534  0.215028707  0.141084226 -0.06081637
#> 10  0.035138095  0.05368090 -0.307694700  0.341491714 -0.26676340
#> 11  0.286894055  0.23187100  0.007686110  0.578740893  0.44693492
#> 12  0.435502685  0.21432103  0.242449852  0.588809524  0.16712606
#> 13  0.253608468 -0.09214066 -0.098905306  0.277616430  0.25755314
#> 14  0.078888999  0.16364136  0.071199602  0.064557345  0.08320137
#> 15  0.350104363  0.45353181  0.018388933  0.530973285 -0.32609301
#> 16  0.207298170 -0.21875552  0.001213982  0.380993208  0.24624390
#> 17  0.800900959  0.10126766 -0.217461760  0.227253140  0.22024101
#> 18  0.007476202  0.26281270 -0.085172497  0.316454411  0.09827690
#> 19 -0.020151748  0.36071573 -0.132185261  0.203064392 -0.24954161
#> 20  0.127970153  0.09667765 -0.007999662 -0.396229771 -0.18147435

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>             F1          F2           F3           F4           F5
#> 1   0.08869153  0.16823809  0.212883791  0.348317590 -0.310964177
#> 2   0.20631049  0.21523477  0.475985472  0.621552674  0.117790636
#> 3   0.10490653  0.01751756  0.310387351  0.269430381  0.006733982
#> 4  -0.13525803  0.02068803  0.003210576  0.201435633  0.292351915
#> 5   0.02798718  0.14291066 -0.022834655  0.539014525  0.419967408
#> 6  -0.38266982  0.06170863  0.049655325  0.193900416  0.283607891
#> 7  -0.17418534  0.06871491 -0.045753786  0.008173169  0.042268116
#> 8   0.30463216 -0.10123770 -0.015522164  0.520169917  0.052166921
#> 9   0.06869663  0.25326360  0.128908112  0.312087861 -0.052957015
#> 10  0.46005280  0.05160237 -0.113101443  0.103236185 -0.020293152
#> 11  0.02826636  0.12107207  0.043873116  0.552414977  0.406659742
#> 12 -0.31629373  0.11634808 -0.078977482 -0.072438884  0.227963190
#> 13  0.15740841  0.35345794  0.057332129  0.199578281  0.184723433
#> 14 -0.08559374 -0.29012323  0.300854695  0.268271592  0.034617603
#> 15 -0.13436961 -0.02012752  0.257726748  0.096135480  0.344943203
#> 16 -0.10860380  0.16765270 -0.172420030  0.299067811  0.012463410
#> 17 -0.05216870  0.26705119 -0.298606143 -0.291950459  0.149570831
#> 18  0.06129000  0.13354347 -0.180166986  0.345116626  0.523316988
#> 19  0.45087402 -0.01273509  0.099684555 -0.066022048 -0.066519674
#> 20 -0.41380804 -0.01945913 -0.324173999  0.055180373  0.109348965