Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>             F1          F2           F3          F4          F5
#> 1  -0.38036615 -0.21980986  0.071327776 -0.38111993 -0.00830205
#> 2   0.17481656  0.16794574  0.213239231 -0.41819745  0.17214745
#> 3   0.11308585 -0.06362724  0.079597584  0.08797572  0.23047917
#> 4   0.12568711  0.34829648  0.063050306  0.41013386 -0.33917919
#> 5   0.26814509  0.25890797  0.091017762  0.22509976 -0.36469504
#> 6   0.07392654  0.15698104 -0.046600838 -0.22697576  0.06924360
#> 7  -0.40425352  0.57654153  0.295548187  0.37463062  0.24269081
#> 8   0.01619801 -0.01472252 -0.302254959  0.08816421  0.09364567
#> 9   0.02243983  0.32234927  0.109377789  0.12726790  0.16434492
#> 10  0.53299951  0.38768583 -0.052832801 -0.02099343  0.09830497
#> 11  0.41253944 -0.20262204  0.015633839 -0.07449286  0.18835461
#> 12 -0.45811477  0.38556840 -0.136485602 -0.14122672 -0.07734020
#> 13  0.40937991 -0.13388338 -0.170063906 -0.42030557  0.21490675
#> 14 -0.10132239  0.15445868  0.373432028 -0.14406240 -0.23885711
#> 15  0.18708900  0.35744399 -0.140225911  0.27795404 -0.16810488
#> 16 -0.10629726  0.52072659  0.152843138 -0.03307797  0.22416492
#> 17 -0.04051268 -0.10774090  0.334237788  0.29140406 -0.02645971
#> 18  0.26094943  0.41994347  0.558097105 -0.23051327 -0.04388799
#> 19  0.04360058  0.30943131 -0.008552963  0.06077279 -0.13012539
#> 20  0.22898718  0.08183893  0.479057260  0.22118647 -0.24244520

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>             F1          F2           F3          F4          F5
#> 1  -0.23892295  0.37230571  0.179989066  0.12388217  0.12376907
#> 2   0.17330824 -0.21292902  0.042770165 -0.15088998 -0.11695913
#> 3   0.01676040  0.38442128  0.223011689 -0.08881750 -0.22646738
#> 4  -0.24579430 -0.14488435 -0.242674995 -0.20166006  0.27531204
#> 5  -0.04787825  0.19250188 -0.192779520 -0.17748719  0.20456854
#> 6   0.19792636  0.07428840 -0.118019135 -0.11677395 -0.12426012
#> 7   0.01482983 -0.47443535 -0.261469349 -0.10848903 -0.23877113
#> 8   0.03003430  0.18026173  0.034087554 -0.32608381 -0.04308847
#> 9   0.07058533  0.06331160 -0.111907621 -0.12257971 -0.09385634
#> 10 -0.01584626  0.05571147 -0.447597644  0.09444494  0.09705816
#> 11  0.14308858 -0.19150173 -0.103333941  0.01427285  0.23839866
#> 12 -0.08872704  0.41456096 -0.102736799  0.03800403  0.14496181
#> 13  0.17995201  0.26290901  0.371409415 -0.24098304  0.43902958
#> 14  0.02329888 -0.18906483  0.099111965 -0.38945952  0.16135189
#> 15 -0.18202937 -0.06509083 -0.353055424 -0.07630949 -0.27831400
#> 16 -0.38028841  0.08781993  0.034138675  0.26099359  0.01658169
#> 17 -0.19381798 -0.03669491 -0.009611462 -0.06331584  0.15197097
#> 18 -0.13857127  0.27159476  0.166418399  0.08098225  0.14194666
#> 19  0.03892712  0.02048004 -0.179770421 -0.03733251  0.16395151
#> 20  0.08144320  0.16337412 -0.281436442 -0.03800741  0.08021853