Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>             F1            F2          F3           F4           F5
#> 1   0.24278415  0.4153749728  0.07257390 -0.097383675  0.118922779
#> 2  -0.47216906  0.1153913614 -0.01789554 -0.075059344  0.093455638
#> 3   0.49079198  0.1847260892  0.23802123 -0.177001312 -0.140452292
#> 4  -0.11097712 -0.3802202818 -0.15402983  0.337133180  0.111773585
#> 5  -0.19826077  0.2666014838 -0.10139952 -0.020183831  0.091185798
#> 6   0.08040509 -0.0498662832 -0.15737951  0.315663726 -0.024124961
#> 7   0.10659921 -0.0005990011  0.04768837 -0.075444307  0.203960072
#> 8  -0.28254580 -0.2176888085 -0.09217525  0.157057442  0.191688505
#> 9   0.34957249  0.0066677001  0.14511333  0.001687393 -0.443038476
#> 10 -0.36820316  0.3259796703  0.22053979  0.158327509 -0.015726437
#> 11  0.33172872 -0.6377648437  0.04015996 -0.223340171 -0.290311143
#> 12 -0.22590087  0.3273022479 -0.00919882 -0.484000025 -0.071073320
#> 13  0.30516764 -0.2733804680  0.45898982  0.487206057 -0.047877940
#> 14 -0.18882878  0.0916783260 -0.14459880 -0.108216350 -0.034722903
#> 15  0.17613277  0.2834408751  0.20501039  0.208716789  0.438267971
#> 16 -0.07265103  0.0548121102  0.35856644  0.034730071 -0.217258859
#> 17  0.16882833  0.3624723330 -0.04259558  0.209204614  0.079426707
#> 18 -0.18255261 -0.0891678661  0.10984542 -0.093106411  0.303157527
#> 19  0.21728633 -0.1274478305 -0.22935147  0.207323269 -0.005307477
#> 20  0.29045760  0.3602251332  0.44401923  0.329927074  0.440654190

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>             F1          F2           F3           F4          F5
#> 1   0.40103075 -0.43485840  0.370745120 -0.187316175 -0.14236763
#> 2  -0.06073891 -0.06502427 -0.006173824  0.009009563 -0.23330581
#> 3   0.41479055 -0.26211354  0.114708621 -0.229117308 -0.21233212
#> 4  -0.27594192  0.25998065  0.021008376 -0.028977667 -0.29700550
#> 5  -0.04007132 -0.59571024  0.440569341 -0.143853971  0.21379966
#> 6   0.29826155 -0.55763938  0.379375075 -0.007750335 -0.34664509
#> 7  -0.20615276 -0.11574713  0.196243912 -0.126638903  0.36689381
#> 8  -0.05066656 -0.24692063  0.101953988  0.068454382 -0.17059727
#> 9   0.61902103  0.12792093 -0.037727946 -0.208935670 -0.52620099
#> 10  0.04333604 -0.23346100  0.412008487  0.080158289 -0.18081302
#> 11  0.20014141  0.08229970  0.426963531 -0.049447468 -0.26821297
#> 12  0.23098480 -0.28265471  0.260298508 -0.345082249 -0.17182670
#> 13 -0.33037429 -0.08670636  0.103794993  0.230594074 -0.46486930
#> 14  0.14002613 -0.48733471  0.035548246 -0.066758475 -0.21346143
#> 15  0.39621018  0.21986498 -0.156959357 -0.255086340 -0.05148005
#> 16  0.24977291 -0.14219109  0.053729554 -0.190876567  0.13851718
#> 17  0.05746184 -0.20561334  0.246529312 -0.248951879 -0.40185774
#> 18 -0.39608793 -0.26271245  0.457060688 -0.042555663 -0.54216706
#> 19  0.12518398 -0.10783761  0.058677248 -0.171676457 -0.24709415
#> 20  0.23176652 -0.65093796  0.110309109 -0.418312438 -0.16535564