Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>              F1            F2          F3          F4          F5
#> 1  -0.070647321  0.0257979643 -0.20222107 -0.04176834  0.13621989
#> 2  -0.095414510 -0.2737771374  0.25150182 -0.24891775 -0.33447739
#> 3  -0.218303502  0.2469698165 -0.27667044  0.01366088  0.08285213
#> 4  -0.290910867  0.2406419671  0.07090453  0.02669455  0.22159413
#> 5   0.231546708 -0.2103251449  0.13383103  0.18352374  0.44304212
#> 6  -0.139613484 -0.4657161513 -0.06203014 -0.01751919  0.36423846
#> 7   0.052344941 -0.1055108661  0.51778685  0.28243489 -0.10767272
#> 8   0.119357029  0.2271558434  0.11030787  0.07289206 -0.24233743
#> 9  -0.081295047  0.0478923732  0.57155436 -0.38380097 -0.15635004
#> 10  0.088169079 -0.1629035497  0.20822820 -0.16779519  0.12796361
#> 11  0.294927466  0.0637521864  0.08443069 -0.12988183 -0.24916999
#> 12  0.194164382  0.0104327929  0.08324025 -0.03151312  0.05259462
#> 13  0.040950417 -0.0172820936  0.38141964 -0.17280860 -0.10858973
#> 14  0.219197881  0.1638429826  0.03958098  0.30569058  0.36099756
#> 15  0.129154665  0.3466516556 -0.13009897  0.14861542 -0.12293408
#> 16 -0.001795029  0.1055081359  0.19764826  0.04870288  0.17270270
#> 17  0.499078771 -0.0007102489 -0.15439844  0.04400699  0.07264588
#> 18  0.410417644  0.1016035181  0.20205769  0.02087216  0.18507019
#> 19 -0.076124669 -0.1073879269 -0.52390256  0.08381648  0.19870981
#> 20  0.186413901  0.0591000032  0.13074355  0.26367211  0.25374526

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>             F1           F2           F3           F4          F5
#> 1  -0.08624374  0.370515895 -0.127508093 -0.329202580  0.17250030
#> 2  -0.02376382 -0.007357015  0.038496333  0.016953547  0.42145391
#> 3  -0.13033609 -0.174582487 -0.202595851  0.035059486 -0.00714043
#> 4  -0.23154613 -0.061804934 -0.014969924 -0.298179659 -0.12770718
#> 5  -0.15049009 -0.323325120  0.495172259  0.121434513 -0.18722541
#> 6   0.22843037  0.164606639  0.044783746  0.001006937  0.06475603
#> 7   0.07960207 -0.224173723  0.023727981  0.079079257 -0.20161899
#> 8  -0.11945980  0.086724025 -0.179353860 -0.200787579  0.50551707
#> 9  -0.20601945 -0.120996014 -0.103617604  0.196122384 -0.10190384
#> 10  0.16307890 -0.159939463 -0.098679660 -0.020728834  0.19752588
#> 11  0.11488707  0.066425275  0.130461297 -0.036790469  0.40443480
#> 12  0.02970226  0.019061322  0.117568683  0.240891687  0.27922304
#> 13  0.39805647  0.054330920  0.235846774 -0.134039878  0.10961294
#> 14  0.01241368  0.270273700  0.141187304 -0.334425487  0.22517102
#> 15 -0.28800016 -0.344443918 -0.009824708  0.005716404 -0.01342264
#> 16  0.17653838  0.199825184 -0.135861101 -0.243494396 -0.06252133
#> 17  0.45246836  0.062710763  0.235776300 -0.096731275  0.31762491
#> 18  0.39688410  0.243949655  0.094738481 -0.196913775  1.00897141
#> 19  0.05268535 -0.344457356  0.079779135 -0.141352656 -0.10120071
#> 20 -0.07765242 -0.332619694  0.172162040 -0.070443874  0.40159590