Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>             F1           F2           F3          F4          F5
#> 1  -0.01465518  0.084388434  0.690757883  0.55912806  0.06967317
#> 2   0.14563697  0.100193787  0.230350625  0.26245820 -0.26649953
#> 3  -0.15370850  0.007137513 -0.385592620 -0.21332057  0.45582497
#> 4  -0.13194587  0.050259875 -0.057564304  0.30252206  0.29934090
#> 5   0.27947975  0.004798876 -0.295152528  0.13625508  0.01247160
#> 6  -0.47783701 -0.121343975 -0.320638938  0.39371711  0.20988014
#> 7  -0.08288434  0.224762634  0.069580178  0.02274156 -0.22480129
#> 8   0.53828617  0.143730685 -0.242955107  0.23520317  0.33169643
#> 9   0.54742708  0.243315971 -0.380082959 -0.04123563 -0.13441918
#> 10  0.13426709  0.456186870  0.037941639 -0.07573616 -0.10735378
#> 11 -0.31610692 -0.204149948 -0.221773001  0.22072076 -0.13716620
#> 12  0.29751683 -0.461074104  0.155417735  0.15859934  0.15040945
#> 13 -0.23627963 -0.242283592  0.147888952 -0.02759048  0.36180645
#> 14  0.15022718  0.020344979  0.163194950  0.16303121  0.19066366
#> 15 -0.17624802  0.173441331 -0.637334061  0.12663445 -0.06180135
#> 16  0.31903526 -0.496057568  0.290177313  0.07519433 -0.04839320
#> 17 -0.09102012  0.111088268  0.202194900  0.38496358  0.03073651
#> 18  0.20889272  0.026833361 -0.371110761  0.14334706  0.14082330
#> 19 -0.04056082 -0.032320581  0.013951965 -0.45646178  0.17510905
#> 20  0.04536413  0.171901469 -0.003955219 -0.09050180 -0.06532294

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>             F1          F2          F3          F4          F5
#> 1  -0.54946904 -0.07917722 -0.02907146  0.27042870  0.33797648
#> 2  -0.23508874 -0.06958596  0.13887605 -0.18857568  0.26101572
#> 3  -0.06424454 -0.08268244 -0.47543447 -0.12514405  0.11085554
#> 4  -0.17269699 -0.08392128  0.20587224 -0.29022369  0.12768948
#> 5  -0.23407931  0.32623418 -0.30838890 -0.08652141  0.02796327
#> 6  -0.16714712 -0.03045301  0.24190999 -0.17333114 -0.06050789
#> 7  -0.43522317  0.26639189 -0.24675168 -0.13551844  0.12682631
#> 8  -0.42592602 -0.13161318  0.08670047  0.21660571 -0.30585501
#> 9  -0.30601723  0.41210671 -0.08829194  0.06181822  0.18232026
#> 10  0.26942679 -0.30833236 -0.05130066  0.14553319  0.38082075
#> 11 -0.24479354  0.47273210 -0.08272606 -0.11320748 -0.17928948
#> 12 -0.13199292  0.23990817  0.18440538 -0.17053572 -0.07321625
#> 13 -0.17819224 -0.13016783 -0.04000753  0.20744917 -0.08130627
#> 14 -0.22040999  0.31599928 -0.26536902 -0.43677106  0.04716789
#> 15  0.07716426  0.36302996  0.08792107 -0.50435787  0.25038265
#> 16  0.02577078 -0.18218427  0.11085420 -0.43067428  0.12737458
#> 17 -0.29840289 -0.31409883  0.27751811 -0.32056125  0.13343711
#> 18  0.44678947 -0.05421929  0.13623015 -0.31915515  0.28830693
#> 19 -0.25917595  0.41111383 -0.09975899 -0.22372515 -0.12773542
#> 20 -0.44027962  0.17671043 -0.02918553 -0.22316246 -0.12731266