Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>             F1          F2          F3          F4            F5
#> 1   0.73336531  0.31032532  0.06849378  0.08880606 -0.0004243534
#> 2   0.22424657  0.31988718  0.06976589  0.31357712 -0.0654961993
#> 3  -0.09296293  0.32669439 -0.02482513  0.42290235 -0.1487509789
#> 4   0.45224080  0.45191314  0.36488784  0.08717540 -0.2156812287
#> 5   0.29457432  0.32634745  0.27555160 -0.17801800  0.2450550909
#> 6  -0.17681326 -0.11936450 -0.11516747  0.05966582  0.0961931100
#> 7   0.37639918  0.02129797  0.34465329 -0.05380258 -0.1406466684
#> 8  -0.07490954  0.10969511 -0.06801741  0.06892758  0.3417112184
#> 9  -0.07046089 -0.14603235  0.19424764 -0.20948146  0.1317570400
#> 10  0.09258927 -0.12767482  0.23003975  0.22201808 -0.0416473273
#> 11 -0.04575005  0.05422567  0.10568004 -0.07584574  0.5007890715
#> 12 -0.09354828  0.48119265  0.26246214  0.30685968 -0.0251387654
#> 13  0.39058455  0.32605605  0.06430739  0.15994946  0.2818004430
#> 14  0.18245303 -0.06296096  0.34856351  0.11656812 -0.0126576717
#> 15  0.20411769  0.07349657  0.58091649 -0.07112746 -0.1383135534
#> 16 -0.02454446  0.38030573 -0.24618966 -0.50649568 -0.0332369766
#> 17 -0.06953061 -0.20666940 -0.19246480  0.08985642  0.0143833130
#> 18 -0.14541418  0.37151597 -0.29004368  0.03938203 -0.1529944713
#> 19  0.21625520  0.14903083  0.06554499  0.09041736  0.1735634511
#> 20 -0.08509182 -0.28174930  0.51227217  0.05240127  0.1657188460

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>              F1           F2           F3           F4          F5
#> 1  -0.058838777 -0.130304943  0.369028193  0.593417656 -0.18215200
#> 2   0.038699714 -0.170760873 -0.138085808  0.043536047  0.01711597
#> 3  -0.169222092  0.173480602  0.511037374  0.143222729 -0.31618552
#> 4   0.006045389  0.088404259  0.077340915  0.166705515 -0.41013182
#> 5  -0.442828653 -0.144577164  0.013846177  0.033570855  0.01958525
#> 6  -0.111050173 -0.020422901 -0.127642113 -0.438956154 -0.35699040
#> 7  -0.065012605 -0.321406947  0.159081592 -0.273509506  0.11654443
#> 8  -0.011112901 -0.290591077  0.331737299  0.230196203 -0.20414357
#> 9   0.118413900 -0.548851231  0.425792601  0.529002713 -0.10363360
#> 10 -0.194527934 -0.324291725 -0.002935162 -0.275095468 -0.20917853
#> 11 -0.125434135 -0.257747269 -0.135344885  0.162791410  0.13033832
#> 12  0.088529258  0.024265578 -0.007834594 -0.000367705 -0.05549178
#> 13 -0.434984969 -0.711005804 -0.017200443  0.128445470 -0.21677310
#> 14  0.222166491 -0.213360330  0.484343962  0.520559411  0.05974031
#> 15  0.197785939  0.173187372  0.421069169 -0.024463997 -0.05980776
#> 16 -0.198030619 -0.408228252  0.588658068  0.228580797 -0.06450048
#> 17 -0.389855821 -0.320875984 -0.035960919  0.141799367 -0.01142753
#> 18  0.337501561 -0.522687689  0.013520548  0.185429886 -0.33557216
#> 19  0.046203021 -0.191359588  0.335210708  0.277685677  0.02237710
#> 20  0.073455277 -0.003970592  0.331683312  0.054243477 -0.13615762