Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>              F1          F2           F3          F4           F5
#> 1   0.267994785  0.09011690  0.256844220 -0.26163224  0.238842308
#> 2   0.061558824  0.33078163  0.073378757  0.20152066  0.107181691
#> 3  -0.154779814  0.10118948  0.020724586  0.39017791  0.181382567
#> 4   0.093911965 -0.07483380 -0.120691187  0.21700290  0.152515884
#> 5  -0.060347213  0.08788627  0.245842525 -0.34633770  0.104069058
#> 6   0.397058758  0.11178714 -0.081555540 -0.13500497  0.007562646
#> 7   0.193737710  0.39255483  0.210255802 -0.03001959 -0.075849066
#> 8   0.382038670 -0.10171437 -0.017545490  0.22374437  0.091868811
#> 9   0.111907535  0.03760580  0.187263693  0.26657439  0.134533564
#> 10  0.007062839  0.51263011  0.258413791  0.33784831  0.131290928
#> 11  0.170933097  0.17657487  0.191164723  0.39032793 -0.104732978
#> 12  0.121043933 -0.02521390 -0.147048842  0.02452740  0.131982041
#> 13  0.209996877  0.39550803 -0.325124481 -0.10772716  0.278574087
#> 14 -0.032215906 -0.24075043 -0.046357325  0.19217252 -0.031122208
#> 15  0.204197356  0.43849901  0.002781791 -0.14836980  0.044048566
#> 16 -0.078411443  0.17500773 -0.006049356 -0.09364484 -0.424324989
#> 17  0.293383644  0.38132713 -0.005274845  0.17619183 -0.074659534
#> 18  0.268344173  0.09677766  0.129618563  0.21105198 -0.075216177
#> 19  0.235965004  0.38686472 -0.053964215 -0.53822404 -0.021318817
#> 20 -0.013496084 -0.24821313 -0.165354551  0.21473209  0.478841652

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>              F1           F2            F3          F4           F5
#> 1  -0.251429945  0.055368585 -0.1493566768  0.08022749  0.191138833
#> 2  -0.273989859 -0.563903460  0.1921261406 -0.20203259 -0.149349797
#> 3   0.432245798  0.155853920  0.1740768499  0.25887576  0.184814932
#> 4   0.064887305 -0.408482612 -0.0402767793  0.03325490 -0.267695442
#> 5  -0.031500248 -0.007404090  0.0009661808  0.14002416  0.120394956
#> 6  -0.020954045 -0.143466329 -0.0750601444  0.26180298  0.187871185
#> 7  -0.006945737 -0.096211268  0.0493935328  0.23082774  0.461470870
#> 8   0.340725398  0.109308008  0.0786478291 -0.27843420  0.550385655
#> 9  -0.164500331 -0.400611945  0.1206051754  0.19699761  0.125247615
#> 10  0.101560256 -0.280840888 -0.0865059372  0.23751574 -0.124381651
#> 11 -0.030693065  0.107482532  0.0869761167  0.20114749  0.009936329
#> 12 -0.330849259 -0.432923364  0.3532718917 -0.07450786 -0.093069625
#> 13 -0.386162925 -0.304420246  0.2488227386 -0.07941765 -0.170936643
#> 14  0.378706392 -0.001624922 -0.2232911980  0.02087857  0.316062041
#> 15 -0.061886584  0.195375426  0.1732116892  0.11851161 -0.122887502
#> 16 -0.153764160 -0.452633049  0.0815430857  0.49420646  0.526307469
#> 17  0.030180275  0.230288023 -0.1233307069  0.26690399  0.046211596
#> 18 -0.089089579 -0.248522293 -0.0414201155 -0.15259047  0.287441243
#> 19 -0.058466914 -0.405105060  0.2623405725  0.19952570  0.036850650
#> 20  0.427190283 -0.320813364  0.3876138794  0.02023173 -0.020420339