Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>             F1          F2            F3           F4           F5
#> 1   0.17471703 -0.07345523 -0.2675224180 -0.016461985  0.131859566
#> 2   0.56372762  0.16957306 -0.1007627536 -0.045990757  0.411294332
#> 3  -0.22449955  0.09186010  0.0369589806 -0.104655765  0.002451712
#> 4   0.63896235  0.29815571 -0.0852431293  0.015201142 -0.451513855
#> 5  -0.01845217 -0.17855777 -0.0741434899  0.298677225  0.438098455
#> 6  -0.03479994 -0.17279067 -0.0776743260 -0.136296505  0.138266904
#> 7   0.47801479 -0.17460920  0.4123742862 -0.200470446 -0.121118638
#> 8  -0.08234559 -0.20504704  0.4427467232 -0.278209716  0.087579646
#> 9   0.21324015  0.03726549 -0.2104299068  0.240673400 -0.066782483
#> 10  0.24811728 -0.33282164  0.1602755166 -0.003335075 -0.215721088
#> 11  0.03208209  0.30410017 -0.2382955604  0.045746891  0.248711152
#> 12  0.27969423  0.08527592  0.3985796748  0.401317525  0.153434400
#> 13  0.07625653 -0.04916791  0.2464073168 -0.223718176  0.069195721
#> 14  0.40027196  0.01657354  0.2767866947 -0.316098142 -0.110760529
#> 15  0.25281186  0.06561033  0.1618805607  0.029036797  0.135733024
#> 16 -0.10415927 -0.04558003  0.0390609697  0.529081002 -0.047924800
#> 17 -0.11226220  0.04858745 -0.0002276991 -0.103555348 -0.169071476
#> 18  0.19181059 -0.03719919 -0.2290754735 -0.003189704 -0.020388346
#> 19  0.34062276  0.32012843  0.3501265861  0.630473546 -0.008598229
#> 20 -0.04408791  0.25860195 -0.2498443398  0.142807006  0.115622270

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>             F1           F2           F3           F4           F5
#> 1  -0.37468727  0.267544239  0.379431411  0.092072423 -0.063395536
#> 2  -0.35514173  0.201801672  0.187809125  0.081584050  0.276498726
#> 3   0.34075490  0.029827461  0.018081686 -0.081435548  0.069575617
#> 4   0.56008251 -0.059289422 -0.326660405 -0.085300260 -0.271063174
#> 5   0.18090639 -0.006662934 -0.004062043 -0.056018727 -0.103853822
#> 6  -0.17318505 -0.227625496  0.268051373  0.049581840  0.072724542
#> 7   0.18830472 -0.575673823  0.094792921 -0.190411571  0.030318070
#> 8  -0.49497642 -0.094899901 -0.187268910 -0.107039912  0.222990309
#> 9  -0.10671904 -0.074753812 -0.215441336  0.043540684 -0.180243373
#> 10  0.13819071  0.200383814  0.251182019  0.398210599  0.351916770
#> 11  0.18294267  0.100883980  0.100368169  0.140975766 -0.043048524
#> 12 -0.37689643  0.038547535 -0.026556652 -0.021784260 -0.372157907
#> 13 -0.26041527  0.105391410  0.332961308  0.027724387 -0.006379742
#> 14  0.15164241 -0.142508621 -0.180439248 -0.257154858 -0.120254237
#> 15 -0.01429101  0.304113288  0.021080097 -0.171892459 -0.289524670
#> 16 -0.41145741 -0.160076170 -0.470983963 -0.105484715  0.292064565
#> 17 -0.19878412 -0.046358033  0.068842960 -0.003079981  0.159552838
#> 18 -0.25334874  0.078506894  0.270100053 -0.102226801 -0.400739267
#> 19 -0.09957540 -0.080938591  0.202546040 -0.087392983  0.125346702
#> 20 -0.32567757 -0.220010893  0.153607703 -0.103977758 -0.235926062