Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>             F1           F2           F3          F4            F5
#> 1  -0.18958901 -0.077147800 -0.207703329 -0.13767984 -0.0922150984
#> 2  -0.03722436  0.444157990 -0.095407467 -0.30256124 -0.0132093269
#> 3   0.49429255  0.092713098  0.402623139  0.55712924  0.0815353594
#> 4   0.29823131 -0.298842305  0.331517652  0.29685506  0.0003845246
#> 5   0.08107691  0.377944184 -0.310929198  0.27520218  0.0702591540
#> 6  -0.15730880  0.345001510 -0.066249893 -0.58510299  0.1540440180
#> 7   0.15424778  0.247624285 -0.151173525  0.37205560 -0.1857639537
#> 8  -0.01201505  0.082962209  0.232880272  0.38948573  0.0570708690
#> 9  -0.20017847  0.221673856  0.135484683 -0.09342216  0.5609343675
#> 10  0.12412864  0.196768757  0.496328887  0.49805187  0.3091883112
#> 11  0.23626946 -0.433292753  0.299770820  0.26398318 -0.1272910430
#> 12 -0.17072199 -0.039073084  0.008082913  0.49483940  0.1375169332
#> 13  0.32006634  0.146924259  0.073936808 -0.18097665  0.2533071704
#> 14 -0.03549833  0.579788352 -0.126517637  0.15093153  0.0091973682
#> 15 -0.09984541  0.048330357  0.251629654 -0.07006010 -0.0701081872
#> 16  0.21251578  0.441201901 -0.282808966  0.13361471 -0.1862424079
#> 17 -0.07657331 -0.194005118  0.372489715  0.05133391  0.1059046023
#> 18  0.27850297  0.007706269  0.178319459 -0.29962126  0.0726723622
#> 19  0.04148473 -0.003440728  0.273971280  0.42645164  0.2533590492
#> 20  0.58178712 -0.255849915 -0.088844804 -0.04261645 -0.0910629568

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>             F1          F2          F3            F4          F5
#> 1  -0.22320157 -0.54165545  0.01424553 -0.0175809769 -0.24498644
#> 2  -0.31254378 -0.14212139  0.03338895 -0.1761952818  0.43499278
#> 3  -0.08715924  0.07319673 -0.24285802 -0.2569511276  0.10585233
#> 4   0.15432427 -0.10302359  0.40043427 -0.1149877579  0.10887841
#> 5  -0.09530175  0.13391771  0.26363158 -0.3191076163 -0.10547581
#> 6  -0.15950315  0.16521397  0.13598064 -0.0411898397 -0.05806839
#> 7  -0.07938148 -0.16350588  0.41316894 -0.1637865617  0.27734589
#> 8   0.02269300 -0.44573226  0.03383133 -0.1193125456  0.42798384
#> 9  -0.02247404  0.21165211  0.35222550  0.2177215937  0.37996944
#> 10  0.19566203 -0.06047999  0.08772293 -0.0177178624  0.06621446
#> 11 -0.38682985 -0.38763349 -0.17709611 -0.1761779653  0.10162870
#> 12 -0.26464019 -0.33076486  0.28832031  0.0008350043  0.09961176
#> 13 -0.26119087  0.21682450  0.10803141 -0.3240616610  0.19946322
#> 14 -0.18410348 -0.31906208 -0.01480082 -0.0662500850 -0.18244138
#> 15 -0.02723975 -0.01569946  0.14946692 -0.0194605651 -0.04662584
#> 16 -0.13418444 -0.20948898 -0.11670846 -0.3112861749  0.26551182
#> 17  0.16272401 -0.40955081  0.02978290  0.0852265327  0.60465020
#> 18 -0.22869149 -0.22435678  0.19719288 -0.1965287860  0.21096979
#> 19  0.17819679 -0.30590107  0.17828473 -0.1099421423  0.41165884
#> 20  0.02191634 -0.25056091  0.13501815 -0.0652945933 -0.26948967