Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>             F1           F2          F3          F4           F5
#> 1  -0.45083475  0.007083734  0.02880531 -0.12400999  0.202501290
#> 2   0.22568368  0.231219110  0.40514385  0.12364996  0.241907995
#> 3   0.27452418  0.057700587 -0.03886111  0.39597754  0.232742951
#> 4   0.38586390  0.080465669  0.22385619 -0.03078661  0.107115223
#> 5  -0.02791947 -0.124435154  0.03786666  0.41420894  0.129822578
#> 6   0.40427992 -0.117576413  0.37704829  0.39189442  0.162561532
#> 7  -0.21200644 -0.006063251  0.16428256 -0.09929628  0.114710867
#> 8   0.33350211 -0.163978497  0.56037797 -0.77529535  0.276799121
#> 9   0.21722915  0.130795587  0.01795629 -0.40307275  0.084515885
#> 10  0.38151883  0.267976165  0.37127680  0.02575441  0.079431662
#> 11  0.10796925 -0.012153993 -0.09619700  0.33691282 -0.047729375
#> 12  0.08175558 -0.003691432  0.23026204  0.12886548  0.001853165
#> 13  0.24604056  0.191699883  0.45187040  0.21031889  0.256290908
#> 14 -0.32660836  0.394399435  0.11755226  0.09005510 -0.117743321
#> 15  0.01934370 -0.269267912  0.10707875  0.23069684 -0.129318926
#> 16  0.22543129 -0.222965564 -0.17492581  0.15632268  0.068688627
#> 17 -0.37476848  0.205521941  0.37374718  0.08016570  0.218245134
#> 18  0.38729416 -0.093477024  0.69404232 -0.12803267  0.075080033
#> 19  0.16733572 -0.333124011  0.34606619 -0.04112832 -0.058382922
#> 20  0.16080709 -0.041871858  0.07150643  0.19053157 -0.327995964

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>             F1            F2          F3            F4          F5
#> 1  -0.39441398  0.1823168233  0.12258654 -0.1446889468 -0.05460015
#> 2   0.17140644  0.1360015346 -0.37283179  0.0002262444 -0.15963338
#> 3  -0.15968902 -0.3079111850  0.03032147 -0.0657147740 -0.49747729
#> 4  -0.41542276  0.0049910486 -0.22988740 -0.0630179395 -0.31238095
#> 5  -0.14308791 -0.2027981335  0.12592776  0.0221851100  0.12222883
#> 6  -0.22237138  0.2012321110 -0.17765240 -0.1980002258 -0.17421255
#> 7  -0.28361474 -0.3450345226 -0.04058652 -0.0302394332 -0.03496579
#> 8  -0.38406829  0.3816998234 -0.34822818 -0.0688333788 -0.15948540
#> 9  -0.23251428 -0.1116994331  0.09422485  0.1323049939 -0.26048590
#> 10  0.10041528  0.1569228468  0.13223553 -0.2631055327 -0.23709690
#> 11 -0.15712280  0.1266192905  0.05725268 -0.4162837484  0.04190519
#> 12 -0.28188289  0.2805606686 -0.08685807 -0.1838493626 -0.01781310
#> 13  0.05252606  0.3561707236 -0.36383377 -0.2973733106 -0.33827653
#> 14 -0.22902588  0.1029757721 -0.47242827  0.0673744664 -0.43583161
#> 15 -0.02036963 -0.2006441068 -0.44277852 -0.2714686440 -0.33091835
#> 16  0.10317491 -0.4860959216  0.22544461  0.0971832853 -0.16497490
#> 17  0.13673294 -0.0430455828  0.00550316 -0.3835343527 -0.01598986
#> 18 -0.12198775  0.2422379223 -0.39786289 -0.3138035672  0.10892028
#> 19 -0.67773789  0.0005534055  0.18037769 -0.1769272697 -0.18758034
#> 20  0.01484874 -0.3619679373 -0.05702674  0.0272717216 -0.52823967