Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for MCA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>             F1          F2           F3          F4           F5
#> 1  -0.04853474  0.22805939  0.084589001  0.50682233  0.208673542
#> 2   0.26048653 -0.02258650  0.096820012 -0.48709164  0.054922003
#> 3  -0.35363493 -0.20508433  0.826300134  0.44999737 -0.066006296
#> 4   0.34795667 -0.02207029 -0.285805138  0.61185828 -0.021084954
#> 5  -0.09745794  0.85372403 -0.204650270 -0.11993702 -0.071226653
#> 6  -0.04976084  0.42680610  0.277210421  0.02643714  0.283314589
#> 7  -0.09407253 -0.01093475 -0.261590149 -0.05090831  0.327287051
#> 8   0.41383290  0.11576379 -0.139639855  0.17377912  0.200799195
#> 9   0.21670681  0.17410531  0.211830492  0.14062148  0.308534304
#> 10  0.04490537  0.21942860  0.100784427  0.67833285  0.090338743
#> 11  0.22916034  0.00915691  0.259515641  0.02780998 -0.001086953
#> 12  0.05400273  0.14462785 -0.374939461  0.14827346  0.163185855
#> 13  0.07684492 -0.01395024  0.287582958  0.09486666  0.135230642
#> 14  0.35102163  0.21738398  0.312063699  0.01051241  0.591961627
#> 15 -0.03892454 -0.16669668  0.122980544 -0.04586808  0.550340540
#> 16  0.06299577  0.27436640 -0.388085785  0.11150329  0.106296410
#> 17 -0.13383903  0.06230044 -0.024722063  0.31725254  0.287458191
#> 18 -0.16782804  0.22676038 -0.119651069 -0.13486042 -0.342476842
#> 19  0.13743610  0.50009009 -0.003014892  0.13157009 -0.145818708
#> 20 -0.17765704  0.63821207  0.065840087  0.49542697 -0.001771869

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>             F1           F2            F3          F4          F5
#> 1   0.02296278 -0.380767054  0.0104634888 -0.22667252  0.17046453
#> 2  -0.01814532  0.380603564 -0.4076948064 -0.04958900  0.04481087
#> 3   0.21298247  0.152108091  0.0504118549 -0.31124867 -0.09769170
#> 4   0.12189003 -0.322976288  0.2715053760 -0.24895867  0.36232504
#> 5  -0.15753820  0.088870960 -0.1494924598  0.32739980  0.34512844
#> 6   0.01599519  0.034606854  0.0007751726 -0.01577428  0.10464885
#> 7   0.24235612  0.152679537  0.0782996405  0.05474678  0.28240180
#> 8   0.05000756  0.162014988  0.1453963464  0.30840830  0.23361694
#> 9   0.04409617 -0.033563712  0.1020111343  0.37545589 -0.10462689
#> 10  0.17255908  0.092771564 -0.1452197894 -0.01415520 -0.10339259
#> 11 -0.02379714  0.330584599 -0.2235153578  0.14673089  0.13494191
#> 12 -0.23419128  0.001395140 -0.2094934494  0.21819125 -0.10402888
#> 13  0.04839851  0.009508094  0.1046134937 -0.27113759 -0.08384802
#> 14 -0.29018027 -0.085081281  0.0769654205  0.34266346  0.34336591
#> 15 -0.01621033  0.219229452 -0.0289591522 -0.06791723  0.05972815
#> 16  0.19944193 -0.114190532 -0.1257181941  0.41159010 -0.10065631
#> 17  0.34760940 -0.127128452 -0.1237261716 -0.60448496  0.11626026
#> 18  0.06630314  0.151047409 -0.0282866977 -0.28180737 -0.03281488
#> 19 -0.39878066  0.132959422 -0.2921725805 -0.35225006  0.28398606
#> 20 -0.22150638  0.228797352 -0.0325311820  0.35842447 -0.27728598