Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>              F1          F2          F3          F4           F5
#> 1   0.292479094 -0.02170056  0.04161898 -0.04996867  0.458373982
#> 2   0.077980465  0.01916429  0.13232323 -0.48214237  0.126342061
#> 3   0.288068306 -0.16617959 -0.02225667 -0.10697569 -0.240931145
#> 4   0.008973974  0.30067643  0.07462273 -0.10165188  0.075947368
#> 5   0.263668097  0.32148875  0.20183509 -0.37361589  0.002491602
#> 6   0.105673222 -0.27303465  0.36934386 -0.04844313 -0.022060121
#> 7   0.471304111 -0.03039108  0.39075768  0.03468182  0.097864596
#> 8   0.066134706  0.21922865  0.52905441  0.07451448  0.016263634
#> 9  -0.060668436 -0.21682261 -0.21516471  0.03219297 -0.391798316
#> 10  0.265904081 -0.20982007  0.11614465  0.05038002 -0.198831027
#> 11 -0.179648655 -0.13465673 -0.16869169 -0.32414343 -0.353099157
#> 12 -0.040852736 -0.03020670  0.46328348  0.18746746  0.202673352
#> 13 -0.248084740  0.17878468  0.05318374 -0.18552726  0.452635735
#> 14 -0.163753409 -0.07937645  0.33872621  0.19458540 -0.144781736
#> 15 -0.385070002 -0.01959636  0.09729652  0.20552610  0.500719008
#> 16  0.225570771 -0.14026136  0.11565868 -0.10600856 -0.323742265
#> 17  0.306600312  0.43044912  0.04546936 -0.03006885  0.530757979
#> 18 -0.597086993  0.20144757  0.27799556 -0.07682929 -0.073425812
#> 19 -0.232009669  0.14018038 -0.03885428 -0.23958656  0.331427466
#> 20 -0.059556468  0.24804817 -0.07266590 -0.16510195 -0.006007474

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>               F1          F2           F3           F4          F5
#> 1  -0.0114058079 -0.27350878 -0.048806960  0.179233189  0.06858788
#> 2   0.0568668311 -0.12837117 -0.185037273  0.536138057  0.24566385
#> 3   0.0200272352  0.13311862 -0.303555065 -0.272605969  0.21106409
#> 4  -0.2371689589 -0.25303910  0.017742315  0.413568796 -0.25603232
#> 5  -0.3754102047  0.16524314 -0.383740772 -0.136094703  0.07623480
#> 6   0.2163059356 -0.10684306 -0.534963703  0.275159212 -0.25812801
#> 7  -0.2491196037  0.10484028 -0.015101193 -0.279450677 -0.11544302
#> 8  -0.0534637221  0.14321440  0.114565896 -0.018730730 -0.20295008
#> 9  -0.0272259970 -0.05538358  0.045865459  0.096077647 -0.39314340
#> 10 -0.0607430590 -0.02863631  0.138586551 -0.044589575 -0.26921963
#> 11  0.3430600701  0.39178460 -0.613643090 -0.018420121  0.21732311
#> 12 -0.0858381602 -0.03080565 -0.013611542  0.106309664 -0.38642776
#> 13  0.1772462347  0.15506694 -0.032758216  0.073636629  0.19043327
#> 14 -0.1408541622  0.01755632 -0.317516820  0.409118581  0.19754958
#> 15  0.1480464435 -0.02798910 -0.391361090 -0.263164757  0.26001650
#> 16 -0.0685673250  0.08782842 -0.105780134 -0.006717933  0.10646168
#> 17  0.0007779257 -0.24687985  0.222662412  0.097709254 -0.36552611
#> 18  0.4012652996 -0.04704815 -0.625326817  0.581825831  0.28416263
#> 19  0.4327708729 -0.17487597 -0.313134656  0.659270336 -0.03573795
#> 20 -0.1303510014  0.09672206 -0.001055489 -0.062105190  0.24566476