Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for MCA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>             F1           F2           F3           F4          F5
#> 1  -0.15419885 -0.057918012  0.277930547  0.229299893 -0.04918302
#> 2   0.08957097  0.026695991 -0.208405489 -0.078197901 -0.19348841
#> 3   0.11456895  0.117071371 -0.195480859  0.092736245  0.29614123
#> 4   0.07628718  0.159418842  0.066748297  0.405563113 -0.05066126
#> 5   0.37499557 -0.001786224  0.016289576 -0.285891721 -0.09335688
#> 6  -0.54194772  0.145129769  0.230727545  0.150035825 -0.33457187
#> 7  -0.26786021  0.065603311  0.079773087  0.488173194 -0.21941057
#> 8   0.28832168  0.074399035  0.091492780  0.184088466 -0.01242387
#> 9   0.27728115  0.093706920 -0.104063781  0.434854932 -0.04710031
#> 10 -0.14421099 -0.180287323 -0.038160422  0.020558312 -0.08142109
#> 11  0.22268565  0.458430691  0.068750304 -0.024362193  0.16780764
#> 12  0.17124939  0.193293958 -0.133319588  0.002785838 -0.23107174
#> 13 -0.24943087 -0.050208032  0.146970619 -0.062659211 -0.38837017
#> 14  0.15097728 -0.117686598  0.106406904 -0.137893907 -0.08240232
#> 15 -0.17448214  0.107584174  0.209175359 -0.041254620  0.43941036
#> 16  0.04330571  0.264451659 -0.102405498 -0.106068735  0.13138017
#> 17  0.22483910 -0.003038597 -0.001654998  0.201022675 -0.00462909
#> 18  0.42957021  0.044867433  0.443339079 -0.191054385  0.03823633
#> 19 -0.22535379 -0.260608313  0.055135452 -0.036844437  0.04827954
#> 20  0.33443331 -0.151864827  0.041472320  0.161527548  0.02822350

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>               F1           F2          F3          F4           F5
#> 1   0.0640386451  0.006257043 -0.06938478 -0.05317135 -0.082899798
#> 2  -0.1852582931  0.263303772 -0.04633991  0.42191053 -0.075999435
#> 3   0.1386393348  0.002817373  0.24211867  0.01240868  0.148926118
#> 4   0.3769954625 -0.097590089 -0.14554015 -0.03741468 -0.112139734
#> 5   0.4236926574  0.270849460  0.01642903  0.14921377  0.182359435
#> 6  -0.1050644857 -0.007546605  0.03114534  0.24556018 -0.034090467
#> 7  -0.3525626086  0.104623418  0.27217710  0.30551478 -0.056693937
#> 8  -0.1174930198  0.431797203 -0.07662027 -0.07693258  0.158790022
#> 9  -0.2729418345 -0.022280850 -0.15939825 -0.36614734 -0.365489695
#> 10 -0.0005098838  0.223998596  0.06843024  0.21650038  0.018658280
#> 11  0.0116266613 -0.084266549 -0.38474856 -0.14021038 -0.120792326
#> 12  0.2756415973  0.426975194 -0.06735015  0.22660215  0.318842341
#> 13 -0.1071001406  0.035988212  0.06449437  0.07954862  0.105784478
#> 14 -0.1404831274  0.204736705  0.06103661  0.41858122 -0.303070677
#> 15  0.1822362591  0.816069847  0.21520598  0.09483830 -0.248471108
#> 16 -0.0521638190  0.077067172  0.13883070 -0.08142245  0.015101550
#> 17 -0.0002407757  0.115996196  0.08681738  0.05589485 -0.005907332
#> 18 -0.3507587145  0.206101911 -0.15661046  0.15655623  0.018818850
#> 19  0.1285238372 -0.069526613 -0.26981690 -0.10218525 -0.261261592
#> 20  0.1008186230  0.072428162 -0.20046342 -0.07227784 -0.178597602