Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>             F1          F2          F3          F4          F5
#> 1   0.27187405  0.20355063  0.10650862  0.05468883  0.51883425
#> 2   0.20552643  0.41172761  0.21486117  0.19770022  0.38110435
#> 3   0.03830301  0.16273983  0.69806141  0.09575654  0.29485453
#> 4  -0.15951879 -0.40765833  0.10185267  0.32269660  0.04710670
#> 5   0.27296444 -0.15604770  0.15755779 -0.22330770  0.34465773
#> 6   0.12216122 -0.31185375  0.01402383  0.13711940 -0.05746263
#> 7  -0.20742231 -0.14257653 -0.36105076  0.38500522  0.12661289
#> 8   0.02367265  0.15832155  0.41031790  0.14847118  0.52907014
#> 9   0.46707015 -0.05903031  0.27298499  0.05219901  0.20848272
#> 10  0.17473784  0.10906156 -0.01471340  0.23474795  0.35888423
#> 11  0.01954252  0.00508386  0.12739148  0.05377768  0.29040656
#> 12 -0.17008904  0.14942145  0.19256734  0.46181981  0.69505655
#> 13  0.01911587  0.12391870  0.23366516 -0.07120636  0.36143165
#> 14 -0.17009421 -0.02449458  0.28237828 -0.02004206  0.16788417
#> 15  0.45766415  0.22166721  0.24037939  0.24389094 -0.08509380
#> 16 -0.28734473 -0.04275752  0.35212882  0.34011160  0.21356641
#> 17  0.03046732  0.26950967 -0.08256041 -0.24428469  0.03840245
#> 18 -0.21467275 -0.00943770  0.28546974 -0.09142191 -0.06552658
#> 19 -0.08641295  0.26708206  0.38300399  0.12310529  0.27158605
#> 20  0.02714639  0.10763054  0.09435654 -0.15419278 -0.07971340

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>             F1           F2           F3          F4          F5
#> 1  -0.16558584  0.061730153  0.394077181 -0.29310349 -0.11057085
#> 2  -0.30126435  0.292657272  0.052092019  0.12302303 -0.36658729
#> 3   0.32522575 -0.054858978  0.158458888  0.14528388  0.07867119
#> 4  -0.10694081  0.019271555  0.003514338 -0.38545443 -0.04236943
#> 5  -0.18626416  0.007902169  0.112842162 -0.33542726  0.33223425
#> 6   0.11655076 -0.104371460  0.271004201  0.03903213  0.11538236
#> 7   0.30294514  0.071252096  0.385511569 -0.04003999  0.29745802
#> 8  -0.09831202  0.173113950  0.113112717 -0.33685817 -0.14206221
#> 9  -0.20382033  0.267407431 -0.247178052 -0.41271876  0.07343375
#> 10  0.26409497  0.350179908 -0.335336918 -0.25388998  0.17011461
#> 11  0.08516457 -0.242358940  0.190850534 -0.19581160  0.26387343
#> 12  0.04444371 -0.331643691 -0.073012835 -0.23261589 -0.13855198
#> 13 -0.02614050  0.074371485 -0.086506324 -0.32308591 -0.00274153
#> 14 -0.36204638 -0.262271327  0.028605013 -0.60290597  0.42942887
#> 15 -0.15977655 -0.019280333 -0.183683452 -0.45306131  0.06681951
#> 16  0.53355896  0.394226873  0.287494070 -0.17869779  0.02633084
#> 17 -0.11928469 -0.188186070  0.567823041 -0.02755488  0.28272151
#> 18  0.03157526  0.197508162 -0.055028961 -0.34029910 -0.06348316
#> 19  0.07022962  0.024330065 -0.248345003 -0.20424415 -0.39832547
#> 20 -0.04988254  0.170000714 -0.039153401 -0.09127885 -0.02040073