Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>             F1           F2           F3           F4          F5
#> 1   0.37863290 -0.101376340  0.213805920 -0.094988527  0.03834275
#> 2   0.15618657  0.082716901  0.075122753 -0.103462613 -0.21472452
#> 3   0.02892040  0.310271779 -0.352654382  0.361922833  0.21765431
#> 4   0.28198944 -0.052456933  0.427942919  0.363693405  0.18958795
#> 5  -0.29122521  0.119309228  0.301641194 -0.093182871 -0.13187273
#> 6   0.09360046 -0.112336059  0.124916941 -0.175799708 -0.01211716
#> 7  -0.14819124  0.262685978  0.683273199 -0.250242320  0.28259912
#> 8  -0.20775293 -0.045244130  0.187826510  0.003780336  0.08348584
#> 9   0.51161166 -0.035013496  0.359178486 -0.034517667 -0.25305243
#> 10  0.26545400  0.443464812  0.740736638  0.157537548 -0.21781597
#> 11  0.76510401  0.312797563 -0.009259178  0.435405496 -0.01501752
#> 12  0.10999807  0.002374119  0.367847339  0.340064630 -0.09281508
#> 13  0.47211587  0.124972033 -0.050104785  0.312139694  0.21577898
#> 14 -0.06956779  0.038378341  0.487761890 -0.166565490 -0.16701530
#> 15  0.31434956  0.197274627  0.524924900  0.228402819 -0.26667925
#> 16 -0.14788562 -0.005497261  0.554540837  0.079725824  0.22076125
#> 17  0.40393675  0.171366486 -0.020598307  0.425682553 -0.10224329
#> 18  0.50026271 -0.205518496  0.101461972  0.117111508  0.27155255
#> 19  0.20345887  0.048558918  0.148990679  0.253508703  0.03318381
#> 20  0.34373131  0.463626954  0.238007576 -0.174407436  0.11922953

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>             F1          F2           F3           F4           F5
#> 1   0.25218470  0.37001418  0.077485377 -0.204360004  0.341554168
#> 2   0.35870855  0.46534633 -0.160876705  0.093447142  0.281597317
#> 3   0.27849462 -0.26250657 -0.206837564 -0.211372592  0.165838636
#> 4  -0.26657905 -0.34222117 -0.020378252 -0.380448490 -0.057716994
#> 5  -0.15282824 -0.15667820 -0.191404531 -0.472823793  0.321360072
#> 6   0.24976313 -0.39236891  0.012280146 -0.284400896  0.258560007
#> 7   0.22218157 -0.03762715 -0.137443318  0.022486190 -0.182561513
#> 8   0.62137637  0.50810739 -0.404122880 -0.203261604 -0.008529715
#> 9   0.09839780 -0.15772471  0.082968519 -0.012629532 -0.054886026
#> 10 -0.01830912 -0.06061984 -0.009479506  0.011097048  0.173613939
#> 11  0.12827650  0.21765934 -0.064944349 -0.175458153  0.217760454
#> 12 -0.12848353  0.18229603 -0.228891776  0.014665578  0.156397809
#> 13  0.30141764 -0.20528035 -0.049396919 -0.256165272 -0.025147126
#> 14 -0.05228198  0.18366703 -0.219763533 -0.105141869 -0.126845760
#> 15  0.17752959  0.20976929  0.026374278 -0.079899833 -0.041295082
#> 16  0.44377175  0.25717368 -0.461933967 -0.006714822 -0.434656848
#> 17 -0.09995708  0.13366156 -0.076367524 -0.222089271  0.051036078
#> 18  0.25831935 -0.05795181 -0.516219610 -0.223373123  0.119796884
#> 19  0.09724505 -0.16156957 -0.105359572 -0.546862804  0.218382842
#> 20  0.19132438 -0.31335292  0.102593818 -0.145288842 -0.149529126