Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>              F1           F2           F3           F4          F5
#> 1   0.433428255 -0.233808770  0.185122398 -0.137936770  0.03002061
#> 2  -0.077652269 -0.017107272  0.363648633 -0.010538847  0.43430990
#> 3   0.140442187 -0.411216017  0.062174517  0.193013374 -0.07264586
#> 4  -0.368255593  0.208159873  0.186112411  0.180045542  0.46128631
#> 5  -0.053272443  0.105208978 -0.284736753 -0.048198129  0.12480605
#> 6   0.323403790  0.256036861  0.160423482 -0.082485343  0.33695794
#> 7   0.086543598  0.119722229 -0.052349262  0.005352615 -0.24032788
#> 8  -0.041509045  0.040642950 -0.010333986 -0.602973161 -0.10201629
#> 9  -0.143658477  0.301880525  0.008819776  0.003395888  0.03603482
#> 10  0.342096131 -0.070708961  0.089253689 -0.349096464  0.28267843
#> 11  0.542105392 -0.247554822 -0.313061704  0.149829885  0.07728103
#> 12  0.007551645  0.057439257 -0.017746454  0.193098941 -0.09477281
#> 13 -0.127096547  0.207698887  0.084859267  0.256594830  0.45596278
#> 14 -0.190183268  0.009381502  0.271683340 -0.020345212  0.19949946
#> 15 -0.174473687 -0.035741838 -0.077239245  0.093520863 -0.00624486
#> 16  0.215795146  0.066610310  0.332588872 -0.349427275 -0.11635946
#> 17  0.116830863  0.372781863  0.083075230 -0.048743176 -0.08347651
#> 18  0.228998260 -0.510361346 -0.264653509 -0.438004882  0.26198856
#> 19 -0.119880435 -0.178751493  0.313062147  0.196099090 -0.42199118
#> 20 -0.128783999  0.224632657  0.162735184  0.227776655  0.03774088

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>              F1          F2          F3          F4           F5
#> 1  -0.134446438 -0.17708236  0.12510134  0.29712675  0.111206209
#> 2  -0.337409851 -0.11409768  0.18499329  0.16296664  0.050454172
#> 3  -0.002445321 -0.24802977 -0.11402265  0.55164153 -0.026880494
#> 4  -0.072591782  0.50694784 -0.32053012  0.08966433  0.110289570
#> 5  -0.021882971  0.06835982 -0.15700796 -0.13746127  0.102640350
#> 6  -0.289628912 -0.02615322 -0.06532050  0.14552373  0.187997874
#> 7  -0.117205204 -0.26527033 -0.41172306  0.73144827 -0.006279428
#> 8   0.088665113 -0.13530281  0.22179011 -0.17560295 -0.378071096
#> 9  -0.187827262 -0.02999274 -0.24245394 -0.04343258 -0.035832691
#> 10  0.027020317 -0.08311651  0.19218171 -0.10655363 -0.300925170
#> 11  0.075790237  0.40942622  0.27741516  0.33326678  0.244465354
#> 12  0.465411080  0.08937588  0.12660945 -0.23753438  0.031395983
#> 13 -0.356106966  0.18785351 -0.23143640 -0.18234521 -0.191597441
#> 14  0.329328949  0.12012649 -0.16655917 -0.13270173 -0.155788426
#> 15 -0.155282561  0.19622514 -0.25172818 -0.10471329  0.052179690
#> 16 -0.060394721 -0.06496146 -0.23502591  0.03177021  0.490605777
#> 17 -0.471447709 -0.09837814 -0.09423322  0.30820363  0.116311473
#> 18 -0.111056617  0.01708733 -0.16540274  0.30759887 -0.021196861
#> 19  0.082115562 -0.31050031 -0.25226053 -0.29313688  0.001680094
#> 20  0.350314892 -0.02857392 -0.16063472  0.16364698  0.339645799