Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>              F1            F2           F3          F4           F5
#> 1   0.013090319  0.0591326327  0.231867176  0.14494595  0.354914259
#> 2   0.184244154 -0.0468818926  0.249436311 -0.30468936  0.004204361
#> 3   0.081821397  0.0774959886 -0.256187853 -0.22085094 -0.045879236
#> 4  -0.351530779 -0.5300486542 -0.238847382 -0.01319730 -0.078764929
#> 5   0.258785950 -0.1230087978 -0.109089577  0.24381875 -0.037498105
#> 6  -0.005667743  0.4803992159 -0.413173683  0.24138610  0.173991218
#> 7   0.077785858  0.0442166585  0.197832889  0.14333194  0.015321374
#> 8   0.144250046  0.4334319134  0.252409756  0.19162225  0.617011789
#> 9  -0.091504859 -0.1169824711 -0.096356801  0.42921448  0.073922696
#> 10 -0.432462600  0.2720900103  0.107501038  0.12177083 -0.076332808
#> 11 -0.055723554  0.4242007682 -0.335701677  0.30666374  0.127135187
#> 12  0.083141285 -0.1321129150 -0.044278281  0.26836900 -0.097970647
#> 13  0.051242321 -0.0217907411 -0.009512419  0.07084820  0.119462184
#> 14 -0.390236722  0.2465359436  0.402427389  0.22482748  0.055146514
#> 15 -0.371216490  0.0007875158  0.027036062  0.15883927 -0.045831101
#> 16  0.446760711 -0.3471180660 -0.066244924  0.04400695  0.227066776
#> 17  0.199431102 -0.2887021477 -0.059935830  0.28571554  0.257803875
#> 18 -0.252658082  0.3459034755  0.120300339  0.10110857  0.318856101
#> 19  0.269962267  0.3271202484 -0.176955985  0.08525187  0.299923919
#> 20  0.154189459  0.0204926282  0.298158135 -0.22477711  0.296359080

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>             F1          F2           F3          F4           F5
#> 1  -0.47829219  0.15960327 -0.181908505  0.04961403 -0.203477421
#> 2  -0.11412182 -0.21456876 -0.075030007 -0.06614029 -0.112167005
#> 3   0.30862286  0.05198685 -0.235336095 -0.22330341  0.002320041
#> 4  -0.29250761 -0.25395993  0.259033266 -0.06529616 -0.163003834
#> 5   0.14927161 -0.16537872  0.425765357  0.52221743  0.245978989
#> 6   0.41891037  0.07955486  0.074190871 -0.01326824  0.025117252
#> 7   0.15458702 -0.22297511 -0.084522170 -0.28571585 -0.210489763
#> 8   0.09645186 -0.15139235  0.175992141  0.05070973 -0.283422263
#> 9  -0.28412715 -0.16592275  0.163725185 -0.02539788 -0.611502560
#> 10 -0.20403183  0.18848406  0.090795433 -0.36470797 -0.106575301
#> 11 -0.02777885 -0.28160390  0.091740272 -0.22899387 -0.138255107
#> 12  0.08880594 -0.13990176  0.343170973  0.06565679 -0.284956016
#> 13  0.05049945 -0.19554504  0.170509587  0.03638030  0.035380105
#> 14  0.32317907 -0.06718570 -0.120630723 -0.31309464 -0.106649494
#> 15 -0.07870535 -0.20001427  0.052013893  0.06372051  0.110578230
#> 16  0.19118573 -0.04824871 -0.118751247 -0.34360397  0.453600462
#> 17 -0.21307305  0.15799417  0.133942939 -0.29701816  0.011216291
#> 18 -0.24257784  0.33186020  0.193665648  0.12816926 -0.112854505
#> 19  0.27957133 -0.01270439  0.007477673 -0.18408149  0.146964716
#> 20  0.08482313  0.14116111  0.401207042  0.26023998 -0.058352115