Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>             F1           F2           F3          F4          F5
#> 1  -0.29156640  0.153178605  0.292825876  0.53892285  0.21697015
#> 2  -0.12950987  0.274838222  0.107367677 -0.24764514  0.54973639
#> 3   0.36010849 -0.009738134 -0.029004289 -0.01437925  0.43605189
#> 4   0.05732802 -0.298128329  0.209368004  0.10349499  0.34143830
#> 5   0.13879643 -0.034186924  0.001398614  0.40122877 -0.27344381
#> 6   0.20647895  0.122191949  0.169239461  0.08934243 -0.13133096
#> 7   0.14761435 -0.153844115 -0.015518609 -0.10369713 -0.10700074
#> 8   0.34005803  0.237446043 -0.069019647  0.51165965  0.33954549
#> 9   0.09273232  0.623106941  0.261824584  0.38102671  0.06100308
#> 10 -0.28242654 -0.219755934  0.289105831  0.01936464 -0.05252436
#> 11  0.25203857 -0.148173129 -0.268954384 -0.00397372  0.64148868
#> 12  0.00332249 -0.342171768 -0.039695051  0.32218767  0.10283677
#> 13  0.14929391 -0.167853666  0.056679441 -0.25158688  0.09770844
#> 14  0.06568258  0.284244771  0.067289065 -0.42581362  0.33367762
#> 15 -0.22037508  0.054245101  0.223314542  0.14581514 -0.07859173
#> 16  0.71186184  0.404389670 -0.186152803  0.37885156 -0.46847969
#> 17  0.08658305  0.232555449  0.109905500 -0.02609608 -0.24716755
#> 18  0.30907624  0.090227327  0.571652928  0.03455643 -0.02172088
#> 19  0.30318641  0.005754507 -0.032508394 -0.12357683  0.03789167
#> 20 -0.23642882  0.091444400  0.131966014  0.07741339  0.07745437

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>             F1           F2          F3           F4           F5
#> 1  -0.08869069  0.310923688  0.28489397 -0.163014464 -0.285273463
#> 2   0.03114366 -0.070608642  0.01202838 -0.008524323 -0.189723307
#> 3   0.03000434 -0.184356271 -0.16174360  0.264847558 -0.284837056
#> 4  -0.03043894 -0.220108848  0.06448830  0.283758426 -0.119631293
#> 5   0.45377766 -0.202225635 -0.11642462  0.151742775 -0.288621941
#> 6  -0.11082960 -0.004606412 -0.03419632  0.266842062  0.129365419
#> 7  -0.13233703 -0.294643625  0.56387406 -0.205163058  0.193529403
#> 8  -0.10209009 -0.197234258  0.46047851  0.052175742  0.166870113
#> 9  -0.11891257  0.296564583  0.25404395 -0.172751668 -0.042329889
#> 10  0.05255277 -0.010854100  0.03392442  0.075917574  0.250823010
#> 11 -0.07615208  0.349690264 -0.08845092  0.052140385 -0.025310623
#> 12  0.12616824 -0.215207075  0.34087012  0.204587952  0.084949715
#> 13 -0.22032410 -0.273360361  0.02647730 -0.055651756  0.072705047
#> 14 -0.52854661  0.319817541 -0.05443644 -0.093472689 -0.196743681
#> 15 -0.23813900  0.266870204  0.02411037 -0.067190883  0.214844417
#> 16  0.22749556  0.176008325  0.16347431 -0.161154777  0.592549635
#> 17  0.15040772 -0.324340112  0.19978337  0.089128191 -0.185196811
#> 18 -0.37913890  0.143738177 -0.03614396 -0.092266831  0.182553291
#> 19 -0.53028003  0.016156516  0.00680066 -0.013869589  0.133086513
#> 20 -0.08732349 -0.276417030 -0.07749657  0.194913221  0.004893915