Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>             F1           F2          F3          F4            F5
#> 1  -0.02640365  0.035491007  0.19284928 -0.23177375  0.4005794216
#> 2   0.20004775 -0.008291659  0.34957394  0.27493755  0.0102008784
#> 3   0.32464408  0.033488714  0.12400855 -0.20633665 -0.2389189221
#> 4   0.25599571  0.395509376  0.22080468  0.02851697  0.0813555377
#> 5   0.19881406  0.053222375 -0.37866591 -0.21500759 -0.3176163227
#> 6   0.20002505  0.182098599 -0.17112141 -0.36209518 -0.1733463130
#> 7  -0.20514630 -0.237661226 -0.03120658 -0.18337503  0.0160723615
#> 8   0.22036296  0.244052741  0.06992233 -0.16788788 -0.2236300414
#> 9  -0.31594878 -0.223674256 -0.15955167 -0.11177713 -0.5211195704
#> 10  0.29235805 -0.148407005  0.01954966  0.08745011 -0.2477958294
#> 11 -0.15829136  0.002529895 -0.05294313  0.15310756  0.0510894118
#> 12  0.16631892  0.229461184  0.18938897 -0.07514404  0.0009939458
#> 13  0.30623402  0.227325172 -0.25795083 -0.01608618  0.0458067512
#> 14  0.09738461  0.221112863  0.04285272  0.11632445 -0.1455934974
#> 15  0.10140598 -0.095293947  0.43403514 -0.17785268  0.2026041535
#> 16  0.17552432 -0.244449152  0.15037190  0.29693670 -0.4799529561
#> 17 -0.03680025  0.139888888  0.22736733 -0.41075057  0.0582798461
#> 18  0.57464997  0.080711818 -0.18331773 -0.11302851 -0.0253971195
#> 19  0.40717607 -0.147284903 -0.02331840  0.05030273 -0.1005405799
#> 20  0.30327013 -0.158312657 -0.03291092 -0.25049690  0.2409484510

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>              F1           F2           F3           F4           F5
#> 1   0.041609663 -0.077910469  0.265484660  0.321963262  0.264402573
#> 2  -0.137630028  0.137819092 -0.589600643  0.007571247  0.098277838
#> 3  -0.143032015  0.120349466  0.003714648  0.125206852  0.070743107
#> 4  -0.387453538 -0.060854504  0.364899144 -0.169451652  0.589415287
#> 5  -0.075002683 -0.018605306 -0.027563856 -0.309232253  0.261667560
#> 6  -0.266855599 -0.131850678 -0.355226128  0.199605799  0.198154626
#> 7  -0.029762933  0.069902807  0.136013202  0.244354494 -0.004788472
#> 8   0.077486936 -0.002934064  0.139470827  0.185524609  0.257531087
#> 9  -0.013322039  0.137073966  0.137827258  0.297485226  0.068145938
#> 10  0.002573231  0.248656963  0.366528929 -0.053735066  0.264249024
#> 11 -0.420248192  0.629701615 -0.443874299  0.400160173 -0.086174942
#> 12  0.147365494  0.027378814 -0.031274234  0.098416849  0.542475690
#> 13 -0.130777355  0.285188635 -0.058431601  0.068694981 -0.334937697
#> 14  0.098003779  0.291289567  0.109496495 -0.527150207  0.003234729
#> 15 -0.200062997  0.127836707  0.075051357 -0.332200708  0.198262234
#> 16 -0.230225668 -0.199207016 -0.092654833  0.261258860 -0.037629376
#> 17  0.336120737  0.132972709 -0.044755849  0.212387595 -0.084329185
#> 18 -0.292566927  0.119177170 -0.148992557 -0.008527427 -0.312781717
#> 19  0.122539725  0.141061763 -0.042828251 -0.031923451  0.046306055
#> 20 -0.299089501  0.156172127  0.248515425  0.169469647 -0.287550403