Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>             F1          F2          F3          F4          F5
#> 1  -0.11393136  0.12196021 -0.03933027  0.07547897  0.22868955
#> 2  -0.36245566  0.34426865 -0.47883474  0.21197807 -0.16705579
#> 3   0.10458280  0.22830858  0.53864849  0.01975142 -0.30258976
#> 4  -0.29542894  0.14813156  0.13537729  0.17444654 -0.28157716
#> 5  -0.02465857  0.05552166  0.17514484  0.15150058 -0.09443768
#> 6   0.10813085  0.21997827 -0.19071388  0.19689191 -0.54721259
#> 7  -0.37242363 -0.01568988  0.18957832  0.04289928 -0.19863331
#> 8   0.17257300  0.17911758  0.08824491 -0.06422399  0.44277505
#> 9   0.07707648  0.19986807  0.25371803  0.18164763 -0.22047288
#> 10 -0.07236992  0.18254891  0.04612638  0.48076918  0.13052350
#> 11  0.40533664  0.06978921 -0.18065696  0.46788388 -0.04925436
#> 12 -0.16644791  0.17671218 -0.01710058 -0.08502750 -0.14945800
#> 13 -0.09970503  0.12432823 -0.13474191  0.57986567  0.59065788
#> 14  0.12139432 -0.21341644  0.15966991  0.10509173  0.09162720
#> 15 -0.19982542 -0.06666130  0.10155244  0.42434594 -0.20052918
#> 16  0.14729423  0.21361720 -0.26987112  0.15856722 -0.42553311
#> 17  0.02571276  0.15473673  0.10631792  0.08367243 -0.36746158
#> 18  0.47560029  0.06889942  0.11271582  0.26929808  0.03044542
#> 19 -0.02630321  0.12540419 -0.37097389  0.29947778  0.33343898
#> 20  0.15506270  0.24345068  0.32268311 -0.18034512 -0.06256898

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>              F1          F2          F3          F4           F5
#> 1  -0.014348453  0.02428119 -0.23553827 -0.04363336  0.095996772
#> 2   0.253798662  0.07319638  0.41986009  0.24987843 -0.005412720
#> 3   0.127891431 -0.07966325  0.08669107 -0.02128859 -0.170614788
#> 4   0.117044566  0.22911437  0.23148916  0.04901558 -0.258025402
#> 5  -0.068236532 -0.33260007  0.18317602 -0.22303068 -0.035566346
#> 6  -0.002347086 -0.08444443 -0.12251044 -0.09982256 -0.157513109
#> 7  -0.198671358  0.17230814  0.25896714 -0.04459512 -0.080602364
#> 8  -0.385907223 -0.18988897  0.29818477 -0.01087143  0.063660402
#> 9  -0.074609791  0.04924708 -0.47131970  0.07851356 -0.092664017
#> 10 -0.154511317  0.22949038  0.43205684 -0.21288387 -0.021507839
#> 11 -0.304891966  0.03488418  0.46609342  0.01606043  0.246270360
#> 12 -0.127321998  0.14709106  0.08247683 -0.25801248  0.174759416
#> 13 -0.246324025  0.13786864  0.02890378  0.11672619  0.166996621
#> 14 -0.223390576  0.13350105  0.13369598 -0.09144535  0.216573789
#> 15 -0.131721425 -0.12198161  0.41155539 -0.24856352 -0.236235957
#> 16  0.093521292 -0.12964924  0.34384056  0.02304465  0.108082767
#> 17 -0.025465795 -0.02249537  0.05746759 -0.29993442  0.005918782
#> 18 -0.138221744  0.35583543 -0.13878305  0.18888950  0.028078939
#> 19 -0.254076079 -0.03922454 -0.29123610  0.31611310 -0.214836059
#> 20 -0.167505264  0.17521753  0.05120108 -0.38165447  0.015001565