Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for MCA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>             F1          F2          F3           F4           F5
#> 1   0.22691898 -0.07403845 -0.06330092  0.093112091 -0.189316577
#> 2  -0.52692916 -0.10917223  0.51595116  0.131281920 -0.104409607
#> 3   0.15906787  0.10303146  0.18287035  0.207357165  0.140566182
#> 4   0.23824690  0.13497711  0.21430982  0.500190810  0.083643890
#> 5   0.36748162  0.43625168  0.32679584 -0.127093582  0.001667244
#> 6   0.18643437  0.16499274 -0.14009909 -0.087386020  0.512831152
#> 7   0.15468013  0.20423695 -0.23663315 -0.009334409  0.219750876
#> 8   0.15719296  0.15961094  0.18655037  0.311367489  0.020136832
#> 9  -0.11790283 -0.14922667  0.16158669  0.059764004  0.045521938
#> 10  0.42725999  0.15009775 -0.10835586  0.193606759  0.331164504
#> 11  0.29044365  0.08571413  0.28325843  0.075564427 -0.206395650
#> 12  0.21250267 -0.20636787  0.08955280  0.178545859  0.054875160
#> 13  0.24024658  0.11943086  0.03473202  0.456405035 -0.038614587
#> 14  0.15041952  0.19619069  0.03661136 -0.048264285  0.096105320
#> 15  0.26947290 -0.21110681  0.15879917  0.154445873 -0.172268594
#> 16  0.02545579  0.02206669 -0.05243706 -0.377111849 -0.274679977
#> 17  0.37862094 -0.13456501  0.00649570  0.038900166 -0.038681946
#> 18  0.63457966  0.25505459  0.17085641  0.435205656  0.364246940
#> 19  0.20554887 -0.09792813 -0.43285991  0.276910357 -0.125299705
#> 20  0.01706394  0.26536108  0.42614332  0.158035026 -0.083198918

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>             F1           F2          F3            F4           F5
#> 1  -0.63430088 -0.121344900 -0.45531720 -0.1465009315 -0.077640393
#> 2   0.11956540 -0.177088445 -0.44598673 -0.3536431003  0.171347266
#> 3  -0.21786678 -0.004245854  0.42909184  0.0890141193 -0.185506457
#> 4  -0.41646041  0.063986156 -0.25603617  0.3591592519  0.006245810
#> 5   0.06805974 -0.085702785 -0.27411814  0.1701365657 -0.201978596
#> 6  -0.27644190  0.480660053  0.12387877  0.0469448085 -0.529351487
#> 7  -0.05490564  0.051872677 -0.28880840 -0.3105381425  0.010050575
#> 8  -0.26687163 -0.351181072 -0.30820930  0.1896135705 -0.020914587
#> 9  -0.36917425  0.020133298 -0.18734077  0.3348599586 -0.086966862
#> 10  0.08599002 -0.108542988 -0.09663937  0.0010100063 -0.247345136
#> 11 -0.08732170  0.054975357 -0.41391414  0.3876796247 -0.144567889
#> 12 -0.43271728 -0.159195673  0.05555465 -0.0007021077  0.336372061
#> 13  0.17189611 -0.461579623 -0.52276178 -0.1265364166  0.140392495
#> 14 -0.04899452 -0.045490903 -0.23774737 -0.1865257394 -0.142315814
#> 15  0.11217577 -0.294887061 -0.31822751 -0.0608935244 -0.312761764
#> 16 -0.30499595  0.161663081 -0.10405246 -0.1929595875 -0.140852570
#> 17  0.03193442  0.065714478 -0.25439654  0.4390370934 -0.469429677
#> 18 -0.03488583  0.066069851 -0.28059537  0.2371313044  0.113849464
#> 19 -0.42376002 -0.018512537 -0.01799450  0.0643201593 -0.023862904
#> 20 -0.46334467 -0.062553301  0.03576626  0.3290005781 -0.003278885