Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>             F1           F2           F3           F4           F5
#> 1  -0.19010595 -0.243431592 -0.051817407  0.093322334  0.208058485
#> 2   0.07507756 -0.054825869 -0.140751993 -0.007101449  0.350879494
#> 3  -0.02019743 -0.270199254 -0.283154359  0.250132342  0.321337461
#> 4   0.49438065 -0.361426411  0.177997769 -0.050967640  0.499248694
#> 5  -0.12746366  0.180595142  0.177001669  0.270114742  0.554851995
#> 6   0.50120776 -0.002844285  0.386590501  0.398611769  0.198697653
#> 7   0.19485723  0.027548669  0.288950360  0.109766288 -0.001576723
#> 8   0.10102135  0.453093473  0.049579684 -0.041557241  0.170892724
#> 9   0.14709476 -0.026740959 -0.016941808  0.049903581 -0.140229491
#> 10  0.20525629  0.026707064  0.003580798 -0.415309546  0.547070274
#> 11 -0.39659313  0.119215455  0.438065156  0.082688409  0.437435204
#> 12  0.22795421 -0.168411364  0.001001511 -0.271818222  0.413595464
#> 13  0.21160475 -0.040532646  0.399500822  0.325818416 -0.087778417
#> 14 -0.15962017 -0.199047967  0.232368300 -0.319666287  0.058779231
#> 15 -0.25044665  0.003613157  0.170360731 -0.270673496  0.418140164
#> 16  0.10129305 -0.028778813 -0.033500393  0.377003861 -0.049581165
#> 17  0.13073639  0.138549104  0.125670690 -0.131057248  0.391295084
#> 18  0.29151922  0.095090905  0.106504577  0.592572928  0.126102235
#> 19  0.57314800  0.032566766 -0.346885202  0.152597479  0.344324356
#> 20  0.04501856 -0.057036029  0.085474890  0.190056167  0.198363530

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>             F1           F2           F3           F4          F5
#> 1  -0.03834001 -0.229683739  0.045330415 -0.166766027  0.15740909
#> 2  -0.26922125  0.122791635 -0.269710091  0.002219027  0.26642644
#> 3  -0.32972110 -0.149240987 -0.103624373 -0.084220336  0.11427533
#> 4  -0.08575114 -0.193112710 -0.115609205 -0.261621730  0.11354759
#> 5   0.03857332  0.082918613 -0.046217226  0.201533343  0.04152795
#> 6   0.28837148 -0.488310077  0.107246423  0.047195546  0.17194559
#> 7  -0.53585993 -0.123209269 -0.047584307  0.184359881 -0.17034479
#> 8  -0.22532776 -0.320039809  0.105648683  0.145596026 -0.08832055
#> 9  -0.50462735  0.057291708  0.050300929  0.205144190 -0.03314557
#> 10 -0.46666826  0.034178748  0.008176913  0.332106361  0.19280844
#> 11 -0.14175706 -0.093603247 -0.210296985  0.010592074 -0.05997436
#> 12  0.12980298 -0.296542656 -0.135059066 -0.335003327  0.29047935
#> 13  0.06327837  0.269161668  0.066615032 -0.377641050  0.18030305
#> 14  0.22697562 -0.129618925 -0.092630776 -0.360322573  0.29997281
#> 15 -0.07602493 -0.054812524 -0.137484815  0.053488923 -0.13021883
#> 16  0.07924652 -0.037872723  0.068966635 -0.023121301  0.16342096
#> 17 -0.25325100 -0.007535311 -0.133604266 -0.199066042 -0.09447030
#> 18 -0.32563048  0.139552647 -0.065947738  0.057745592  0.04478623
#> 19 -0.56857401 -0.110620006  0.081975730 -0.030521657 -0.19732612
#> 20 -0.31940812 -0.488057138 -0.285504432 -0.151879757 -0.14302331