Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for MCA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>             F1           F2            F3           F4           F5
#> 1   0.25583082 -0.064629902 -0.3101843095 -0.154150197 -0.009214535
#> 2   0.20839794 -0.154896464  0.0015106005  0.145074512  0.019701644
#> 3   0.16038801  0.068034788 -0.1239055459  0.311067785 -0.016602702
#> 4   0.41801575  0.075635699  0.3750933170  0.234371988 -0.333060504
#> 5   0.02089422  0.007175445 -0.1373517229  0.065056819 -0.242740996
#> 6  -0.08530464 -0.008700705  0.2410337368 -0.049707621  0.183605283
#> 7  -0.58123638  0.377242791 -0.5234919665  0.461752458  0.444320676
#> 8   0.10009581 -0.227561785  0.1121396361  0.041310508 -0.184934594
#> 9   0.27400981  0.108830065  0.3213473818 -0.110207034  0.231662584
#> 10  0.32078940  0.031348759 -0.0008425321 -0.202735731  0.235536559
#> 11 -0.07032258  0.372137917  0.4288272971 -0.297524556  0.161614807
#> 12  0.22223562  0.350882411  0.0002653956 -0.126956800  0.247025844
#> 13  0.56995374  0.202264646  0.2800789093  0.393582015 -0.135825453
#> 14 -0.41424881 -0.066051789  0.2247749561  0.006355641  0.224472846
#> 15  0.10794767 -0.158242485  0.1721660307  0.200750816  0.058498857
#> 16  0.20343088 -0.194765724  0.0885301625 -0.422847412  0.111167817
#> 17  0.39881965  0.221427538  0.5960431512 -0.283792044  0.538758570
#> 18  0.09885279 -0.122910515 -0.2653861763 -0.246257047  0.085860311
#> 19 -0.33971027 -0.376136906  0.5261857568 -0.366215991  0.235809365
#> 20  0.30427866 -0.012924772  0.1132563063 -0.162553474  0.448903839

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>               F1          F2          F3          F4           F5
#> 1  -4.171927e-05 -0.10710657 -0.24877230 -0.06753261  0.184090817
#> 2  -2.123614e-01 -0.38157710  0.18990023 -0.43370922 -0.035565060
#> 3  -1.302529e-02  0.10886420 -0.02593583 -0.20757734  0.009834617
#> 4  -4.286091e-01  0.23033599 -0.17461034  0.02560171 -0.127900628
#> 5   6.478344e-03  0.19532098 -0.20888878  0.19371914  0.283141466
#> 6   8.956452e-02 -0.01502533  0.14691530 -0.09907261 -0.235371518
#> 7  -3.391999e-01  0.08444364  0.08952949 -0.36486633 -0.161504597
#> 8  -3.403857e-02  0.06828915  0.25918971 -0.25625918 -0.256872450
#> 9   2.489135e-02 -0.12728724 -0.43752867 -0.37339619 -0.172126435
#> 10 -3.794833e-01  0.24832775  0.18393376  0.18315309 -0.010349659
#> 11 -1.843564e-01  0.25811305  0.52291546 -0.08626059 -0.160702478
#> 12 -1.897244e-01  0.14549493 -0.09983265  0.06621263 -0.071415422
#> 13 -2.607904e-01  0.16968999  0.13732328 -0.39087754  0.280138172
#> 14 -6.441200e-02  0.07563693 -0.24476012  0.05740406  0.119671507
#> 15  2.007073e-01 -0.10724377 -0.37468914 -0.30804986 -0.041998661
#> 16  1.781883e-01  0.16260053 -0.16235166 -0.23778966  0.223122429
#> 17 -1.103046e-01 -0.06586190  0.21442336 -0.24106958  0.039534202
#> 18 -1.468345e-01  0.19120946 -0.17532994 -0.10876650  0.299871481
#> 19 -4.938341e-02  0.05331162 -0.03974973  0.46580832  0.111031866
#> 20 -1.412484e-01  0.08485122  0.42266052  0.02424422  0.141506316