Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>             F1          F2          F3          F4          F5
#> 1  -0.19259219 -0.34770571  0.29039761  0.38136143 -0.08923038
#> 2   0.25389920 -0.29725828 -0.06794500 -0.35539776  0.02175684
#> 3  -0.01318282  0.36574518  0.13839879  0.04400467 -0.08249660
#> 4  -0.48874376  0.45443536  0.26534463  0.15995477  0.16771076
#> 5  -0.36558779  0.13364133 -0.04839531  0.11952816  0.25940183
#> 6  -0.27468951 -0.13848588  0.57316083 -0.28912627 -0.05344535
#> 7   0.39517140  0.10816830 -0.01202371  0.15818234 -0.12789261
#> 8  -0.13178028  0.19634672  0.27152351  0.35361439 -0.28316741
#> 9   0.19560779  0.41727893 -0.02429464  0.21817747  0.20287283
#> 10  0.29481610 -0.06255537 -0.48080170  0.45826124 -0.11989260
#> 11  0.30669850  0.15399356  0.02449116 -0.10966612  0.42747774
#> 12  0.25680106  0.12641629  0.20456362  0.08691559  0.13228901
#> 13 -0.05550330  0.18633213  0.12473254 -0.07191871 -0.17315464
#> 14 -0.08329060  0.16131644 -0.10005594  0.33719945  0.26221231
#> 15 -0.10356656  0.22833114 -0.26677062 -0.12047848 -0.09729383
#> 16  0.38889129  0.20055410 -0.11137593 -0.37592121 -0.32829918
#> 17  0.24473062  0.59828866  0.04675354  0.10722045  0.07501331
#> 18 -0.06530006  0.11855629 -0.11607232  0.33122316  0.28872628
#> 19 -0.05418363 -0.07238851  0.30548206 -0.13778242  0.10550998
#> 20  0.28643447 -0.07813013  0.06565075 -0.01423465 -0.10330020

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>              F1           F2            F3          F4           F5
#> 1  -0.112027853 -0.082347207 -0.0254274648 -0.20970295 -0.358587130
#> 2  -0.282117953 -0.142911589 -0.1386786160  0.07545922 -0.216184623
#> 3  -0.376166375 -0.165332852 -0.0009323501 -0.50679512 -0.078816908
#> 4  -0.002515632  0.010968504  0.0960984795  0.27267166 -0.116256788
#> 5  -0.439130285 -0.167019447  0.0776297721  0.30359690 -0.064847051
#> 6  -0.033670288  0.215682655  0.1276632309  0.06565739  0.019423622
#> 7   0.019279503 -0.231679266  0.3277478861  0.04799111 -0.005452886
#> 8  -0.334181361 -0.204514299  0.0354231865 -0.01918490  0.011467833
#> 9  -0.066915180 -0.080512544  0.1210098484  0.15667936 -0.358319914
#> 10 -0.375721736 -0.397593223 -0.0022954402  0.03391677  0.103571842
#> 11 -0.514877923 -0.024269791  0.0103010016 -0.21751914  0.113872683
#> 12  0.062827052  0.157557332  0.2922082387 -0.32974017 -0.239940139
#> 13  0.017199188 -0.095256541 -0.1189885875  0.30964027 -0.230369653
#> 14 -0.010986783  0.177097009 -0.1581666993  0.15122841 -0.177918827
#> 15 -0.125967367 -0.009321041  0.4130228040  0.09299932 -0.033100530
#> 16  0.094203082  0.232005119  0.0889342528 -0.08865248 -0.254884682
#> 17  0.019244842  0.182442798 -0.0402866231 -0.20225334  0.081275760
#> 18 -0.002713328 -0.024387226  0.0601826647  0.16004249  0.138115340
#> 19  0.023115534 -0.176910495  0.3682166355 -0.08405417  0.278146048
#> 20 -0.216303543  0.163267552  0.2714633010  0.20792744 -0.148309877