Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>              F1           F2          F3            F4           F5
#> 1   0.015031151 -0.093879784 -0.24581872  0.0549331858 -0.033605826
#> 2  -0.124151394 -0.091757540 -0.07775016  0.3475647445  0.187791359
#> 3   0.060248091 -0.221445578 -0.22312545 -0.2450180647  0.207225271
#> 4   0.008855272  0.184246165  0.04543981 -0.1003353960  0.162310287
#> 5   0.080528917 -0.009450387 -0.00958895  0.2541288881 -0.378916727
#> 6   0.519118083  0.363828804  0.04929003  0.2130365740  0.556604427
#> 7  -0.136758401 -0.011626607 -0.10346508 -0.2101157033  0.142756890
#> 8   0.143250341  0.118772202 -0.05830812 -0.2594812415 -0.081694793
#> 9  -0.063468731 -0.108213070  0.12947098  0.6623018771 -0.119091978
#> 10  0.313822530  0.186305885 -0.35240844 -0.0007117739  0.327153129
#> 11  0.009965746 -0.017637959 -0.11771236  0.1082404601  0.153712928
#> 12  0.094034362  0.042909967 -0.52957663  0.0442916991  0.250422308
#> 13 -0.210374918  0.425066497  0.58623511  0.1278534544 -0.282734803
#> 14  0.083542557 -0.117324925 -0.02108181  0.3430050914  0.292593790
#> 15 -0.140201422  0.156467139  0.40663571  0.0818911405  0.097073314
#> 16  0.021362461 -0.035103576 -0.24926597  0.4291115388  0.208027908
#> 17  0.224282987 -0.087643391 -0.41110612 -0.0923562198  0.207053981
#> 18  0.187945417  0.246275882  0.04534689 -0.1662466419 -0.068717460
#> 19 -0.003575711 -0.108212050 -0.20475327  0.3556753518  0.214966982
#> 20  0.120891438  0.149949703  0.10226344  0.0845836968 -0.009900026

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>              F1          F2           F3           F4          F5
#> 1  -0.357699979 -0.41857300 -0.194697641 -0.004867324  0.21149508
#> 2  -0.032120241 -0.20875940 -0.039375421 -0.307793752  0.26975616
#> 3   0.009282344 -0.11695499 -0.456274979 -0.162786646  0.29942937
#> 4   0.046289453 -0.14707145 -0.001796145 -0.038396737  0.19331479
#> 5  -0.149982424 -0.15263763  0.025471694 -0.166250417 -0.44822560
#> 6  -0.207315858  0.56271075 -0.095782237  0.033454446  0.33865775
#> 7  -0.202544020 -0.41478935 -0.190788270 -0.142595059 -0.27807209
#> 8  -0.090287308 -0.11066903 -0.075474104  0.148149782  0.12713837
#> 9  -0.184565824 -0.10523734 -0.011866131 -0.303938081 -0.20820261
#> 10 -0.222337363  0.19638211  0.017125669 -0.343168155  0.11768500
#> 11  0.045709147  0.05923306  0.039086847  0.191429442  0.19119944
#> 12 -0.067255383 -0.44742711 -0.321824746  0.171411905  0.15748384
#> 13 -0.195617512  0.49196122 -0.031799663  0.077502862  0.01714089
#> 14 -0.133953958 -0.04910395 -0.093189350  0.120179221  0.15895057
#> 15 -0.245090219 -0.15440438 -0.202032238 -0.374619190 -0.38523884
#> 16  0.212233710 -0.25803316  0.314149528 -0.310402323  0.08131782
#> 17 -0.146359599 -0.35466105 -0.066684615 -0.064986443  0.52695217
#> 18 -0.174283863 -0.15228303 -0.492195812  0.120569153  0.06641055
#> 19  0.102629605  0.03220274  0.359107737 -0.096490680 -0.05507876
#> 20 -0.195796708 -0.11167065 -0.191513828 -0.335110364  0.32630524