Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>              F1          F2          F3          F4          F5
#> 1   0.162582830 -0.23265371 -0.11716741  0.13825147 -0.00366079
#> 2   0.280267538  0.23696896  0.12679792 -0.13006804  0.03091977
#> 3   0.317626829  0.44068110 -0.19474158  0.19005267  0.17308051
#> 4   0.267409682 -0.31712920 -0.18272207  0.15036904  0.27084074
#> 5  -0.322727527 -0.04877549  0.02033787  0.07822927  0.29375232
#> 6   0.337919971  0.08803980  0.62145558  0.32804373  0.05890903
#> 7   0.447263868  0.28480658 -0.10187180 -0.12231324  0.23805783
#> 8   0.076969022  0.17833378  0.29628452  0.04273469  0.41399695
#> 9   0.140908742 -0.08793844 -0.18073869  0.08273998 -0.15567622
#> 10 -0.252529483 -0.31966974  0.07946135  0.17131489  0.04026991
#> 11  0.300695405  0.06206589  0.13228688  0.03406019  0.14076703
#> 12  0.007091051 -0.02119961 -0.21698741  0.37135073 -0.20523194
#> 13  0.169587972  0.59154558 -0.13768450  0.09525445 -0.03900281
#> 14  0.002442803 -0.15973584  0.24786000 -0.39446228  0.51617142
#> 15 -0.005348210  0.33573770  0.33985217 -0.25279044  0.27658369
#> 16 -0.045010064  0.12422367  0.10664829  0.31137463  0.12820069
#> 17 -0.206534362 -0.18931844  0.49554986 -0.03883516  0.22854119
#> 18 -0.111420721  0.32579264 -0.09377322  0.19144408  0.29596931
#> 19  0.146348655  0.37046470  0.19162259 -0.06960001 -0.05085829
#> 20  0.196218559 -0.20560395  0.69442804  0.13541422 -0.16142798

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>              F1           F2           F3          F4           F5
#> 1   0.128201806  0.250760350  0.030057261 -0.00669123  0.197332485
#> 2   0.014033634  0.171352883 -0.050602413 -0.08728603  0.304130497
#> 3   0.546356938  0.067707289  0.358347953 -0.20634464 -0.390725639
#> 4  -0.272211784 -0.007914657 -0.069816988 -0.02729844 -0.566605623
#> 5   0.153044189 -0.035247117 -0.096111800 -0.31903431 -0.273915541
#> 6  -0.254541480  0.068940404  0.157603838 -0.20805112 -0.165315923
#> 7   0.200989412 -0.063787775 -0.124533783 -0.04663391 -0.231169981
#> 8  -0.176834318 -0.267086933 -0.137661021  0.06492807 -0.175395602
#> 9   0.018493733 -0.032295523 -0.087334652 -0.25253666 -0.327164850
#> 10 -0.334995444 -0.236501259  0.039192886  0.23366592  0.113358004
#> 11  0.437642450 -0.298900720  0.003023894 -0.03062306  0.002890916
#> 12  0.254779273 -0.464828176 -0.135128109  0.02489411  0.129349818
#> 13 -0.179077194  0.114950337 -0.010773822 -0.07052224  0.092538620
#> 14  0.174563250  0.054464515  0.251957501 -0.21426668 -0.501122513
#> 15  0.232664597  0.427755708  0.042607973  0.20728659  0.206079553
#> 16  0.058766125 -0.047410401 -0.149770420 -0.44351525 -0.034893751
#> 17  0.202571933  0.118191571  0.176087527  0.16353492 -0.096645669
#> 18  0.186296467 -0.017242921 -0.374966583 -0.51175463 -0.648775589
#> 19  0.007064004  0.149563929 -0.124432751 -0.09443630  0.037702835
#> 20  0.002345882  0.262852862 -0.057654133 -0.17127846  0.032727414