Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>             F1          F2           F3           F4           F5
#> 1   0.45952706  0.31828620  0.521025689 -0.017737111 -0.207201854
#> 2   0.12685959 -0.11561912  0.012754217  0.157278700  0.288909807
#> 3  -0.04870776  0.17274092  0.265957440  0.414788396 -0.009854215
#> 4   0.22924483 -0.28801414  0.417347079 -0.228662146  0.140192172
#> 5   0.21435333  0.69129381  0.119159432  0.331344085 -0.150070508
#> 6   0.02460727  0.51929289  0.164525939  0.015914502 -0.142178401
#> 7  -0.01382663  0.17669148  0.234854112 -0.080567745 -0.075594229
#> 8   0.30695506  0.15101415  0.161281830  0.039060572 -0.399340096
#> 9   0.15469256 -0.32964487  0.182983132  0.417278523 -0.330975726
#> 10 -0.38189806  0.38849666  0.207952314  0.065088664  0.135031277
#> 11  0.07553692  0.17028788  0.201719829 -0.101086788  0.123527199
#> 12  0.13650858  0.22965403  0.159118686  0.356994521 -0.260111404
#> 13 -0.08742914  0.08748481  0.115125556  0.464181743 -0.141801603
#> 14  0.14049112 -0.24972253 -0.006340555  0.212371352  0.035168785
#> 15  0.02597174  0.21489022  0.008672529  0.256388503 -0.235255282
#> 16 -0.16544720  0.19329164  0.363275938  0.378975085  0.023854779
#> 17  0.15819223 -0.43704286  0.569204864 -0.001819605 -0.185888748
#> 18 -0.15251987  0.02169928 -0.222771540  0.177870773 -0.188861784
#> 19 -0.21729478  0.13980574  0.467036840  0.190030486  0.085436952
#> 20 -0.13710062  0.01642041 -0.017963506  0.098090915 -0.470595562

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>             F1           F2          F3           F4           F5
#> 1  -0.36155142 -0.300717548  0.03426601  0.130565221  0.097068099
#> 2  -0.30132988 -0.117987972 -0.16546864  0.037675945 -0.079937791
#> 3  -0.42159123 -0.102952464 -0.40769445  0.143226486  0.167952183
#> 4   0.30575739 -0.312031198 -0.25620188  0.043833648 -0.205773182
#> 5  -0.17125003 -0.134226419 -0.30780070  0.142693021  0.448428985
#> 6   0.09778731 -0.284741110 -0.10715153  0.052719824 -0.295474852
#> 7  -0.05751713 -0.197821201  0.15441668  0.303182932  0.201907645
#> 8   0.23154779 -0.420997698  0.16127517 -0.001795102 -0.135449554
#> 9  -0.23749248 -0.079708454 -0.24455097  0.094097072  0.058955251
#> 10  0.08342504 -0.207757592 -0.01661114 -0.148055493 -0.089557268
#> 11 -0.39990905 -0.116881990 -0.20679628 -0.201414108 -0.119711372
#> 12 -0.55695527 -0.193304305  0.23984539  0.194291330  0.187108603
#> 13 -0.17089590 -0.342446095 -0.08136256 -0.135202582  0.017930654
#> 14 -0.26817002  0.043738915 -0.33928371 -0.334726443 -0.303985502
#> 15 -0.21157775 -0.216967353  0.18892818 -0.117318822  0.064947262
#> 16  0.26171014 -0.002387317 -0.05655113  0.402439096 -0.259999036
#> 17 -0.20417389 -0.152867528 -0.03150072  0.066379705  0.187553456
#> 18 -0.20135846  0.351605627 -0.39302833  0.006903172  0.008324486
#> 19  0.08785740 -0.029309699  0.07912735 -0.104303266  0.080730992
#> 20 -0.13749116  0.130824565  0.22201823 -0.159981134  0.254362376