Predict New Coordinates

Predict the projection of new individuals/rows or variables/columns.

Usage

# S4 method for CA
predict(object, newdata, margin = 1)

# S4 method for PCA
predict(object, newdata, margin = 1)

Arguments

object

A CA or PCA object.

newdata

An object of supplementary points coercible to a matrix for which to compute principal coordinates.

margin

A length-one numeric vector giving the subscript which the data will be predicted: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

Value

A data.frame of coordinates.

See also

Author

N. Frerebeau

Examples

## Create a matrix
A <- matrix(data = sample(1:10, 100, TRUE), nrow = 10, ncol = 10)

## Compute correspondence analysis
X <- ca(A, sup_row = 8:10, sup_col = 7:10)

## Predict new row coordinates
Y <- matrix(data = sample(1:10, 120, TRUE), nrow = 20, ncol = 6)
predict(X, Y, margin = 1)
#>             F1          F2          F3          F4          F5
#> 1   0.01902940  0.26741123  0.06229576  0.58910819 -0.39027087
#> 2  -0.35157763  0.21295636 -0.04961846 -0.21306218  0.16066801
#> 3  -0.11172473  0.41459767  0.25133524  0.06573648 -0.24164298
#> 4   0.31380558 -0.25513096 -0.53352789  0.30383239 -0.12930160
#> 5   0.08765077  0.78540210  0.30504918  0.10552292 -0.52953846
#> 6  -0.21769453  0.28205305  0.18490012  0.04452441 -0.06454854
#> 7   0.21759789 -0.02823393 -0.01962753  0.01410808 -0.01779875
#> 8   0.47674067  0.11188437  0.09919843  0.07012721 -0.05483874
#> 9   0.53220544 -0.22017748  0.12951234  0.15430954  0.23799424
#> 10  0.23301819  0.52904513 -0.30913344  0.05152453 -0.15550369
#> 11 -0.12223590  0.01465750  0.04164649  0.28310913 -0.25248969
#> 12  0.05458232  0.46519155  0.08681422 -0.32321214 -0.22558341
#> 13  0.04825337  0.14277045  0.19015702 -0.13509559 -0.10121899
#> 14 -0.13599121  0.34187572  0.14970499 -0.03720439 -0.20117222
#> 15  0.10585355  0.26022389  0.29944339 -0.31651446 -0.13091116
#> 16  0.01162517  0.02804532  0.13553770  0.22874649 -0.05489591
#> 17  0.29806743  0.47404952  0.07893001  0.04289141 -0.13325462
#> 18  0.05388119 -0.26084213  0.46830483  0.04885476  0.61292960
#> 19  0.18233254 -0.07931418  0.34367498  0.18309674  0.08353582
#> 20 -0.25126515 -0.03749193 -0.16696282 -0.07571598  0.36286530

## Predict new column coordinates
Z <- matrix(data = sample(1:10, 140, TRUE), nrow = 7, ncol = 20)
predict(X, Z, margin = 2)
#>              F1          F2          F3          F4          F5
#> 1  -0.144606302 -0.17937468  0.17798334 -0.14957616  0.08090004
#> 2  -0.266431746  0.05449727  0.20050459 -0.18958778  0.17620985
#> 3  -0.329948627 -0.36257024  0.10343122 -0.12043406 -0.28250016
#> 4   0.133641745 -0.31264116  0.14665027 -0.24674689 -0.15858326
#> 5  -0.243291175 -0.12899505 -0.17067111 -0.10720534 -0.21433291
#> 6  -0.177294410  0.42029120 -0.02048229  0.05493606 -0.27047137
#> 7  -0.211447118  0.10317697  0.20138014 -0.28103380  0.11458538
#> 8   0.060145662  0.09744734  0.26067152  0.25831862 -0.24106644
#> 9  -0.551612463 -0.18337026  0.19523102 -0.29723731  0.11600422
#> 10 -0.367803054 -0.16008489  0.11187436 -0.24906316  0.07897495
#> 11 -0.044444831 -0.44825123  0.30703204  0.12535960 -0.14099868
#> 12  0.007928875 -0.21253318  0.04489412 -0.30046562 -0.18557973
#> 13 -0.345866534 -0.14663732  0.12678623 -0.47431844 -0.03414371
#> 14 -0.078029779  0.33460628 -0.13345771 -0.48391871  0.11871483
#> 15 -0.851812976  0.12000967  0.05675247 -0.57201774  0.15881639
#> 16 -0.077762171 -0.03988073  0.29175311 -0.19162510 -0.03481283
#> 17 -0.055442916  0.02900402 -0.07758486 -0.19394909 -0.28895664
#> 18 -0.220649229 -0.38582247  0.20134718  0.05042437  0.06924263
#> 19 -0.238971078  0.01862691  0.20583601 -0.07970844 -0.13240843
#> 20 -0.267993050 -0.18791520 -0.13482491  0.07680061 -0.13883811