Extract Parts of an Object

Operators acting on objects to extract parts.

Usage

# S4 method for CA,ANY,missing
[[(x, i)

# S4 method for PCA,ANY,missing
[[(x, i)

Arguments

x

An object from which to extract element(s) or in which to replace element(s).

i

A character string specifying elements to extract. Any unambiguous substring can be given (see details).

Value

A list.

Details

If i is "data", returns a list with the following elements:

data

A numeric matrix of raw data.

mean

A numeric vector giving the variables means (PCA).

sd

A numeric vector giving the variables standard deviations (PCA).

If i is "rows", returns a list with the following elements:

coord

A numeric matrix of rows/individuals coordinates.

cos2

A numeric matrix of rows/individuals squared cosine.

masses

A numeric vector giving the rows masses/individual weights.

sup

A logical vector specifying whether a point is a supplementary observation or not.

If i is "columns", returns a list with the following elements:

coord

A numeric matrix of columns/variables coordinates.

cor

A numeric matrix of correlation between variables and the dimensions (PCA).

cos2

A numeric matrix of columns/variables squared cosine.

masses

A numeric vector giving the columns masses/variable weights.

sup

A logical vector specifying whether a point is a supplementary observation or not.

If i is "eigenvalues", returns a numeric vector of eigenvalues.

See also

Author

N. Frerebeau

Examples

## Load data
data("iris")

## Compute principal components analysis
X <- pca(iris, scale = TRUE, sup_row = 8:10, sup_col = 1)
#> 1 qualitative variable was removed: Species.

## Get results for the individuals
X[["rows"]]
#> $coordinates
#>             F1           F2
#> 1   -2.1799607  0.107860619
#> 2   -1.6996139 -0.926588874
#> 3   -1.9288393 -0.527485410
#> 4   -1.7585967 -0.705022642
#> 5   -2.2760301  0.314750517
#> 6   -2.2841520  1.069266219
#> 7   -1.9994783 -0.054120776
#> 11  -2.3350128  0.536316749
#> 12  -2.0097181 -0.069676614
#> 13  -1.7840271 -0.971497378
#> 14  -1.8952870 -1.015526377
#> 15  -2.7344808  1.112957446
#> 16  -2.8386721  2.074363046
#> 17  -2.4324986  1.010560888
#> 18  -2.0955476  0.152769123
#> 19  -2.2724958  0.817467817
#> 20  -2.3466691  0.788115151
#> 21  -1.9726315 -0.055000281
#> 22  -2.1661866  0.626133757
#> 23  -2.4243767  0.256045186
#> 24  -1.6233228 -0.127164668
#> 25  -1.8984582 -0.025647615
#> 26  -1.6254407 -0.897236208
#> 27  -1.8408919  0.020140394
#> 28  -2.1428741  0.122536952
#> 29  -2.0838914 -0.099029280
#> 30  -1.8175794 -0.483456411
#> 31  -1.7215100 -0.690346310
#> 32  -1.8779785  0.005464061
#> 33  -2.8037034  1.318967839
#> 34  -2.8524463  1.556089908
#> 35  -1.7585967 -0.705022642
#> 36  -1.9659259 -0.542161743
#> 37  -2.2170474  0.093184286
#> 38  -2.3604432  0.269842013
#> 39  -1.7367006 -0.941265207
#> 40  -2.0468047 -0.084352947
#> 41  -2.1326343  0.138092790
#> 42  -0.9798020 -2.344585993
#> 43  -1.9288393 -0.527485410
#> 44  -1.7681350  0.316847300
#> 45  -2.1139094  0.891728987
#> 46  -1.6152008 -0.881680370
#> 47  -2.3939955  0.757882980
#> 48  -1.8917527 -0.512809077
#> 49  -2.3350128  0.536316749
#> 50  -1.9878220 -0.305919178
#> 51   0.3450638  0.510411957
#> 52   0.3553036  0.525967795
#> 53   0.5997195  0.377783228
#> 54   0.8656684 -1.499239964
#> 55   0.7766677 -0.286915466
#> 56   0.5707548 -0.391408807
#> 57   0.4178206  0.807118863
#> 58   0.2567533 -1.529809908
#> 59   0.5117721 -0.169842576
#> 60   0.5287175 -0.641448199
#> 61   0.7152040 -2.328016836
#> 62   0.4361824  0.068158999
#> 63   0.7084985 -1.840855375
#> 64   0.6332719 -0.110257739
#> 65   0.1409057 -0.316605905
#> 66   0.3298732  0.259493060
#> 67   0.5474423  0.112187998
#> 68   0.2652383 -0.791729549
#> 69   1.3159972 -1.542931190
#> 70   0.4676168 -1.189953508
#> 71   0.7198029  0.704722305
#> 72   0.3853216 -0.464790472
#> 73   1.1761357 -0.863556163
#> 74   0.5605150 -0.406964645
#> 75   0.4005122 -0.213871574
#> 76   0.4259426  0.052603161
#> 77   0.7664279 -0.302471304
#> 78   0.9017017  0.275386670
#> 79   0.6435117 -0.094701901
#> 80   0.1387878 -1.086677445
#> 81   0.5265996 -1.411519740
#> 82   0.4050998 -1.471104576
#> 83   0.3598912 -0.731265207
#> 84   1.1425834 -0.375515196
#> 85   0.5474423  0.112187998
#> 86   0.2475780  0.984656096
#> 87   0.5255462  0.348430562
#> 88   1.0140150 -1.440534633
#> 89   0.2302695 -0.036334342
#> 90   0.6735297 -1.085460167
#> 91   0.6413938 -0.864773441
#> 92   0.5001159  0.081955827
#> 93   0.4930472 -0.923478773
#> 94   0.3528226 -1.736699807
#> 95   0.5555643 -0.642327704
#> 96   0.1829431 -0.066566513
#> 97   0.3634255 -0.228547907
#> 98   0.4005122 -0.213871574
#> 99   0.1338371 -1.322040504
#> 100  0.4224083 -0.450114139
#> 101  1.6596649  1.402087727
#> 102  1.3958227 -0.240789684
#> 103  1.5731339  0.587107682
#> 104  1.3047041  0.201463273
#> 105  1.6204604  0.617339853
#> 106  1.8327404  0.689842013
#> 107  1.1966153 -0.832444487
#> 108  1.5643105  0.304197604
#> 109  1.7631548 -0.596743655
#> 110  1.4085435  2.037433755
#> 111  0.9998890  0.838568312
#> 112  1.4699960 -0.211437019
#> 113  1.4247874  0.528402351
#> 114  1.6352879 -0.624337311
#> 115  1.7218189  0.190642734
#> 116  1.3273016  1.002646490
#> 117  1.1715481  0.393676839
#> 118  1.1856853  2.404546038
#> 119  2.4971040 -0.003871575
#> 120  1.5014304 -1.469549526
#> 121  1.4756482  1.061351821
#> 122  1.3099931 -0.018343948
#> 123  1.9775527  0.245830045
#> 124  1.2372363 -0.315050854
#> 125  1.2107526  1.178424712
#> 126  1.1648425  0.880838301
#> 127  1.1040803 -0.122837289
#> 128  0.9490282  0.305618841
#> 129  1.6540127  0.129298887
#> 130  1.1139818  0.347888830
#> 131  1.6706197  0.112863543
#> 132  0.9055992  2.270700031
#> 133  1.7384258  0.174207390
#> 134  0.9621009 -0.213533802
#> 135  1.2552597 -0.598840438
#> 136  1.8161334  0.706277356
#> 137  1.3308359  1.505363789
#> 138  1.0754787  0.600566737
#> 139  0.9119416  0.290942508
#> 140  1.2916314  0.720615916
#> 141  1.6190440  0.884694094
#> 142  1.3491977  0.766403925
#> 143  1.3958227 -0.240789684
#> 144  1.5498214  1.090704487
#> 145  1.5484050  1.358058728
#> 146  1.4823537  0.574190360
#> 147  1.5508748 -0.669245814
#> 148  1.2291144  0.439464848
#> 149  1.1722495  1.431102620
#> 150  1.0232015  0.334971507
#> 8   -2.0468047 -0.084352947
#> 9   -1.6035446 -1.133478772
#> 10  -1.8430098 -0.749931146
#> 
#> $contributions
#>              F1           F2
#> 1   1.453343684 0.0106993487
#> 2   0.883428800 0.7895973615
#> 3   1.137792525 0.2558894786
#> 4   0.945809064 0.4571278240
#> 5   1.584261926 0.0911096611
#> 6   1.595588876 1.0514853344
#> 7   1.222656219 0.0026937641
#> 11  1.667437334 0.2645295822
#> 12  1.235211341 0.0044648369
#> 13  0.973360803 0.8679901399
#> 14  1.098552806 0.9484489111
#> 15  2.286760836 1.1391702764
#> 16  2.464344658 3.9573190026
#> 17  1.809573309 0.9391965689
#> 18  1.342969212 0.0214636064
#> 19  1.579345535 0.6145721192
#> 20  1.684126390 0.5712298316
#> 21  1.190043653 0.0027820270
#> 22  1.435035728 0.3605502185
#> 23  1.797509409 0.0602927030
#> 24  0.805899240 0.0148718312
#> 25  1.102232091 0.0006049584
#> 26  0.808003463 0.7403636858
#> 27  1.036400344 0.0003730503
#> 28  1.404314222 0.0138091094
#> 29  1.328070529 0.0090190079
#> 30  1.010317225 0.2149543180
#> 31  0.906337800 0.4382940161
#> 32  1.078579659 0.0000274576
#> 33  2.404003572 1.5999261560
#> 34  2.488318246 2.2269009410
#> 35  0.945809064 0.4571278240
#> 36  1.181966834 0.2703268996
#> 37  1.503214416 0.0079857717
#> 38  1.703954895 0.0669654302
#> 39  0.922403375 0.8148084750
#> 40  1.281220300 0.0065438305
#> 41  1.390925123 0.0175377362
#> 42  0.293594279 5.0554970160
#> 43  1.137792525 0.2558894786
#> 44  0.956096741 0.0923276004
#> 45  1.366607256 0.7313028984
#> 46  0.797855133 0.7149141115
#> 47  1.752740637 0.5282455348
#> 48  1.094459487 0.2418482414
#> 49  1.667437334 0.2645295822
#> 50  1.208442464 0.0860686388
#> 51  0.036414113 0.2395924925
#> 52  0.038607369 0.2544191706
#> 53  0.109993587 0.1312553072
#> 54  0.229178638 2.0671587361
#> 55  0.184476671 0.0757075942
#> 56  0.099625421 0.1408940695
#> 57  0.053388873 0.5991099343
#> 58  0.020160574 2.1523181354
#> 59  0.080098468 0.0265292309
#> 60  0.085490582 0.3784031185
#> 61  0.156433881 4.9842952461
#> 62  0.058184494 0.0042724593
#> 63  0.153514287 3.1165268007
#> 64  0.122645427 0.0111802030
#> 65  0.006071949 0.0921869714
#> 66  0.033278603 0.0619274246
#> 67  0.091653220 0.0115750885
#> 68  0.021515104 0.5764813136
#> 69  0.529640167 2.1893976274
#> 70  0.066873073 1.3022414468
#> 71  0.158452140 0.4567384375
#> 72  0.045406471 0.1986762522
#> 73  0.423044289 0.6858240370
#> 74  0.096082758 0.1523157768
#> 75  0.049057166 0.0420666239
#> 76  0.055484680 0.0025448091
#> 77  0.179644343 0.0841394918
#> 78  0.248654774 0.0697456870
#> 79  0.126643778 0.0082480050
#> 80  0.005890792 1.0860075196
#> 81  0.084807053 1.8323373666
#> 82  0.050187444 1.9903003497
#> 83  0.039610792 0.4917918728
#> 84  0.399251696 0.1296840484
#> 85  0.091653220 0.0115750885
#> 86  0.018745415 0.8916628530
#> 87  0.084468126 0.1116513580
#> 88  0.314455886 1.9084418605
#> 89  0.016216010 0.0012141326
#> 90  0.138734505 1.0835758275
#> 91  0.125811545 0.6877588893
#> 92  0.076491334 0.0061771934
#> 93  0.074344362 0.7843056805
#> 94  0.038070081 2.7738376469
#> 95  0.094392957 0.3794415052
#> 96  0.010235356 0.0040751457
#> 97  0.040392607 0.0480381224
#> 98  0.049057166 0.0420666239
#> 99  0.005478023 1.6073892100
#> 100 0.054567719 0.1863274482
#> 101 0.842387320 1.8079312085
#> 102 0.595842510 0.0533221045
#> 103 0.756837054 0.3170056865
#> 104 0.520589025 0.0373270182
#> 105 0.803059702 0.3504936509
#> 106 1.027242319 0.4376539041
#> 107 0.437905211 0.6372973156
#> 108 0.748370960 0.0851026540
#> 109 0.950718311 0.3274968644
#> 110 0.606752411 3.8176711884
#> 111 0.305755696 0.6467082790
#> 112 0.660850565 0.0411143642
#> 113 0.620827724 0.2567798899
#> 114 0.817823172 0.3584842707
#> 115 0.906663024 0.0334250480
#> 116 0.538778450 0.9245431850
#> 117 0.419750477 0.1425316314
#> 118 0.429941982 5.3173802664
#> 119 1.906971687 0.0000137850
#> 120 0.689416059 1.9860948268
#> 121 0.665942333 1.0359773621
#> 122 0.524818373 0.0003094690
#> 123 1.195988781 0.0555778122
#> 124 0.468140574 0.0912836189
#> 125 0.448313474 1.2771303006
#> 126 0.414959229 0.7135491729
#> 127 0.372796904 0.0138768844
#> 128 0.275441407 0.0858997255
#> 129 0.836659379 0.0153752114
#> 130 0.379513426 0.1113044416
#> 131 0.853544550 0.0117149082
#> 132 0.250808937 4.7418854471
#> 133 0.924236944 0.0279103111
#> 134 0.283081973 0.0419338551
#> 135 0.481879173 0.3298023644
#> 136 1.008710388 0.4587563529
#> 137 0.541651566 2.0840803395
#> 138 0.353732210 0.3317065721
#> 139 0.254334345 0.0778477176
#> 140 0.510209050 0.4775724073
#> 141 0.801656426 0.7198098450
#> 142 0.556701191 0.5401905310
#> 143 0.595842510 0.0533221045
#> 144 0.734571902 1.0940715530
#> 145 0.733229824 1.6961670731
#> 146 0.672008322 0.3032098599
#> 147 0.735570728 0.4119104916
#> 148 0.462014453 0.1776150699
#> 149 0.420253277 1.8835323329
#> 150 0.320179346 0.1031922926
#> 
#> $cos2
#>              F1           F2
#> 1   0.997537883 2.442069e-03
#> 2   0.770275213 2.289389e-01
#> 3   0.929285621 6.949892e-02
#> 4   0.861533164 1.384668e-01
#> 5   0.981140144 1.876324e-02
#> 6   0.820250332 1.797497e-01
#> 7   0.997006440 7.304536e-04
#> 11  0.948921305 5.006041e-02
#> 12  0.997254638 1.198698e-03
#> 13  0.771015900 2.286355e-01
#> 14  0.775718588 2.227084e-01
#> 15  0.857879117 1.421129e-01
#> 16  0.651874030 3.480990e-01
#> 17  0.849498051 1.466163e-01
#> 18  0.993184069 5.278443e-03
#> 19  0.884472859 1.144510e-01
#> 20  0.898639841 1.013588e-01
#> 21  0.995527356 7.739112e-04
#> 22  0.920782450 7.693067e-02
#> 23  0.985634960 1.099383e-02
#> 24  0.983278582 6.033923e-03
#> 25  0.988580295 1.804280e-04
#> 26  0.766287189 2.334871e-01
#> 27  0.996619952 1.192913e-04
#> 28  0.996181238 3.257461e-03
#> 29  0.997744694 2.253183e-03
#> 30  0.933168940 6.602192e-02
#> 31  0.861052131 1.384664e-01
#> 32  0.993951874 8.414246e-06
#> 33  0.814686002 1.802995e-01
#> 34  0.769931500 2.291324e-01
#> 35  0.861533164 1.384668e-01
#> 36  0.926583440 7.047060e-02
#> 37  0.998037699 1.763122e-03
#> 38  0.984800233 1.287005e-02
#> 39  0.771150422 2.265235e-01
#> 40  0.997960094 1.694966e-03
#> 41  0.992507874 4.161436e-03
#> 42  0.146925900 8.413062e-01
#> 43  0.929285621 6.949892e-02
#> 44  0.946393484 3.039071e-02
#> 45  0.848161656 1.509286e-01
#> 46  0.765582021 2.281188e-01
#> 47  0.906464208 9.084652e-02
#> 48  0.931348538 6.843763e-02
#> 49  0.948921305 5.006041e-02
#> 50  0.976803096 2.313478e-02
#> 51  0.277995466 6.082479e-01
#> 52  0.311683132 6.830184e-01
#> 53  0.664146963 2.635440e-01
#> 54  0.249361414 7.479428e-01
#> 55  0.878241702 1.198537e-01
#> 56  0.637260992 2.996949e-01
#> 57  0.210729871 7.863598e-01
#> 58  0.027308118 9.694709e-01
#> 59  0.758623302 8.355378e-02
#> 60  0.387926000 5.709851e-01
#> 61  0.086194377 9.132535e-01
#> 62  0.926162213 2.261503e-02
#> 63  0.128022840 8.642691e-01
#> 64  0.898327863 2.723157e-02
#> 65  0.130964651 6.612032e-01
#> 66  0.593595505 3.673229e-01
#> 67  0.958380776 4.024891e-02
#> 68  0.090772461 8.087892e-01
#> 69  0.420435646 5.779399e-01
#> 70  0.133260226 8.629393e-01
#> 71  0.504882110 4.839482e-01
#> 72  0.406578554 5.915782e-01
#> 73  0.645293645 3.478754e-01
#> 74  0.519247169 2.737241e-01
#> 75  0.734870489 2.095489e-01
#> 76  0.958138539 1.461335e-02
#> 77  0.812191218 1.264981e-01
#> 78  0.912896321 8.514938e-02
#> 79  0.978657862 2.119510e-02
#> 80  0.016044824 9.836341e-01
#> 81  0.122144323 8.775790e-01
#> 82  0.070309187 9.272035e-01
#> 83  0.194938448 8.048312e-01
#> 84  0.891022209 9.624269e-02
#> 85  0.958380776 4.024891e-02
#> 86  0.059437245 9.401643e-01
#> 87  0.672034771 2.953945e-01
#> 88  0.330736771 6.674844e-01
#> 89  0.946744678 2.357189e-02
#> 90  0.277285564 7.201808e-01
#> 91  0.342463747 6.225436e-01
#> 92  0.891508689 2.394109e-02
#> 93  0.221487022 7.770072e-01
#> 94  0.039508517 9.572545e-01
#> 95  0.426831711 5.705601e-01
#> 96  0.490691231 6.496627e-02
#> 97  0.698527791 2.762532e-01
#> 98  0.734870489 2.095489e-01
#> 99  0.009680752 9.445952e-01
#> 100 0.467996371 5.314018e-01
#> 101 0.576108690 4.111629e-01
#> 102 0.961696898 2.861893e-02
#> 103 0.877177165 1.221776e-01
#> 104 0.958249977 2.284791e-02
#> 105 0.870879776 1.263950e-01
#> 106 0.851672810 1.206619e-01
#> 107 0.657810722 3.183482e-01
#> 108 0.886186912 3.351128e-02
#> 109 0.885499558 1.014339e-01
#> 110 0.321916434 6.735496e-01
#> 111 0.577480507 4.061726e-01
#> 112 0.978244776 2.023845e-02
#> 113 0.873487112 1.201395e-01
#> 114 0.848034409 1.236129e-01
#> 115 0.885718143 1.085828e-02
#> 116 0.607788840 3.468240e-01
#> 117 0.884632536 9.989009e-02
#> 118 0.191705611 7.884282e-01
#> 119 0.992440547 2.385649e-06
#> 120 0.508966080 4.875811e-01
#> 121 0.651305390 3.369285e-01
#> 122 0.949464886 1.861772e-04
#> 123 0.939398567 1.451656e-02
#> 124 0.929464957 6.026831e-02
#> 125 0.513528277 4.864713e-01
#> 126 0.594651369 3.400324e-01
#> 127 0.968111697 1.198353e-02
#> 128 0.898455967 9.317487e-02
#> 129 0.989867285 6.049073e-03
#> 130 0.787377903 7.679059e-02
#> 131 0.967682818 4.416571e-03
#> 132 0.132869218 8.353566e-01
#> 133 0.977445867 9.815525e-03
#> 134 0.899008289 4.428491e-02
#> 135 0.716813201 1.631402e-01
#> 136 0.867820065 1.312456e-01
#> 137 0.428750425 5.485781e-01
#> 138 0.748409494 2.333771e-01
#> 139 0.890507966 9.063963e-02
#> 140 0.754859896 2.349616e-01
#> 141 0.744129523 2.221864e-01
#> 142 0.695525700 2.244282e-01
#> 143 0.961696898 2.861893e-02
#> 144 0.666268263 3.299894e-01
#> 145 0.547542725 4.211977e-01
#> 146 0.810397292 1.215923e-01
#> 147 0.830568366 1.546654e-01
#> 148 0.874671814 1.118173e-01
#> 149 0.391519529 5.835188e-01
#> 150 0.903130625 9.679300e-02
#> 8   0.997960094 1.694966e-03
#> 9   0.666045138 3.327885e-01
#> 10  0.856123843 1.417503e-01
#> 
#> $masses
#>           1           2           3           4           5           6 
#> 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 
#>           7          14          15          16          17          18 
#> 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 
#>          19          20          21          22          23          24 
#> 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 
#>          25          26          27          28          29          30 
#> 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 
#>          31          32          33          34          35          36 
#> 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 
#>          37          38          39          40          41          42 
#> 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 
#>          43          44          45          46          47          48 
#> 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 
#>          49          50          51          52          53          54 
#> 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 
#>          55          56          57          58          59          60 
#> 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 
#>          61          62          63          64          65          66 
#> 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 
#>          67          68          69          70          71          72 
#> 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 
#>          73          74          75          76          77          78 
#> 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 
#>          79          80          81          82          83          84 
#> 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 
#>          85          86          87          88          89          90 
#> 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 
#>          91          92          93          94          95          96 
#> 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 
#>          97          98          99         100         101         102 
#> 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 
#>         103         104         105         106         107         108 
#> 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 
#>         109         110         111         112         113         114 
#> 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 
#>         115         116         117         118         119         120 
#> 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 
#>         121         122         123         124         125         126 
#> 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 
#>         127         128         129         130         131         132 
#> 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 
#>         133         134         135         136         137         138 
#> 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 
#>         139         140         141         142         143         144 
#> 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 
#>         145         146         147         148         149         150 
#> 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 0.006802721 
#>           8           9          10 
#> 0.006802721 0.006802721 0.006802721 
#> 
#> $supplement
#>   [1] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [13] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [25] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [37] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [49] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [61] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [73] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [85] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#>  [97] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [109] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [121] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [133] FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
#> [145] FALSE FALSE FALSE  TRUE  TRUE  TRUE
#>