Object Summaries

Provides a summary of the results of a multivariate data analysis.

# S4 method for CA
summary(object, margin = 1, active = TRUE, sup = TRUE, rank = 3)

# S4 method for PCA
summary(object, margin = 1, active = TRUE, sup = TRUE, rank = 3)

Arguments

object

A CA or PCA object.

margin

A length-one numeric vector giving the subscript which the data will be summarized: 1 indicates individuals/rows (the default), 2 indicates variables/columns.

active

A logical scalar: should the active observations be summarized?

sup

A logical scalar: should the supplementary observations be summarized?

rank

An integer value specifying the maximal number of components to be kept in the results.

Author

N. Frerebeau

Examples

## Load data
data("zuni", package = "folio")

## Compute correspondence analysis
X <- ca(zuni, rank = 5, sup_row = 1:50)

## Rows summary
summary(X, margin = 1)
#> --- Correspondence Analysis (CA) ------------------------------------------------
#> 
#> Eigenvalues:
#>    eigenvalues variance cumulative
#> F1       0.795   39.122     39.122
#> F2       0.547   26.948     66.071
#> F3       0.305   15.018     81.088
#> F4       0.218   10.748     91.837
#> F5       0.166    8.163    100.000
#> 
#> Active rows:
#>        inertia CA1_coord CA1_contrib CA1_cos2 CA2_coord CA2_contrib CA2_cos2
#> LZ1009   3.883     0.880       0.037    0.076     1.167       0.094    0.133
#> LZ1008   2.198     0.560       0.011    0.039     1.717       0.146    0.364
#> LZ1007   5.146     0.120       0.001    0.001     1.001       0.089    0.095
#> LZ1004   4.303     0.107       0.001    0.002     0.642       0.057    0.073
#> LZ1001   1.833     1.562       0.100    0.433     0.189       0.002    0.006
#> LZ0988  11.161     2.861       0.950    0.676    -1.884       0.598    0.293
#> LZ0987   1.865     0.321       0.004    0.015     1.184       0.069    0.204
#> LZ0986   5.433     0.810       0.045    0.065     2.080       0.429    0.432
#> LZ0983  19.695     2.861       1.676    0.676    -1.884       1.055    0.293
#> LZ0982  29.543     2.861       2.514    0.676    -1.884       1.582    0.293
#>        CA3_coord CA3_contrib CA3_cos2
#> LZ1009    -0.891       0.099    0.078
#> LZ1008    -1.180       0.124    0.172
#> LZ1007    -1.122       0.202    0.119
#> LZ1004    -0.446       0.049    0.035
#> LZ1001     0.976       0.102    0.169
#> LZ0988    -0.536       0.087    0.024
#> LZ0987    -0.444       0.018    0.029
#> LZ0986    -1.083       0.208    0.117
#> LZ0983    -0.536       0.153    0.024
#> LZ0982    -0.536       0.230    0.024
#> (370 more)
#> 
#> Supplementary rows:
#>        CA1_coord CA1_cos2 CA2_coord CA2_cos2 CA3_coord CA3_cos2
#> LZ1105     0.056    0.001     0.920    0.165    -1.042    0.211
#> LZ1103     0.318    0.067     0.396    0.105    -0.469    0.147
#> LZ1100     0.134    0.014     0.626    0.295     0.106    0.008
#> LZ1099     0.841    0.054     2.064    0.323    -0.523    0.021
#> LZ1097     0.617    0.041     0.774    0.064    -0.465    0.023
#> LZ1096     2.246    0.696    -0.788    0.086    -0.791    0.086
#> LZ1095     0.293    0.024     1.078    0.330    -0.322    0.029
#> LZ1092     0.542    0.074     1.288    0.420    -0.115    0.003
#> LZ1091     0.270    0.034     0.961    0.427    -0.028    0.000
#> LZ1090     2.656    0.728    -1.538    0.244    -0.448    0.021
#> (50 more)

## Columns summary
summary(X, margin = 2)
#> --- Correspondence Analysis (CA) ------------------------------------------------
#> 
#> Eigenvalues:
#>    eigenvalues variance cumulative
#> F1       0.795   39.122     39.122
#> F2       0.547   26.948     66.071
#> F3       0.305   15.018     81.088
#> F4       0.218   10.748     91.837
#> F5       0.166    8.163    100.000
#> 
#> Active columns:
#>      inertia CA1_coord CA1_contrib CA1_cos2 CA2_coord CA2_contrib CA2_cos2
#> LINO 651.912     2.550      62.455    0.761    -1.394      27.076    0.227
#> KIAT 148.177     1.246       2.382    0.128     0.492       0.539    0.020
#> RED  322.621     1.376       7.560    0.186     1.403      11.403    0.193
#> GALL 150.034     0.997       1.823    0.097     1.795       8.590    0.313
#> ESC  119.395     0.537       0.395    0.026     1.264       3.183    0.146
#> PUBW 255.598     0.632       2.620    0.081     1.445      19.891    0.426
#> RES  159.994     0.604       1.505    0.075     1.339      10.740    0.367
#> TULA 143.094    -0.367       3.882    0.216    -0.063       0.168    0.006
#> PINE  23.086    -0.476       0.372    0.128    -0.229       0.125    0.030
#> PUBR 147.602     0.731       0.846    0.046     1.619       6.024    0.223
#>      CA3_coord CA3_contrib CA3_cos2
#> LINO    -0.296       2.195    0.010
#> KIAT     1.441       8.301    0.171
#> RED      2.353      57.559    0.544
#> GALL     0.162       0.126    0.003
#> ESC     -0.340       0.414    0.011
#> PUBW    -0.789      10.654    0.127
#> RES     -0.439       2.069    0.039
#> TULA    -0.130       1.268    0.027
#> PINE     0.095       0.039    0.005
#> PUBR    -0.959       3.792    0.078
#> (18 more)